+++ .env.example
@@ -3,5 +3,5 @@ BROWSERSTACK_USERNAME=
 BROWSERSTACK_ACCESS_KEY=
 DEV_BROWSERSTACK_TEST=
 API_CONVERSION_URLS=
-API_STATS_CONSENT_URLS=
+API_STATS_URLS=
 API_PROOF_CONSENT_URLS=
