+++ .env.production
@@ -3,5 +3,5 @@ BROWSERSTACK_USERNAME=
 BROWSERSTACK_ACCESS_KEY=
 DEV_BROWSERSTACK_TEST=
 API_CONVERSION_URLS=https://b.time1.me/v1/b
-API_STATS_CONSENT_URLS=https://c.time1.me/v1/log/consent
+API_STATS_URLS=https://c.time1.me/v1/log/consent
 API_PROOF_CONSENT_URLS=https://c.time1.me/v1/log/consent/proof
