+++ .env.test
@@ -1,0 +1,3 @@
+API_CONVERSION_URLS=https://api.test/v1/conversion,https://api.test/v2/conversion
+API_STATS_URLS=https://api.test/v1/stats-consent
+API_PROOF_CONSENT_URLS=https://api.test/v1/consent-proof
