+++ browserstack/site/sdk.js
@@ -1,4 +1,4 @@
 /*! isdk 1.0.1 https://github.com/https://github.com/TimeOne-Group/isdk#readme @license GPL-3.0 */
-!function(){"use strict";function t(t,e){var r=Object.keys(t);if(Object.getOwnPropertySymbols){var n=Object.getOwnPropertySymbols(t);e&&(n=n.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),r.push.apply(r,n)}return r}function e(e){for(var r=1;r<arguments.length;r++){var n=null!=arguments[r]?arguments[r]:{};r%2?t(Object(n),!0).forEach((function(t){s(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):t(Object(n)).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}function r(t){return r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},r(t)}function n(t,e,r,n,o,i,a){try{var u=t[i](a),s=u.value}catch(t){return void r(t)}u.done?e(s):Promise.resolve(s).then(n,o)}function o(t){return function(){var e=this,r=arguments;return new Promise((function(o,i){var a=t.apply(e,r);function u(t){n(a,o,i,u,s,"next",t)}function s(t){n(a,o,i,u,s,"throw",t)}u(void 0)}))}}function i(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function a(t,e){for(var r=0;r<e.length;r++){var n=e[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(t,n.key,n)}}function u(t,e,r){return e&&a(t.prototype,e),r&&a(t,r),Object.defineProperty(t,"prototype",{writable:!1}),t}function s(t,e,r){return e in t?Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}):t[e]=r,t}function c(t,e){return l(t)||function(t,e){var r=null==t?null:"undefined"!=typeof Symbol&&t[Symbol.iterator]||t["@@iterator"];if(null==r)return;var n,o,i=[],a=!0,u=!1;try{for(r=r.call(t);!(a=(n=r.next()).done)&&(i.push(n.value),!e||i.length!==e);a=!0);}catch(t){u=!0,o=t}finally{try{a||null==r.return||r.return()}finally{if(u)throw o}}return i}(t,e)||p(t,e)||y()}function f(t){return function(t){if(Array.isArray(t))return d(t)}(t)||h(t)||p(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function l(t){if(Array.isArray(t))return t}function h(t){if("undefined"!=typeof Symbol&&null!=t[Symbol.iterator]||null!=t["@@iterator"])return Array.from(t)}function p(t,e){if(t){if("string"==typeof t)return d(t,e);var r=Object.prototype.toString.call(t).slice(8,-1);return"Object"===r&&t.constructor&&(r=t.constructor.name),"Map"===r||"Set"===r?Array.from(t):"Arguments"===r||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r)?d(t,e):void 0}}function d(t,e){(null==e||e>t.length)&&(e=t.length);for(var r=0,n=new Array(e);r<e;r++)n[r]=t[r];return n}function y(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}function v(t,e){return function(t,e){if(e.get)return e.get.call(t);return e.value}(t,m(t,e,"get"))}function g(t,e,r){return function(t,e,r){if(e.set)e.set.call(t,r);else{if(!e.writable)throw new TypeError("attempted to set read only private field");e.value=r}}(t,m(t,e,"set"),r),r}function m(t,e,r){if(!e.has(t))throw new TypeError("attempted to "+r+" private field on non-instance");return e.get(t)}function b(t,e,r){if(!e.has(t))throw new TypeError("attempted to get private field on non-instance");return r}function w(t,e){if(e.has(t))throw new TypeError("Cannot initialize the same private elements twice on an object")}function S(t,e,r){w(t,e),e.set(t,r)}function O(t,e){w(t,e),e.add(t)}var x="undefined"!=typeof globalThis?globalThis:"undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:{};function k(t){var e={exports:{}};return t(e,e.exports),e.exports}var T,E,j=function(t){return t&&t.Math==Math&&t},A=j("object"==("undefined"==typeof globalThis?"undefined":r(globalThis))&&globalThis)||j("object"==("undefined"==typeof window?"undefined":r(window))&&window)||j("object"==("undefined"==typeof self?"undefined":r(self))&&self)||j("object"==r(x)&&x)||function(){return this}()||Function("return this")(),_=function(t){try{return!!t()}catch(t){return!0}},P=!_((function(){var t=function(){}.bind();return"function"!=typeof t||t.hasOwnProperty("prototype")})),I=Function.prototype,L=I.bind,R=I.call,C=P&&L.bind(R,R),U=P?function(t){return t&&C(t)}:function(t){return t&&function(){return R.apply(t,arguments)}},D=U({}.toString),F=U("".slice),M=function(t){return F(D(t),8,-1)},N=A.Object,B=U("".split),G=_((function(){return!N("z").propertyIsEnumerable(0)}))?function(t){return"String"==M(t)?B(t,""):N(t)}:N,W=A.TypeError,z=function(t){if(null==t)throw W("Can't call method on "+t);return t},q=function(t){return G(z(t))},V=Object.defineProperty,H=function(t,e){try{V(A,t,{value:e,configurable:!0,writable:!0})}catch(r){A[t]=e}return e},Q="__core-js_shared__",$=A[Q]||H(Q,{}),J=k((function(t){(t.exports=function(t,e){return $[t]||($[t]=void 0!==e?e:{})})("versions",[]).push({version:"3.22.4",mode:"global",copyright:"© 2014-2022 Denis Pushkarev (zloirock.ru)",license:"https://github.com/zloirock/core-js/blob/v3.22.4/LICENSE",source:"https://github.com/zloirock/core-js"})})),K=A.Object,Y=function(t){return K(z(t))},X=U({}.hasOwnProperty),Z=Object.hasOwn||function(t,e){return X(Y(t),e)},tt=0,et=Math.random(),rt=U(1..toString),nt=function(t){return"Symbol("+(void 0===t?"":t)+")_"+rt(++tt+et,36)},ot=function(t){return"function"==typeof t},it=function(t){return ot(t)?t:void 0},at=function(t,e){return arguments.length<2?it(A[t]):A[t]&&A[t][e]},ut=at("navigator","userAgent")||"",st=A.process,ct=A.Deno,ft=st&&st.versions||ct&&ct.version,lt=ft&&ft.v8;lt&&(E=(T=lt.split("."))[0]>0&&T[0]<4?1:+(T[0]+T[1])),!E&&ut&&(!(T=ut.match(/Edge\/(\d+)/))||T[1]>=74)&&(T=ut.match(/Chrome\/(\d+)/))&&(E=+T[1]);var ht,pt=E,dt=!!Object.getOwnPropertySymbols&&!_((function(){var t=Symbol();return!String(t)||!(Object(t)instanceof Symbol)||!Symbol.sham&&pt&&pt<41})),yt=dt&&!Symbol.sham&&"symbol"==r(Symbol.iterator),vt=J("wks"),gt=A.Symbol,mt=gt&&gt.for,bt=yt?gt:gt&&gt.withoutSetter||nt,wt=function(t){if(!Z(vt,t)||!dt&&"string"!=typeof vt[t]){var e="Symbol."+t;dt&&Z(gt,t)?vt[t]=gt[t]:vt[t]=yt&&mt?mt(e):bt(e)}return vt[t]},St=function(t){return"object"==r(t)?null!==t:ot(t)},Ot=A.String,xt=A.TypeError,kt=function(t){if(St(t))return t;throw xt(Ot(t)+" is not an object")},Tt=!_((function(){return 7!=Object.defineProperty({},1,{get:function(){return 7}})[1]})),Et=Tt&&_((function(){return 42!=Object.defineProperty((function(){}),"prototype",{value:42,writable:!1}).prototype})),jt=A.document,At=St(jt)&&St(jt.createElement),_t=function(t){return At?jt.createElement(t):{}},Pt=!Tt&&!_((function(){return 7!=Object.defineProperty(_t("div"),"a",{get:function(){return 7}}).a})),It=Function.prototype.call,Lt=P?It.bind(It):function(){return It.apply(It,arguments)},Rt=U({}.isPrototypeOf),Ct=A.Object,Ut=yt?function(t){return"symbol"==r(t)}:function(t){var e=at("Symbol");return ot(e)&&Rt(e.prototype,Ct(t))},Dt=A.String,Ft=function(t){try{return Dt(t)}catch(t){return"Object"}},Mt=A.TypeError,Nt=function(t){if(ot(t))return t;throw Mt(Ft(t)+" is not a function")},Bt=function(t,e){var r=t[e];return null==r?void 0:Nt(r)},Gt=A.TypeError,Wt=A.TypeError,zt=wt("toPrimitive"),qt=function(t,e){if(!St(t)||Ut(t))return t;var r,n=Bt(t,zt);if(n){if(void 0===e&&(e="default"),r=Lt(n,t,e),!St(r)||Ut(r))return r;throw Wt("Can't convert object to primitive value")}return void 0===e&&(e="number"),function(t,e){var r,n;if("string"===e&&ot(r=t.toString)&&!St(n=Lt(r,t)))return n;if(ot(r=t.valueOf)&&!St(n=Lt(r,t)))return n;if("string"!==e&&ot(r=t.toString)&&!St(n=Lt(r,t)))return n;throw Gt("Can't convert object to primitive value")}(t,e)},Vt=function(t){var e=qt(t,"string");return Ut(e)?e:e+""},Ht=A.TypeError,Qt=Object.defineProperty,$t=Object.getOwnPropertyDescriptor,Jt="enumerable",Kt="configurable",Yt="writable",Xt={f:Tt?Et?function(t,e,r){if(kt(t),e=Vt(e),kt(r),"function"==typeof t&&"prototype"===e&&"value"in r&&Yt in r&&!r.writable){var n=$t(t,e);n&&n.writable&&(t[e]=r.value,r={configurable:Kt in r?r.configurable:n.configurable,enumerable:Jt in r?r.enumerable:n.enumerable,writable:!1})}return Qt(t,e,r)}:Qt:function(t,e,r){if(kt(t),e=Vt(e),kt(r),Pt)try{return Qt(t,e,r)}catch(t){}if("get"in r||"set"in r)throw Ht("Accessors not supported");return"value"in r&&(t[e]=r.value),t}},Zt=Math.ceil,te=Math.floor,ee=function(t){var e=+t;return e!=e||0===e?0:(e>0?te:Zt)(e)},re=Math.max,ne=Math.min,oe=function(t,e){var r=ee(t);return r<0?re(r+e,0):ne(r,e)},ie=Math.min,ae=function(t){return(e=t.length)>0?ie(ee(e),9007199254740991):0;var e},ue=function(t){return function(e,r,n){var o,i=q(e),a=ae(i),u=oe(n,a);if(t&&r!=r){for(;a>u;)if((o=i[u++])!=o)return!0}else for(;a>u;u++)if((t||u in i)&&i[u]===r)return t||u||0;return!t&&-1}},se={includes:ue(!0),indexOf:ue(!1)},ce={},fe=se.indexOf,le=U([].push),he=function(t,e){var r,n=q(t),o=0,i=[];for(r in n)!Z(ce,r)&&Z(n,r)&&le(i,r);for(;e.length>o;)Z(n,r=e[o++])&&(~fe(i,r)||le(i,r));return i},pe=["constructor","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","toLocaleString","toString","valueOf"],de=Object.keys||function(t){return he(t,pe)},ye=Tt&&!Et?Object.defineProperties:function(t,e){kt(t);for(var r,n=q(e),o=de(e),i=o.length,a=0;i>a;)Xt.f(t,r=o[a++],n[r]);return t},ve={f:ye},ge=at("document","documentElement"),me=J("keys"),be=function(t){return me[t]||(me[t]=nt(t))},we=be("IE_PROTO"),Se=function(){},Oe=function(t){return"<script>"+t+"</"+"script>"},xe=function(t){t.write(Oe("")),t.close();var e=t.parentWindow.Object;return t=null,e},ke=function(){try{ht=new ActiveXObject("htmlfile")}catch(t){}var t,e;ke="undefined"!=typeof document?document.domain&&ht?xe(ht):((e=_t("iframe")).style.display="none",ge.appendChild(e),e.src=String("javascript:"),(t=e.contentWindow.document).open(),t.write(Oe("document.F=Object")),t.close(),t.F):xe(ht);for(var r=pe.length;r--;)delete ke.prototype[pe[r]];return ke()};ce[we]=!0;var Te=Object.create||function(t,e){var r;return null!==t?(Se.prototype=kt(t),r=new Se,Se.prototype=null,r[we]=t):r=ke(),void 0===e?r:ve.f(r,e)},Ee=wt("unscopables"),je=Array.prototype;null==je[Ee]&&Xt.f(je,Ee,{configurable:!0,value:Te(null)});var Ae=function(t){je[Ee][t]=!0},_e={},Pe=U(Function.toString);ot($.inspectSource)||($.inspectSource=function(t){return Pe(t)});var Ie,Le,Re,Ce=$.inspectSource,Ue=A.WeakMap,De=ot(Ue)&&/native code/.test(Ce(Ue)),Fe=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}},Me=Tt?function(t,e,r){return Xt.f(t,e,Fe(1,r))}:function(t,e,r){return t[e]=r,t},Ne="Object already initialized",Be=A.TypeError,Ge=A.WeakMap;if(De||$.state){var We=$.state||($.state=new Ge),ze=U(We.get),qe=U(We.has),Ve=U(We.set);Ie=function(t,e){if(qe(We,t))throw new Be(Ne);return e.facade=t,Ve(We,t,e),e},Le=function(t){return ze(We,t)||{}},Re=function(t){return qe(We,t)}}else{var He=be("state");ce[He]=!0,Ie=function(t,e){if(Z(t,He))throw new Be(Ne);return e.facade=t,Me(t,He,e),e},Le=function(t){return Z(t,He)?t[He]:{}},Re=function(t){return Z(t,He)}}var Qe,$e,Je,Ke={set:Ie,get:Le,has:Re,enforce:function(t){return Re(t)?Le(t):Ie(t,{})},getterFor:function(t){return function(e){var r;if(!St(e)||(r=Le(e)).type!==t)throw Be("Incompatible receiver, "+t+" required");return r}}},Ye={}.propertyIsEnumerable,Xe=Object.getOwnPropertyDescriptor,Ze={f:Xe&&!Ye.call({1:2},1)?function(t){var e=Xe(this,t);return!!e&&e.enumerable}:Ye},tr=Object.getOwnPropertyDescriptor,er={f:Tt?tr:function(t,e){if(t=q(t),e=Vt(e),Pt)try{return tr(t,e)}catch(t){}if(Z(t,e))return Fe(!Lt(Ze.f,t,e),t[e])}},rr=Function.prototype,nr=Tt&&Object.getOwnPropertyDescriptor,or=Z(rr,"name"),ir={EXISTS:or,PROPER:or&&"something"===function(){}.name,CONFIGURABLE:or&&(!Tt||Tt&&nr(rr,"name").configurable)},ar=k((function(t){var e=Xt.f,r=ir.CONFIGURABLE,n=Ke.enforce,o=Ke.get,i=!_((function(){return 8!==e((function(){}),"length",{value:8}).length})),a=String(String).split("String"),u=t.exports=function(t,o,u){"Symbol("===String(o).slice(0,7)&&(o="["+String(o).replace(/^Symbol\(([^)]*)\)/,"$1")+"]"),u&&u.getter&&(o="get "+o),u&&u.setter&&(o="set "+o),(!Z(t,"name")||r&&t.name!==o)&&e(t,"name",{value:o,configurable:!0}),i&&u&&Z(u,"arity")&&t.length!==u.arity&&e(t,"length",{value:u.arity});var s=n(t);return Z(s,"source")||(s.source=a.join("string"==typeof o?o:"")),t};Function.prototype.toString=u((function(){return ot(this)&&o(this).source||Ce(this)}),"toString")})),ur=function(t,e,r,n){var o=!!n&&!!n.unsafe,i=!!n&&!!n.enumerable,a=!!n&&!!n.noTargetGet,u=n&&void 0!==n.name?n.name:e;return ot(r)&&ar(r,u,n),t===A?(i?t[e]=r:H(e,r),t):(o?!a&&t[e]&&(i=!0):delete t[e],i?t[e]=r:Me(t,e,r),t)},sr=pe.concat("length","prototype"),cr={f:Object.getOwnPropertyNames||function(t){return he(t,sr)}},fr={f:Object.getOwnPropertySymbols},lr=U([].concat),hr=at("Reflect","ownKeys")||function(t){var e=cr.f(kt(t)),r=fr.f;return r?lr(e,r(t)):e},pr=function(t,e,r){for(var n=hr(e),o=Xt.f,i=er.f,a=0;a<n.length;a++){var u=n[a];Z(t,u)||r&&Z(r,u)||o(t,u,i(e,u))}},dr=/#|\.prototype\./,yr=function(t,e){var r=gr[vr(t)];return r==br||r!=mr&&(ot(e)?_(e):!!e)},vr=yr.normalize=function(t){return String(t).replace(dr,".").toLowerCase()},gr=yr.data={},mr=yr.NATIVE="N",br=yr.POLYFILL="P",wr=yr,Sr=er.f,Or=function(t,e){var n,o,i,a,u,s=t.target,c=t.global,f=t.stat;if(n=c?A:f?A[s]||H(s,{}):(A[s]||{}).prototype)for(o in e){if(a=e[o],i=t.noTargetGet?(u=Sr(n,o))&&u.value:n[o],!wr(c?o:s+(f?".":"#")+o,t.forced)&&void 0!==i){if(r(a)==r(i))continue;pr(a,i)}(t.sham||i&&i.sham)&&Me(a,"sham",!0),ur(n,o,a,t)}},xr=!_((function(){function t(){}return t.prototype.constructor=null,Object.getPrototypeOf(new t)!==t.prototype})),kr=be("IE_PROTO"),Tr=A.Object,Er=Tr.prototype,jr=xr?Tr.getPrototypeOf:function(t){var e=Y(t);if(Z(e,kr))return e[kr];var r=e.constructor;return ot(r)&&e instanceof r?r.prototype:e instanceof Tr?Er:null},Ar=wt("iterator"),_r=!1;[].keys&&("next"in(Je=[].keys())?($e=jr(jr(Je)))!==Object.prototype&&(Qe=$e):_r=!0);var Pr=null==Qe||_((function(){var t={};return Qe[Ar].call(t)!==t}));Pr&&(Qe={}),ot(Qe[Ar])||ur(Qe,Ar,(function(){return this}));var Ir={IteratorPrototype:Qe,BUGGY_SAFARI_ITERATORS:_r},Lr=Xt.f,Rr=wt("toStringTag"),Cr=function(t,e,r){t&&!r&&(t=t.prototype),t&&!Z(t,Rr)&&Lr(t,Rr,{configurable:!0,value:e})},Ur=Ir.IteratorPrototype,Dr=function(){return this},Fr=function(t,e,r,n){var o=e+" Iterator";return t.prototype=Te(Ur,{next:Fe(+!n,r)}),Cr(t,o,!1),_e[o]=Dr,t},Mr=A.String,Nr=A.TypeError,Br=Object.setPrototypeOf||("__proto__"in{}?function(){var t,e=!1,n={};try{(t=U(Object.getOwnPropertyDescriptor(Object.prototype,"__proto__").set))(n,[]),e=n instanceof Array}catch(t){}return function(n,o){return kt(n),function(t){if("object"==r(t)||ot(t))return t;throw Nr("Can't set "+Mr(t)+" as a prototype")}(o),e?t(n,o):n.__proto__=o,n}}():void 0),Gr=ir.PROPER,Wr=ir.CONFIGURABLE,zr=Ir.IteratorPrototype,qr=Ir.BUGGY_SAFARI_ITERATORS,Vr=wt("iterator"),Hr="keys",Qr="values",$r="entries",Jr=function(){return this},Kr=Xt.f,Yr="Array Iterator",Xr=Ke.set,Zr=Ke.getterFor(Yr),tn=function(t,e,r,n,o,i,a){Fr(r,e,n);var u,s,c,f=function(t){if(t===o&&y)return y;if(!qr&&t in p)return p[t];switch(t){case Hr:case Qr:case $r:return function(){return new r(this,t)}}return function(){return new r(this)}},l=e+" Iterator",h=!1,p=t.prototype,d=p[Vr]||p["@@iterator"]||o&&p[o],y=!qr&&d||f(o),v="Array"==e&&p.entries||d;if(v&&(u=jr(v.call(new t)))!==Object.prototype&&u.next&&(jr(u)!==zr&&(Br?Br(u,zr):ot(u[Vr])||ur(u,Vr,Jr)),Cr(u,l,!0)),Gr&&o==Qr&&d&&d.name!==Qr&&(Wr?Me(p,"name",Qr):(h=!0,y=function(){return Lt(d,this)})),o)if(s={values:f(Qr),keys:i?y:f(Hr),entries:f($r)},a)for(c in s)(qr||h||!(c in p))&&ur(p,c,s[c]);else Or({target:e,proto:!0,forced:qr||h},s);return p[Vr]!==y&&ur(p,Vr,y,{name:o}),_e[e]=y,s}(Array,"Array",(function(t,e){Xr(this,{type:Yr,target:q(t),index:0,kind:e})}),(function(){var t=Zr(this),e=t.target,r=t.kind,n=t.index++;return!e||n>=e.length?(t.target=void 0,{value:void 0,done:!0}):"keys"==r?{value:n,done:!1}:"values"==r?{value:e[n],done:!1}:{value:[n,e[n]],done:!1}}),"values"),en=_e.Arguments=_e.Array;if(Ae("keys"),Ae("values"),Ae("entries"),Tt&&"values"!==en.name)try{Kr(en,"name",{value:"values"})}catch(t){}var rn={};rn[wt("toStringTag")]="z";var nn="[object z]"===String(rn),on=wt("toStringTag"),an=A.Object,un="Arguments"==M(function(){return arguments}()),sn=nn?M:function(t){var e,r,n;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(r=function(t,e){try{return t[e]}catch(t){}}(e=an(t),on))?r:un?M(e):"Object"==(n=M(e))&&ot(e.callee)?"Arguments":n},cn=nn?{}.toString:function(){return"[object "+sn(this)+"]"};nn||ur(Object.prototype,"toString",cn,{unsafe:!0});var fn=A.Array,ln=Math.max,hn=function(t,e,r){for(var n,o,i,a,u=ae(t),s=oe(e,u),c=oe(void 0===r?u:r,u),f=fn(ln(c-s,0)),l=0;s<c;s++,l++)n=f,o=l,i=t[s],a=void 0,(a=Vt(o))in n?Xt.f(n,a,Fe(0,i)):n[a]=i;return f.length=l,f},pn=cr.f,dn="object"==("undefined"==typeof window?"undefined":r(window))&&window&&Object.getOwnPropertyNames?Object.getOwnPropertyNames(window):[],yn={f:function(t){return dn&&"Window"==M(t)?function(t){try{return pn(t)}catch(t){return hn(dn)}}(t):pn(q(t))}},vn=_((function(){if("function"==typeof ArrayBuffer){var t=new ArrayBuffer(8);Object.isExtensible(t)&&Object.defineProperty(t,"a",{value:8})}})),gn=Object.isExtensible,mn=_((function(){gn(1)}))||vn?function(t){return!!St(t)&&((!vn||"ArrayBuffer"!=M(t))&&(!gn||gn(t)))}:gn,bn=!_((function(){return Object.isExtensible(Object.preventExtensions({}))})),wn=k((function(t){var e=Xt.f,n=!1,o=nt("meta"),i=0,a=function(t){e(t,o,{value:{objectID:"O"+i++,weakData:{}}})},u=t.exports={enable:function(){u.enable=function(){},n=!0;var t=cr.f,e=U([].splice),r={};r[o]=1,t(r).length&&(cr.f=function(r){for(var n=t(r),i=0,a=n.length;i<a;i++)if(n[i]===o){e(n,i,1);break}return n},Or({target:"Object",stat:!0,forced:!0},{getOwnPropertyNames:yn.f}))},fastKey:function(t,e){if(!St(t))return"symbol"==r(t)?t:("string"==typeof t?"S":"P")+t;if(!Z(t,o)){if(!mn(t))return"F";if(!e)return"E";a(t)}return t[o].objectID},getWeakData:function(t,e){if(!Z(t,o)){if(!mn(t))return!0;if(!e)return!1;a(t)}return t[o].weakData},onFreeze:function(t){return bn&&n&&mn(t)&&!Z(t,o)&&a(t),t}};ce[o]=!0})),Sn=U(U.bind),On=function(t,e){return Nt(t),void 0===e?t:P?Sn(t,e):function(){return t.apply(e,arguments)}},xn=wt("iterator"),kn=Array.prototype,Tn=wt("iterator"),En=function(t){if(null!=t)return Bt(t,Tn)||Bt(t,"@@iterator")||_e[sn(t)]},jn=A.TypeError,An=function(t,e){var r=arguments.length<2?En(t):e;if(Nt(r))return kt(Lt(r,t));throw jn(Ft(t)+" is not iterable")},_n=function(t,e,r){var n,o;kt(t);try{if(!(n=Bt(t,"return"))){if("throw"===e)throw r;return r}n=Lt(n,t)}catch(t){o=!0,n=t}if("throw"===e)throw r;if(o)throw n;return kt(n),r},Pn=A.TypeError,In=function(t,e){this.stopped=t,this.result=e},Ln=In.prototype,Rn=function(t,e,n){var o,i,a,u,s,c,f,l,h=n&&n.that,p=!(!n||!n.AS_ENTRIES),d=!(!n||!n.IS_ITERATOR),y=!(!n||!n.INTERRUPTED),v=On(e,h),g=function(t){return o&&_n(o,"normal",t),new In(!0,t)},m=function(t){return p?(kt(t),y?v(t[0],t[1],g):v(t[0],t[1])):y?v(t,g):v(t)};if(d)o=t;else{if(!(i=En(t)))throw Pn(Ft(t)+" is not iterable");if(void 0!==(l=i)&&(_e.Array===l||kn[xn]===l)){for(a=0,u=ae(t);u>a;a++)if((s=m(t[a]))&&Rt(Ln,s))return s;return new In(!1)}o=An(t,i)}for(c=o.next;!(f=Lt(c,o)).done;){try{s=m(f.value)}catch(t){_n(o,"throw",t)}if("object"==r(s)&&s&&Rt(Ln,s))return s}return new In(!1)},Cn=A.TypeError,Un=function(t,e){if(Rt(e,t))return t;throw Cn("Incorrect invocation")},Dn=wt("iterator"),Fn=!1;try{var Mn=0,Nn={next:function(){return{done:!!Mn++}},return:function(){Fn=!0}};Nn[Dn]=function(){return this},Array.from(Nn,(function(){throw 2}))}catch(t){}var Bn=function(t,e,r){for(var n in e)ur(t,n,e[n],r);return t},Gn=Array.isArray||function(t){return"Array"==M(t)},Wn=function(){},zn=[],qn=at("Reflect","construct"),Vn=/^\s*(?:class|function)\b/,Hn=U(Vn.exec),Qn=!Vn.exec(Wn),$n=function(t){if(!ot(t))return!1;try{return qn(Wn,zn,t),!0}catch(t){return!1}},Jn=function(t){if(!ot(t))return!1;switch(sn(t)){case"AsyncFunction":case"GeneratorFunction":case"AsyncGeneratorFunction":return!1}try{return Qn||!!Hn(Vn,Ce(t))}catch(t){return!0}};Jn.sham=!0;var Kn=!qn||_((function(){var t;return $n($n.call)||!$n(Object)||!$n((function(){t=!0}))||t}))?Jn:$n,Yn=wt("species"),Xn=A.Array,Zn=function(t,e){return new(function(t){var e;return Gn(t)&&(e=t.constructor,(Kn(e)&&(e===Xn||Gn(e.prototype))||St(e)&&null===(e=e[Yn]))&&(e=void 0)),void 0===e?Xn:e}(t))(0===e?0:e)},to=U([].push),eo=function(t){var e=1==t,r=2==t,n=3==t,o=4==t,i=6==t,a=7==t,u=5==t||i;return function(s,c,f,l){for(var h,p,d=Y(s),y=G(d),v=On(c,f),g=ae(y),m=0,b=l||Zn,w=e?b(s,g):r||a?b(s,0):void 0;g>m;m++)if((u||m in y)&&(p=v(h=y[m],m,d),t))if(e)w[m]=p;else if(p)switch(t){case 3:return!0;case 5:return h;case 6:return m;case 2:to(w,h)}else switch(t){case 4:return!1;case 7:to(w,h)}return i?-1:n||o?o:w}},ro={forEach:eo(0),map:eo(1),filter:eo(2),some:eo(3),every:eo(4),find:eo(5),findIndex:eo(6),filterReject:eo(7)},no=wn.getWeakData,oo=Ke.set,io=Ke.getterFor,ao=ro.find,uo=ro.findIndex,so=U([].splice),co=0,fo=function(t){return t.frozen||(t.frozen=new lo)},lo=function(){this.entries=[]},ho=function(t,e){return ao(t.entries,(function(t){return t[0]===e}))};lo.prototype={get:function(t){var e=ho(this,t);if(e)return e[1]},has:function(t){return!!ho(this,t)},set:function(t,e){var r=ho(this,t);r?r[1]=e:this.entries.push([t,e])},delete:function(t){var e=uo(this.entries,(function(e){return e[0]===t}));return~e&&so(this.entries,e,1),!!~e}};var po={getConstructor:function(t,e,r,n){var o=t((function(t,o){Un(t,i),oo(t,{type:e,id:co++,frozen:void 0}),null!=o&&Rn(o,t[n],{that:t,AS_ENTRIES:r})})),i=o.prototype,a=io(e),u=function(t,e,r){var n=a(t),o=no(kt(e),!0);return!0===o?fo(n).set(e,r):o[n.id]=r,t};return Bn(i,{delete:function(t){var e=a(this);if(!St(t))return!1;var r=no(t);return!0===r?fo(e).delete(t):r&&Z(r,e.id)&&delete r[e.id]},has:function(t){var e=a(this);if(!St(t))return!1;var r=no(t);return!0===r?fo(e).has(t):r&&Z(r,e.id)}}),Bn(i,r?{get:function(t){var e=a(this);if(St(t)){var r=no(t);return!0===r?fo(e).get(t):r?r[e.id]:void 0}},set:function(t,e){return u(this,t,e)}}:{add:function(t){return u(this,t,!0)}}),o}};!function(t,e,r){var n=-1!==t.indexOf("Map"),o=-1!==t.indexOf("Weak"),i=n?"set":"add",a=A[t],u=a&&a.prototype,s=a,c={},f=function(t){var e=U(u[t]);ur(u,t,"add"==t?function(t){return e(this,0===t?0:t),this}:"delete"==t?function(t){return!(o&&!St(t))&&e(this,0===t?0:t)}:"get"==t?function(t){return o&&!St(t)?void 0:e(this,0===t?0:t)}:"has"==t?function(t){return!(o&&!St(t))&&e(this,0===t?0:t)}:function(t,r){return e(this,0===t?0:t,r),this})};if(wr(t,!ot(a)||!(o||u.forEach&&!_((function(){(new a).entries().next()})))))s=r.getConstructor(e,t,n,i),wn.enable();else if(wr(t,!0)){var l=new s,h=l[i](o?{}:-0,1)!=l,p=_((function(){l.has(1)})),d=function(t,e){if(!e&&!Fn)return!1;var r=!1;try{var n={};n[Dn]=function(){return{next:function(){return{done:r=!0}}}},t(n)}catch(t){}return r}((function(t){new a(t)})),y=!o&&_((function(){for(var t=new a,e=5;e--;)t[i](e,e);return!t.has(-0)}));d||((s=e((function(t,e){Un(t,u);var r=function(t,e,r){var n,o;return Br&&ot(n=e.constructor)&&n!==r&&St(o=n.prototype)&&o!==r.prototype&&Br(t,o),t}(new a,t,s);return null!=e&&Rn(e,r[i],{that:r,AS_ENTRIES:n}),r}))).prototype=u,u.constructor=s),(p||y)&&(f("delete"),f("has"),n&&f("get")),(y||h)&&f(i),o&&u.clear&&delete u.clear}c[t]=s,Or({global:!0,forced:s!=a},c),Cr(s,t),o||r.setStrong(s,t,n)}("WeakSet",(function(t){return function(){return t(this,arguments.length?arguments[0]:void 0)}}),po);var yo={CSSRuleList:0,CSSStyleDeclaration:0,CSSValueList:0,ClientRectList:0,DOMRectList:0,DOMStringList:0,DOMTokenList:1,DataTransferItemList:0,FileList:0,HTMLAllCollection:0,HTMLCollection:0,HTMLFormElement:0,HTMLSelectElement:0,MediaList:0,MimeTypeArray:0,NamedNodeMap:0,NodeList:1,PaintRequestList:0,Plugin:0,PluginArray:0,SVGLengthList:0,SVGNumberList:0,SVGPathSegList:0,SVGPointList:0,SVGStringList:0,SVGTransformList:0,SourceBufferList:0,StyleSheetList:0,TextTrackCueList:0,TextTrackList:0,TouchList:0},vo=_t("span").classList,go=vo&&vo.constructor&&vo.constructor.prototype,mo=go===Object.prototype?void 0:go,bo=wt("iterator"),wo=wt("toStringTag"),So=tn.values,Oo=function(t,e){if(t){if(t[bo]!==So)try{Me(t,bo,So)}catch(e){t[bo]=So}if(t[wo]||Me(t,wo,e),yo[e])for(var r in tn)if(t[r]!==tn[r])try{Me(t,r,tn[r])}catch(e){t[r]=tn[r]}}};for(var xo in yo)Oo(A[xo]&&A[xo].prototype,xo);Oo(mo,"DOMTokenList");var ko=wt("iterator"),To=!_((function(){var t=new URL("b?a=1&b=2&c=3","http://a"),e=t.searchParams,r="";return t.pathname="c%20d",e.forEach((function(t,n){e.delete("b"),r+=n+t})),!e.sort||"http://a/c%20d?a=1&c=3"!==t.href||"3"!==e.get("c")||"a=1"!==String(new URLSearchParams("?a=1"))||!e[ko]||"a"!==new URL("https://a@b").username||"b"!==new URLSearchParams(new URLSearchParams("a=b")).get("a")||"xn--e1aybc"!==new URL("http://тест").host||"#%D0%B1"!==new URL("http://a#б").hash||"a1c3"!==r||"x"!==new URL("http://x",void 0).host})),Eo=A.String,jo=function(t){if("Symbol"===sn(t))throw TypeError("Cannot convert a Symbol value to a string");return Eo(t)},Ao=A.TypeError,_o=function(t,e){if(t<e)throw Ao("Not enough arguments");return t},Po=Math.floor,Io=function(t,e){for(var r,n,o=t.length,i=1;i<o;){for(n=i,r=t[i];n&&e(t[n-1],r)>0;)t[n]=t[--n];n!==i++&&(t[n]=r)}return t},Lo=function(t,e,r,n){for(var o=e.length,i=r.length,a=0,u=0;a<o||u<i;)t[a+u]=a<o&&u<i?n(e[a],r[u])<=0?e[a++]:r[u++]:a<o?e[a++]:r[u++];return t},Ro=function t(e,r){var n=e.length,o=Po(n/2);return n<8?Io(e,r):Lo(e,t(hn(e,0,o),r),t(hn(e,o),r),r)},Co=wt("iterator"),Uo="URLSearchParams",Do="URLSearchParamsIterator",Fo=Ke.set,Mo=Ke.getterFor(Uo),No=Ke.getterFor(Do),Bo=Object.getOwnPropertyDescriptor,Go=function(t){if(!Tt)return A[t];var e=Bo(A,t);return e&&e.value},Wo=Go("fetch"),zo=Go("Request"),qo=Go("Headers"),Vo=zo&&zo.prototype,Ho=qo&&qo.prototype,Qo=A.RegExp,$o=A.TypeError,Jo=A.decodeURIComponent,Ko=A.encodeURIComponent,Yo=U("".charAt),Xo=U([].join),Zo=U([].push),ti=U("".replace),ei=U([].shift),ri=U([].splice),ni=U("".split),oi=U("".slice),ii=/\+/g,ai=Array(4),ui=function(t){return ai[t-1]||(ai[t-1]=Qo("((?:%[\\da-f]{2}){"+t+"})","gi"))},si=function(t){try{return Jo(t)}catch(e){return t}},ci=function(t){var e=ti(t,ii," "),r=4;try{return Jo(e)}catch(t){for(;r;)e=ti(e,ui(r--),si);return e}},fi=/[!'()~]|%20/g,li={"!":"%21","'":"%27","(":"%28",")":"%29","~":"%7E","%20":"+"},hi=function(t){return li[t]},pi=function(t){return ti(Ko(t),fi,hi)},di=Fr((function(t,e){Fo(this,{type:Do,iterator:An(Mo(t).entries),kind:e})}),"Iterator",(function(){var t=No(this),e=t.kind,r=t.iterator.next(),n=r.value;return r.done||(r.value="keys"===e?n.key:"values"===e?n.value:[n.key,n.value]),r}),!0),yi=function(t){this.entries=[],this.url=null,void 0!==t&&(St(t)?this.parseObject(t):this.parseQuery("string"==typeof t?"?"===Yo(t,0)?oi(t,1):t:jo(t)))};yi.prototype={type:Uo,bindURL:function(t){this.url=t,this.update()},parseObject:function(t){var e,r,n,o,i,a,u,s=En(t);if(s)for(r=(e=An(t,s)).next;!(n=Lt(r,e)).done;){if(i=(o=An(kt(n.value))).next,(a=Lt(i,o)).done||(u=Lt(i,o)).done||!Lt(i,o).done)throw $o("Expected sequence with length 2");Zo(this.entries,{key:jo(a.value),value:jo(u.value)})}else for(var c in t)Z(t,c)&&Zo(this.entries,{key:c,value:jo(t[c])})},parseQuery:function(t){if(t)for(var e,r,n=ni(t,"&"),o=0;o<n.length;)(e=n[o++]).length&&(r=ni(e,"="),Zo(this.entries,{key:ci(ei(r)),value:ci(Xo(r,"="))}))},serialize:function(){for(var t,e=this.entries,r=[],n=0;n<e.length;)t=e[n++],Zo(r,pi(t.key)+"="+pi(t.value));return Xo(r,"&")},update:function(){this.entries.length=0,this.parseQuery(this.url.query)},updateURL:function(){this.url&&this.url.update()}};var vi=function(){Un(this,gi);var t=arguments.length>0?arguments[0]:void 0;Fo(this,new yi(t))},gi=vi.prototype;if(Bn(gi,{append:function(t,e){_o(arguments.length,2);var r=Mo(this);Zo(r.entries,{key:jo(t),value:jo(e)}),r.updateURL()},delete:function(t){_o(arguments.length,1);for(var e=Mo(this),r=e.entries,n=jo(t),o=0;o<r.length;)r[o].key===n?ri(r,o,1):o++;e.updateURL()},get:function(t){_o(arguments.length,1);for(var e=Mo(this).entries,r=jo(t),n=0;n<e.length;n++)if(e[n].key===r)return e[n].value;return null},getAll:function(t){_o(arguments.length,1);for(var e=Mo(this).entries,r=jo(t),n=[],o=0;o<e.length;o++)e[o].key===r&&Zo(n,e[o].value);return n},has:function(t){_o(arguments.length,1);for(var e=Mo(this).entries,r=jo(t),n=0;n<e.length;)if(e[n++].key===r)return!0;return!1},set:function(t,e){_o(arguments.length,1);for(var r,n=Mo(this),o=n.entries,i=!1,a=jo(t),u=jo(e),s=0;s<o.length;s++)(r=o[s]).key===a&&(i?ri(o,s--,1):(i=!0,r.value=u));i||Zo(o,{key:a,value:u}),n.updateURL()},sort:function(){var t=Mo(this);Ro(t.entries,(function(t,e){return t.key>e.key?1:-1})),t.updateURL()},forEach:function(t){for(var e,r=Mo(this).entries,n=On(t,arguments.length>1?arguments[1]:void 0),o=0;o<r.length;)n((e=r[o++]).value,e.key,this)},keys:function(){return new di(this,"keys")},values:function(){return new di(this,"values")},entries:function(){return new di(this,"entries")}},{enumerable:!0}),ur(gi,Co,gi.entries,{name:"entries"}),ur(gi,"toString",(function(){return Mo(this).serialize()}),{enumerable:!0}),Cr(vi,Uo),Or({global:!0,forced:!To},{URLSearchParams:vi}),!To&&ot(qo)){var mi=U(Ho.has),bi=U(Ho.set),wi=function(t){if(St(t)){var e,r=t.body;if(sn(r)===Uo)return e=t.headers?new qo(t.headers):new qo,mi(e,"content-type")||bi(e,"content-type","application/x-www-form-urlencoded;charset=UTF-8"),Te(t,{body:Fe(0,jo(r)),headers:Fe(0,e)})}return t};if(ot(Wo)&&Or({global:!0,enumerable:!0,noTargetGet:!0,forced:!0},{fetch:function(t){return Wo(t,arguments.length>1?wi(arguments[1]):{})}}),ot(zo)){var Si=function(t){return Un(this,Vo),new zo(t,arguments.length>1?wi(arguments[1]):{})};Vo.constructor=Si,Si.prototype=Vo,Or({global:!0,forced:!0,noTargetGet:!0},{Request:Si})}}k((function(t){var e=function(t){var e,r=Object.prototype,n=r.hasOwnProperty,o="function"==typeof Symbol?Symbol:{},i=o.iterator||"@@iterator",a=o.asyncIterator||"@@asyncIterator",u=o.toStringTag||"@@toStringTag";function s(t,e,r){return Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{s({},"")}catch(t){s=function(t,e,r){return t[e]=r}}function c(t,e,r,n){var o=e&&e.prototype instanceof v?e:v,i=Object.create(o.prototype),a=new A(n||[]);return i._invoke=function(t,e,r){var n=l;return function(o,i){if(n===p)throw new Error("Generator is already running");if(n===d){if("throw"===o)throw i;return P()}for(r.method=o,r.arg=i;;){var a=r.delegate;if(a){var u=T(a,r);if(u){if(u===y)continue;return u}}if("next"===r.method)r.sent=r._sent=r.arg;else if("throw"===r.method){if(n===l)throw n=d,r.arg;r.dispatchException(r.arg)}else"return"===r.method&&r.abrupt("return",r.arg);n=p;var s=f(t,e,r);if("normal"===s.type){if(n=r.done?d:h,s.arg===y)continue;return{value:s.arg,done:r.done}}"throw"===s.type&&(n=d,r.method="throw",r.arg=s.arg)}}}(t,r,a),i}function f(t,e,r){try{return{type:"normal",arg:t.call(e,r)}}catch(t){return{type:"throw",arg:t}}}t.wrap=c;var l="suspendedStart",h="suspendedYield",p="executing",d="completed",y={};function v(){}function g(){}function m(){}var b={};s(b,i,(function(){return this}));var w=Object.getPrototypeOf,S=w&&w(w(_([])));S&&S!==r&&n.call(S,i)&&(b=S);var O=m.prototype=v.prototype=Object.create(b);function x(t){["next","throw","return"].forEach((function(e){s(t,e,(function(t){return this._invoke(e,t)}))}))}function k(t,e){function r(o,i,a,u){var s=f(t[o],t,i);if("throw"!==s.type){var c=s.arg,l=c.value;return l&&"object"==typeof l&&n.call(l,"__await")?e.resolve(l.__await).then((function(t){r("next",t,a,u)}),(function(t){r("throw",t,a,u)})):e.resolve(l).then((function(t){c.value=t,a(c)}),(function(t){return r("throw",t,a,u)}))}u(s.arg)}var o;this._invoke=function(t,n){function i(){return new e((function(e,o){r(t,n,e,o)}))}return o=o?o.then(i,i):i()}}function T(t,r){var n=t.iterator[r.method];if(n===e){if(r.delegate=null,"throw"===r.method){if(t.iterator.return&&(r.method="return",r.arg=e,T(t,r),"throw"===r.method))return y;r.method="throw",r.arg=new TypeError("The iterator does not provide a 'throw' method")}return y}var o=f(n,t.iterator,r.arg);if("throw"===o.type)return r.method="throw",r.arg=o.arg,r.delegate=null,y;var i=o.arg;return i?i.done?(r[t.resultName]=i.value,r.next=t.nextLoc,"return"!==r.method&&(r.method="next",r.arg=e),r.delegate=null,y):i:(r.method="throw",r.arg=new TypeError("iterator result is not an object"),r.delegate=null,y)}function E(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function j(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function A(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(E,this),this.reset(!0)}function _(t){if(t){var r=t[i];if(r)return r.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var o=-1,a=function r(){for(;++o<t.length;)if(n.call(t,o))return r.value=t[o],r.done=!1,r;return r.value=e,r.done=!0,r};return a.next=a}}return{next:P}}function P(){return{value:e,done:!0}}return g.prototype=m,s(O,"constructor",m),s(m,"constructor",g),g.displayName=s(m,u,"GeneratorFunction"),t.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===g||"GeneratorFunction"===(e.displayName||e.name))},t.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,m):(t.__proto__=m,s(t,u,"GeneratorFunction")),t.prototype=Object.create(O),t},t.awrap=function(t){return{__await:t}},x(k.prototype),s(k.prototype,a,(function(){return this})),t.AsyncIterator=k,t.async=function(e,r,n,o,i){void 0===i&&(i=Promise);var a=new k(c(e,r,n,o),i);return t.isGeneratorFunction(r)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},x(O),s(O,u,"Generator"),s(O,i,(function(){return this})),s(O,"toString",(function(){return"[object Generator]"})),t.keys=function(t){var e=[];for(var r in t)e.push(r);return e.reverse(),function r(){for(;e.length;){var n=e.pop();if(n in t)return r.value=n,r.done=!1,r}return r.done=!0,r}},t.values=_,A.prototype={constructor:A,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=e,this.done=!1,this.delegate=null,this.method="next",this.arg=e,this.tryEntries.forEach(j),!t)for(var r in this)"t"===r.charAt(0)&&n.call(this,r)&&!isNaN(+r.slice(1))&&(this[r]=e)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var r=this;function o(n,o){return u.type="throw",u.arg=t,r.next=n,o&&(r.method="next",r.arg=e),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var a=this.tryEntries[i],u=a.completion;if("root"===a.tryLoc)return o("end");if(a.tryLoc<=this.prev){var s=n.call(a,"catchLoc"),c=n.call(a,"finallyLoc");if(s&&c){if(this.prev<a.catchLoc)return o(a.catchLoc,!0);if(this.prev<a.finallyLoc)return o(a.finallyLoc)}else if(s){if(this.prev<a.catchLoc)return o(a.catchLoc,!0)}else{if(!c)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return o(a.finallyLoc)}}}},abrupt:function(t,e){for(var r=this.tryEntries.length-1;r>=0;--r){var o=this.tryEntries[r];if(o.tryLoc<=this.prev&&n.call(o,"finallyLoc")&&this.prev<o.finallyLoc){var i=o;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,y):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),y},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.finallyLoc===t)return this.complete(r.completion,r.afterLoc),j(r),y}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.tryLoc===t){var n=r.completion;if("throw"===n.type){var o=n.arg;j(r)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,r,n){return this.delegate={iterator:_(t),resultName:r,nextLoc:n},"next"===this.method&&(this.arg=e),y}},t}(t.exports);try{regeneratorRuntime=e}catch(t){"object"==typeof globalThis?globalThis.regeneratorRuntime=e:Function("r","regeneratorRuntime = r")(e)}}));var Oi="undefined"!=typeof globalThis&&globalThis||"undefined"!=typeof self&&self||void 0!==Oi&&Oi,xi="URLSearchParams"in Oi,ki="Symbol"in Oi&&"iterator"in Symbol,Ti="FileReader"in Oi&&"Blob"in Oi&&function(){try{return new Blob,!0}catch(t){return!1}}(),Ei="FormData"in Oi,ji="ArrayBuffer"in Oi;if(ji)var Ai=["[object Int8Array]","[object Uint8Array]","[object Uint8ClampedArray]","[object Int16Array]","[object Uint16Array]","[object Int32Array]","[object Uint32Array]","[object Float32Array]","[object Float64Array]"],_i=ArrayBuffer.isView||function(t){return t&&Ai.indexOf(Object.prototype.toString.call(t))>-1};function Pi(t){if("string"!=typeof t&&(t=String(t)),/[^a-z0-9\-#$%&'*+.^_`|~!]/i.test(t)||""===t)throw new TypeError('Invalid character in header field name: "'+t+'"');return t.toLowerCase()}function Ii(t){return"string"!=typeof t&&(t=String(t)),t}function Li(t){var e={next:function(){var e=t.shift();return{done:void 0===e,value:e}}};return ki&&(e[Symbol.iterator]=function(){return e}),e}function Ri(t){this.map={},t instanceof Ri?t.forEach((function(t,e){this.append(e,t)}),this):Array.isArray(t)?t.forEach((function(t){this.append(t[0],t[1])}),this):t&&Object.getOwnPropertyNames(t).forEach((function(e){this.append(e,t[e])}),this)}function Ci(t){if(t.bodyUsed)return Promise.reject(new TypeError("Already read"));t.bodyUsed=!0}function Ui(t){return new Promise((function(e,r){t.onload=function(){e(t.result)},t.onerror=function(){r(t.error)}}))}function Di(t){var e=new FileReader,r=Ui(e);return e.readAsArrayBuffer(t),r}function Fi(t){if(t.slice)return t.slice(0);var e=new Uint8Array(t.byteLength);return e.set(new Uint8Array(t)),e.buffer}function Mi(){return this.bodyUsed=!1,this._initBody=function(t){var e;this.bodyUsed=this.bodyUsed,this._bodyInit=t,t?"string"==typeof t?this._bodyText=t:Ti&&Blob.prototype.isPrototypeOf(t)?this._bodyBlob=t:Ei&&FormData.prototype.isPrototypeOf(t)?this._bodyFormData=t:xi&&URLSearchParams.prototype.isPrototypeOf(t)?this._bodyText=t.toString():ji&&Ti&&((e=t)&&DataView.prototype.isPrototypeOf(e))?(this._bodyArrayBuffer=Fi(t.buffer),this._bodyInit=new Blob([this._bodyArrayBuffer])):ji&&(ArrayBuffer.prototype.isPrototypeOf(t)||_i(t))?this._bodyArrayBuffer=Fi(t):this._bodyText=t=Object.prototype.toString.call(t):this._bodyText="",this.headers.get("content-type")||("string"==typeof t?this.headers.set("content-type","text/plain;charset=UTF-8"):this._bodyBlob&&this._bodyBlob.type?this.headers.set("content-type",this._bodyBlob.type):xi&&URLSearchParams.prototype.isPrototypeOf(t)&&this.headers.set("content-type","application/x-www-form-urlencoded;charset=UTF-8"))},Ti&&(this.blob=function(){var t=Ci(this);if(t)return t;if(this._bodyBlob)return Promise.resolve(this._bodyBlob);if(this._bodyArrayBuffer)return Promise.resolve(new Blob([this._bodyArrayBuffer]));if(this._bodyFormData)throw new Error("could not read FormData body as blob");return Promise.resolve(new Blob([this._bodyText]))},this.arrayBuffer=function(){if(this._bodyArrayBuffer){var t=Ci(this);return t||(ArrayBuffer.isView(this._bodyArrayBuffer)?Promise.resolve(this._bodyArrayBuffer.buffer.slice(this._bodyArrayBuffer.byteOffset,this._bodyArrayBuffer.byteOffset+this._bodyArrayBuffer.byteLength)):Promise.resolve(this._bodyArrayBuffer))}return this.blob().then(Di)}),this.text=function(){var t,e,r,n=Ci(this);if(n)return n;if(this._bodyBlob)return t=this._bodyBlob,e=new FileReader,r=Ui(e),e.readAsText(t),r;if(this._bodyArrayBuffer)return Promise.resolve(function(t){for(var e=new Uint8Array(t),r=new Array(e.length),n=0;n<e.length;n++)r[n]=String.fromCharCode(e[n]);return r.join("")}(this._bodyArrayBuffer));if(this._bodyFormData)throw new Error("could not read FormData body as text");return Promise.resolve(this._bodyText)},Ei&&(this.formData=function(){return this.text().then(Gi)}),this.json=function(){return this.text().then(JSON.parse)},this}Ri.prototype.append=function(t,e){t=Pi(t),e=Ii(e);var r=this.map[t];this.map[t]=r?r+", "+e:e},Ri.prototype.delete=function(t){delete this.map[Pi(t)]},Ri.prototype.get=function(t){return t=Pi(t),this.has(t)?this.map[t]:null},Ri.prototype.has=function(t){return this.map.hasOwnProperty(Pi(t))},Ri.prototype.set=function(t,e){this.map[Pi(t)]=Ii(e)},Ri.prototype.forEach=function(t,e){for(var r in this.map)this.map.hasOwnProperty(r)&&t.call(e,this.map[r],r,this)},Ri.prototype.keys=function(){var t=[];return this.forEach((function(e,r){t.push(r)})),Li(t)},Ri.prototype.values=function(){var t=[];return this.forEach((function(e){t.push(e)})),Li(t)},Ri.prototype.entries=function(){var t=[];return this.forEach((function(e,r){t.push([r,e])})),Li(t)},ki&&(Ri.prototype[Symbol.iterator]=Ri.prototype.entries);var Ni=["DELETE","GET","HEAD","OPTIONS","POST","PUT"];function Bi(t,e){if(!(this instanceof Bi))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');var r,n,o=(e=e||{}).body;if(t instanceof Bi){if(t.bodyUsed)throw new TypeError("Already read");this.url=t.url,this.credentials=t.credentials,e.headers||(this.headers=new Ri(t.headers)),this.method=t.method,this.mode=t.mode,this.signal=t.signal,o||null==t._bodyInit||(o=t._bodyInit,t.bodyUsed=!0)}else this.url=String(t);if(this.credentials=e.credentials||this.credentials||"same-origin",!e.headers&&this.headers||(this.headers=new Ri(e.headers)),this.method=(r=e.method||this.method||"GET",n=r.toUpperCase(),Ni.indexOf(n)>-1?n:r),this.mode=e.mode||this.mode||null,this.signal=e.signal||this.signal,this.referrer=null,("GET"===this.method||"HEAD"===this.method)&&o)throw new TypeError("Body not allowed for GET or HEAD requests");if(this._initBody(o),!("GET"!==this.method&&"HEAD"!==this.method||"no-store"!==e.cache&&"no-cache"!==e.cache)){var i=/([?&])_=[^&]*/;if(i.test(this.url))this.url=this.url.replace(i,"$1_="+(new Date).getTime());else{this.url+=(/\?/.test(this.url)?"&":"?")+"_="+(new Date).getTime()}}}function Gi(t){var e=new FormData;return t.trim().split("&").forEach((function(t){if(t){var r=t.split("="),n=r.shift().replace(/\+/g," "),o=r.join("=").replace(/\+/g," ");e.append(decodeURIComponent(n),decodeURIComponent(o))}})),e}function Wi(t,e){if(!(this instanceof Wi))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');e||(e={}),this.type="default",this.status=void 0===e.status?200:e.status,this.ok=this.status>=200&&this.status<300,this.statusText=void 0===e.statusText?"":""+e.statusText,this.headers=new Ri(e.headers),this.url=e.url||"",this._initBody(t)}Bi.prototype.clone=function(){return new Bi(this,{body:this._bodyInit})},Mi.call(Bi.prototype),Mi.call(Wi.prototype),Wi.prototype.clone=function(){return new Wi(this._bodyInit,{status:this.status,statusText:this.statusText,headers:new Ri(this.headers),url:this.url})},Wi.error=function(){var t=new Wi(null,{status:0,statusText:""});return t.type="error",t};var zi=[301,302,303,307,308];Wi.redirect=function(t,e){if(-1===zi.indexOf(e))throw new RangeError("Invalid status code");return new Wi(null,{status:e,headers:{location:t}})};var qi,Vi=Oi.DOMException;try{new Vi}catch(t){(Vi=function(t,e){this.message=t,this.name=e;var r=Error(t);this.stack=r.stack}).prototype=Object.create(Error.prototype),Vi.prototype.constructor=Vi}function Hi(t,e){return new Promise((function(r,n){var o=new Bi(t,e);if(o.signal&&o.signal.aborted)return n(new Vi("Aborted","AbortError"));var i=new XMLHttpRequest;function a(){i.abort()}i.onload=function(){var t,e,n={status:i.status,statusText:i.statusText,headers:(t=i.getAllResponseHeaders()||"",e=new Ri,t.replace(/\r?\n[\t ]+/g," ").split("\r").map((function(t){return 0===t.indexOf("\n")?t.substr(1,t.length):t})).forEach((function(t){var r=t.split(":"),n=r.shift().trim();if(n){var o=r.join(":").trim();e.append(n,o)}})),e)};n.url="responseURL"in i?i.responseURL:n.headers.get("X-Request-URL");var o="response"in i?i.response:i.responseText;setTimeout((function(){r(new Wi(o,n))}),0)},i.onerror=function(){setTimeout((function(){n(new TypeError("Network request failed"))}),0)},i.ontimeout=function(){setTimeout((function(){n(new TypeError("Network request failed"))}),0)},i.onabort=function(){setTimeout((function(){n(new Vi("Aborted","AbortError"))}),0)},i.open(o.method,function(t){try{return""===t&&Oi.location.href?Oi.location.href:t}catch(e){return t}}(o.url),!0),"include"===o.credentials?i.withCredentials=!0:"omit"===o.credentials&&(i.withCredentials=!1),"responseType"in i&&(Ti?i.responseType="blob":ji&&o.headers.get("Content-Type")&&-1!==o.headers.get("Content-Type").indexOf("application/octet-stream")&&(i.responseType="arraybuffer")),!e||"object"!=typeof e.headers||e.headers instanceof Ri?o.headers.forEach((function(t,e){i.setRequestHeader(e,t)})):Object.getOwnPropertyNames(e.headers).forEach((function(t){i.setRequestHeader(t,Ii(e.headers[t]))})),o.signal&&(o.signal.addEventListener("abort",a),i.onreadystatechange=function(){4===i.readyState&&o.signal.removeEventListener("abort",a)}),i.send(void 0===o._bodyInit?null:o._bodyInit)}))}Hi.polyfill=!0,Oi.fetch||(Oi.fetch=Hi,Oi.Headers=Ri,Oi.Request=Bi,Oi.Response=Wi);var Qi=new Uint8Array(16);function $i(){if(!qi&&!(qi="undefined"!=typeof crypto&&crypto.getRandomValues&&crypto.getRandomValues.bind(crypto)||"undefined"!=typeof msCrypto&&"function"==typeof msCrypto.getRandomValues&&msCrypto.getRandomValues.bind(msCrypto)))throw new Error("crypto.getRandomValues() not supported. See https://github.com/uuidjs/uuid#getrandomvalues-not-supported");return qi(Qi)}var Ji=/^(?:[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|00000000-0000-0000-0000-000000000000)$/i;function Ki(t){return"string"==typeof t&&Ji.test(t)}for(var Yi=[],Xi=0;Xi<256;++Xi)Yi.push((Xi+256).toString(16).substr(1));function Zi(t,e,r){var n=(t=t||{}).random||(t.rng||$i)();if(n[6]=15&n[6]|64,n[8]=63&n[8]|128,e){r=r||0;for(var o=0;o<16;++o)e[r+o]=n[o];return e}return function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:0,r=(Yi[t[e+0]]+Yi[t[e+1]]+Yi[t[e+2]]+Yi[t[e+3]]+"-"+Yi[t[e+4]]+Yi[t[e+5]]+"-"+Yi[t[e+6]]+Yi[t[e+7]]+"-"+Yi[t[e+8]]+Yi[t[e+9]]+"-"+Yi[t[e+10]]+Yi[t[e+11]]+Yi[t[e+12]]+Yi[t[e+13]]+Yi[t[e+14]]+Yi[t[e+15]]).toLowerCase();if(!Ki(r))throw TypeError("Stringified UUID is invalid");return r}(n)}
+!function(){"use strict";function t(t,e){var r=Object.keys(t);if(Object.getOwnPropertySymbols){var n=Object.getOwnPropertySymbols(t);e&&(n=n.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),r.push.apply(r,n)}return r}function e(e){for(var r=1;r<arguments.length;r++){var n=null!=arguments[r]?arguments[r]:{};r%2?t(Object(n),!0).forEach((function(t){c(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):t(Object(n)).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}function r(t){return r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},r(t)}function n(t,e,r,n,o,i,a){try{var u=t[i](a),c=u.value}catch(t){return void r(t)}u.done?e(c):Promise.resolve(c).then(n,o)}function o(t){return function(){var e=this,r=arguments;return new Promise((function(o,i){var a=t.apply(e,r);function u(t){n(a,o,i,u,c,"next",t)}function c(t){n(a,o,i,u,c,"throw",t)}u(void 0)}))}}function i(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function a(t,e){for(var r=0;r<e.length;r++){var n=e[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(t,n.key,n)}}function u(t,e,r){return e&&a(t.prototype,e),r&&a(t,r),Object.defineProperty(t,"prototype",{writable:!1}),t}function c(t,e,r){return e in t?Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}):t[e]=r,t}function s(t,e){return l(t)||function(t,e){var r=null==t?null:"undefined"!=typeof Symbol&&t[Symbol.iterator]||t["@@iterator"];if(null==r)return;var n,o,i=[],a=!0,u=!1;try{for(r=r.call(t);!(a=(n=r.next()).done)&&(i.push(n.value),!e||i.length!==e);a=!0);}catch(t){u=!0,o=t}finally{try{a||null==r.return||r.return()}finally{if(u)throw o}}return i}(t,e)||p(t,e)||y()}function f(t){return function(t){if(Array.isArray(t))return d(t)}(t)||h(t)||p(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function l(t){if(Array.isArray(t))return t}function h(t){if("undefined"!=typeof Symbol&&null!=t[Symbol.iterator]||null!=t["@@iterator"])return Array.from(t)}function p(t,e){if(t){if("string"==typeof t)return d(t,e);var r=Object.prototype.toString.call(t).slice(8,-1);return"Object"===r&&t.constructor&&(r=t.constructor.name),"Map"===r||"Set"===r?Array.from(t):"Arguments"===r||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r)?d(t,e):void 0}}function d(t,e){(null==e||e>t.length)&&(e=t.length);for(var r=0,n=new Array(e);r<e;r++)n[r]=t[r];return n}function y(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}function v(t,e){return function(t,e){if(e.get)return e.get.call(t);return e.value}(t,b(t,e,"get"))}function g(t,e,r){return function(t,e,r){if(e.set)e.set.call(t,r);else{if(!e.writable)throw new TypeError("attempted to set read only private field");e.value=r}}(t,b(t,e,"set"),r),r}function b(t,e,r){if(!e.has(t))throw new TypeError("attempted to "+r+" private field on non-instance");return e.get(t)}function m(t,e,r){if(!e.has(t))throw new TypeError("attempted to get private field on non-instance");return r}function w(t,e){if(e.has(t))throw new TypeError("Cannot initialize the same private elements twice on an object")}function S(t,e,r){w(t,e),e.set(t,r)}function O(t,e){w(t,e),e.add(t)}var x="undefined"!=typeof globalThis?globalThis:"undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:{};function k(t){var e={exports:{}};return t(e,e.exports),e.exports}var E,T,j=function(t){return t&&t.Math==Math&&t},A=j("object"==("undefined"==typeof globalThis?"undefined":r(globalThis))&&globalThis)||j("object"==("undefined"==typeof window?"undefined":r(window))&&window)||j("object"==("undefined"==typeof self?"undefined":r(self))&&self)||j("object"==r(x)&&x)||function(){return this}()||Function("return this")(),_=function(t){try{return!!t()}catch(t){return!0}},P=!_((function(){var t=function(){}.bind();return"function"!=typeof t||t.hasOwnProperty("prototype")})),I=Function.prototype,L=I.bind,R=I.call,C=P&&L.bind(R,R),U=P?function(t){return t&&C(t)}:function(t){return t&&function(){return R.apply(t,arguments)}},D=U({}.toString),F=U("".slice),M=function(t){return F(D(t),8,-1)},B=A.Object,N=U("".split),G=_((function(){return!B("z").propertyIsEnumerable(0)}))?function(t){return"String"==M(t)?N(t,""):B(t)}:B,z=A.TypeError,W=function(t){if(null==t)throw z("Can't call method on "+t);return t},V=function(t){return G(W(t))},q=Object.defineProperty,H=function(t,e){try{q(A,t,{value:e,configurable:!0,writable:!0})}catch(r){A[t]=e}return e},Q="__core-js_shared__",$=A[Q]||H(Q,{}),K=k((function(t){(t.exports=function(t,e){return $[t]||($[t]=void 0!==e?e:{})})("versions",[]).push({version:"3.22.4",mode:"global",copyright:"© 2014-2022 Denis Pushkarev (zloirock.ru)",license:"https://github.com/zloirock/core-js/blob/v3.22.4/LICENSE",source:"https://github.com/zloirock/core-js"})})),Y=A.Object,J=function(t){return Y(W(t))},X=U({}.hasOwnProperty),Z=Object.hasOwn||function(t,e){return X(J(t),e)},tt=0,et=Math.random(),rt=U(1..toString),nt=function(t){return"Symbol("+(void 0===t?"":t)+")_"+rt(++tt+et,36)},ot=function(t){return"function"==typeof t},it=function(t){return ot(t)?t:void 0},at=function(t,e){return arguments.length<2?it(A[t]):A[t]&&A[t][e]},ut=at("navigator","userAgent")||"",ct=A.process,st=A.Deno,ft=ct&&ct.versions||st&&st.version,lt=ft&&ft.v8;lt&&(T=(E=lt.split("."))[0]>0&&E[0]<4?1:+(E[0]+E[1])),!T&&ut&&(!(E=ut.match(/Edge\/(\d+)/))||E[1]>=74)&&(E=ut.match(/Chrome\/(\d+)/))&&(T=+E[1]);var ht,pt=T,dt=!!Object.getOwnPropertySymbols&&!_((function(){var t=Symbol();return!String(t)||!(Object(t)instanceof Symbol)||!Symbol.sham&&pt&&pt<41})),yt=dt&&!Symbol.sham&&"symbol"==r(Symbol.iterator),vt=K("wks"),gt=A.Symbol,bt=gt&&gt.for,mt=yt?gt:gt&&gt.withoutSetter||nt,wt=function(t){if(!Z(vt,t)||!dt&&"string"!=typeof vt[t]){var e="Symbol."+t;dt&&Z(gt,t)?vt[t]=gt[t]:vt[t]=yt&&bt?bt(e):mt(e)}return vt[t]},St=function(t){return"object"==r(t)?null!==t:ot(t)},Ot=A.String,xt=A.TypeError,kt=function(t){if(St(t))return t;throw xt(Ot(t)+" is not an object")},Et=!_((function(){return 7!=Object.defineProperty({},1,{get:function(){return 7}})[1]})),Tt=Et&&_((function(){return 42!=Object.defineProperty((function(){}),"prototype",{value:42,writable:!1}).prototype})),jt=A.document,At=St(jt)&&St(jt.createElement),_t=function(t){return At?jt.createElement(t):{}},Pt=!Et&&!_((function(){return 7!=Object.defineProperty(_t("div"),"a",{get:function(){return 7}}).a})),It=Function.prototype.call,Lt=P?It.bind(It):function(){return It.apply(It,arguments)},Rt=U({}.isPrototypeOf),Ct=A.Object,Ut=yt?function(t){return"symbol"==r(t)}:function(t){var e=at("Symbol");return ot(e)&&Rt(e.prototype,Ct(t))},Dt=A.String,Ft=function(t){try{return Dt(t)}catch(t){return"Object"}},Mt=A.TypeError,Bt=function(t){if(ot(t))return t;throw Mt(Ft(t)+" is not a function")},Nt=function(t,e){var r=t[e];return null==r?void 0:Bt(r)},Gt=A.TypeError,zt=A.TypeError,Wt=wt("toPrimitive"),Vt=function(t,e){if(!St(t)||Ut(t))return t;var r,n=Nt(t,Wt);if(n){if(void 0===e&&(e="default"),r=Lt(n,t,e),!St(r)||Ut(r))return r;throw zt("Can't convert object to primitive value")}return void 0===e&&(e="number"),function(t,e){var r,n;if("string"===e&&ot(r=t.toString)&&!St(n=Lt(r,t)))return n;if(ot(r=t.valueOf)&&!St(n=Lt(r,t)))return n;if("string"!==e&&ot(r=t.toString)&&!St(n=Lt(r,t)))return n;throw Gt("Can't convert object to primitive value")}(t,e)},qt=function(t){var e=Vt(t,"string");return Ut(e)?e:e+""},Ht=A.TypeError,Qt=Object.defineProperty,$t=Object.getOwnPropertyDescriptor,Kt="enumerable",Yt="configurable",Jt="writable",Xt={f:Et?Tt?function(t,e,r){if(kt(t),e=qt(e),kt(r),"function"==typeof t&&"prototype"===e&&"value"in r&&Jt in r&&!r.writable){var n=$t(t,e);n&&n.writable&&(t[e]=r.value,r={configurable:Yt in r?r.configurable:n.configurable,enumerable:Kt in r?r.enumerable:n.enumerable,writable:!1})}return Qt(t,e,r)}:Qt:function(t,e,r){if(kt(t),e=qt(e),kt(r),Pt)try{return Qt(t,e,r)}catch(t){}if("get"in r||"set"in r)throw Ht("Accessors not supported");return"value"in r&&(t[e]=r.value),t}},Zt=Math.ceil,te=Math.floor,ee=function(t){var e=+t;return e!=e||0===e?0:(e>0?te:Zt)(e)},re=Math.max,ne=Math.min,oe=function(t,e){var r=ee(t);return r<0?re(r+e,0):ne(r,e)},ie=Math.min,ae=function(t){return(e=t.length)>0?ie(ee(e),9007199254740991):0;var e},ue=function(t){return function(e,r,n){var o,i=V(e),a=ae(i),u=oe(n,a);if(t&&r!=r){for(;a>u;)if((o=i[u++])!=o)return!0}else for(;a>u;u++)if((t||u in i)&&i[u]===r)return t||u||0;return!t&&-1}},ce={includes:ue(!0),indexOf:ue(!1)},se={},fe=ce.indexOf,le=U([].push),he=function(t,e){var r,n=V(t),o=0,i=[];for(r in n)!Z(se,r)&&Z(n,r)&&le(i,r);for(;e.length>o;)Z(n,r=e[o++])&&(~fe(i,r)||le(i,r));return i},pe=["constructor","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","toLocaleString","toString","valueOf"],de=Object.keys||function(t){return he(t,pe)},ye=Et&&!Tt?Object.defineProperties:function(t,e){kt(t);for(var r,n=V(e),o=de(e),i=o.length,a=0;i>a;)Xt.f(t,r=o[a++],n[r]);return t},ve={f:ye},ge=at("document","documentElement"),be=K("keys"),me=function(t){return be[t]||(be[t]=nt(t))},we=me("IE_PROTO"),Se=function(){},Oe=function(t){return"<script>"+t+"</"+"script>"},xe=function(t){t.write(Oe("")),t.close();var e=t.parentWindow.Object;return t=null,e},ke=function(){try{ht=new ActiveXObject("htmlfile")}catch(t){}var t,e;ke="undefined"!=typeof document?document.domain&&ht?xe(ht):((e=_t("iframe")).style.display="none",ge.appendChild(e),e.src=String("javascript:"),(t=e.contentWindow.document).open(),t.write(Oe("document.F=Object")),t.close(),t.F):xe(ht);for(var r=pe.length;r--;)delete ke.prototype[pe[r]];return ke()};se[we]=!0;var Ee=Object.create||function(t,e){var r;return null!==t?(Se.prototype=kt(t),r=new Se,Se.prototype=null,r[we]=t):r=ke(),void 0===e?r:ve.f(r,e)},Te=wt("unscopables"),je=Array.prototype;null==je[Te]&&Xt.f(je,Te,{configurable:!0,value:Ee(null)});var Ae=function(t){je[Te][t]=!0},_e={},Pe=U(Function.toString);ot($.inspectSource)||($.inspectSource=function(t){return Pe(t)});var Ie,Le,Re,Ce=$.inspectSource,Ue=A.WeakMap,De=ot(Ue)&&/native code/.test(Ce(Ue)),Fe=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}},Me=Et?function(t,e,r){return Xt.f(t,e,Fe(1,r))}:function(t,e,r){return t[e]=r,t},Be="Object already initialized",Ne=A.TypeError,Ge=A.WeakMap;if(De||$.state){var ze=$.state||($.state=new Ge),We=U(ze.get),Ve=U(ze.has),qe=U(ze.set);Ie=function(t,e){if(Ve(ze,t))throw new Ne(Be);return e.facade=t,qe(ze,t,e),e},Le=function(t){return We(ze,t)||{}},Re=function(t){return Ve(ze,t)}}else{var He=me("state");se[He]=!0,Ie=function(t,e){if(Z(t,He))throw new Ne(Be);return e.facade=t,Me(t,He,e),e},Le=function(t){return Z(t,He)?t[He]:{}},Re=function(t){return Z(t,He)}}var Qe,$e,Ke,Ye={set:Ie,get:Le,has:Re,enforce:function(t){return Re(t)?Le(t):Ie(t,{})},getterFor:function(t){return function(e){var r;if(!St(e)||(r=Le(e)).type!==t)throw Ne("Incompatible receiver, "+t+" required");return r}}},Je={}.propertyIsEnumerable,Xe=Object.getOwnPropertyDescriptor,Ze={f:Xe&&!Je.call({1:2},1)?function(t){var e=Xe(this,t);return!!e&&e.enumerable}:Je},tr=Object.getOwnPropertyDescriptor,er={f:Et?tr:function(t,e){if(t=V(t),e=qt(e),Pt)try{return tr(t,e)}catch(t){}if(Z(t,e))return Fe(!Lt(Ze.f,t,e),t[e])}},rr=Function.prototype,nr=Et&&Object.getOwnPropertyDescriptor,or=Z(rr,"name"),ir={EXISTS:or,PROPER:or&&"something"===function(){}.name,CONFIGURABLE:or&&(!Et||Et&&nr(rr,"name").configurable)},ar=k((function(t){var e=Xt.f,r=ir.CONFIGURABLE,n=Ye.enforce,o=Ye.get,i=!_((function(){return 8!==e((function(){}),"length",{value:8}).length})),a=String(String).split("String"),u=t.exports=function(t,o,u){"Symbol("===String(o).slice(0,7)&&(o="["+String(o).replace(/^Symbol\(([^)]*)\)/,"$1")+"]"),u&&u.getter&&(o="get "+o),u&&u.setter&&(o="set "+o),(!Z(t,"name")||r&&t.name!==o)&&e(t,"name",{value:o,configurable:!0}),i&&u&&Z(u,"arity")&&t.length!==u.arity&&e(t,"length",{value:u.arity});var c=n(t);return Z(c,"source")||(c.source=a.join("string"==typeof o?o:"")),t};Function.prototype.toString=u((function(){return ot(this)&&o(this).source||Ce(this)}),"toString")})),ur=function(t,e,r,n){var o=!!n&&!!n.unsafe,i=!!n&&!!n.enumerable,a=!!n&&!!n.noTargetGet,u=n&&void 0!==n.name?n.name:e;return ot(r)&&ar(r,u,n),t===A?(i?t[e]=r:H(e,r),t):(o?!a&&t[e]&&(i=!0):delete t[e],i?t[e]=r:Me(t,e,r),t)},cr=pe.concat("length","prototype"),sr={f:Object.getOwnPropertyNames||function(t){return he(t,cr)}},fr={f:Object.getOwnPropertySymbols},lr=U([].concat),hr=at("Reflect","ownKeys")||function(t){var e=sr.f(kt(t)),r=fr.f;return r?lr(e,r(t)):e},pr=function(t,e,r){for(var n=hr(e),o=Xt.f,i=er.f,a=0;a<n.length;a++){var u=n[a];Z(t,u)||r&&Z(r,u)||o(t,u,i(e,u))}},dr=/#|\.prototype\./,yr=function(t,e){var r=gr[vr(t)];return r==mr||r!=br&&(ot(e)?_(e):!!e)},vr=yr.normalize=function(t){return String(t).replace(dr,".").toLowerCase()},gr=yr.data={},br=yr.NATIVE="N",mr=yr.POLYFILL="P",wr=yr,Sr=er.f,Or=function(t,e){var n,o,i,a,u,c=t.target,s=t.global,f=t.stat;if(n=s?A:f?A[c]||H(c,{}):(A[c]||{}).prototype)for(o in e){if(a=e[o],i=t.noTargetGet?(u=Sr(n,o))&&u.value:n[o],!wr(s?o:c+(f?".":"#")+o,t.forced)&&void 0!==i){if(r(a)==r(i))continue;pr(a,i)}(t.sham||i&&i.sham)&&Me(a,"sham",!0),ur(n,o,a,t)}},xr=!_((function(){function t(){}return t.prototype.constructor=null,Object.getPrototypeOf(new t)!==t.prototype})),kr=me("IE_PROTO"),Er=A.Object,Tr=Er.prototype,jr=xr?Er.getPrototypeOf:function(t){var e=J(t);if(Z(e,kr))return e[kr];var r=e.constructor;return ot(r)&&e instanceof r?r.prototype:e instanceof Er?Tr:null},Ar=wt("iterator"),_r=!1;[].keys&&("next"in(Ke=[].keys())?($e=jr(jr(Ke)))!==Object.prototype&&(Qe=$e):_r=!0);var Pr=null==Qe||_((function(){var t={};return Qe[Ar].call(t)!==t}));Pr&&(Qe={}),ot(Qe[Ar])||ur(Qe,Ar,(function(){return this}));var Ir={IteratorPrototype:Qe,BUGGY_SAFARI_ITERATORS:_r},Lr=Xt.f,Rr=wt("toStringTag"),Cr=function(t,e,r){t&&!r&&(t=t.prototype),t&&!Z(t,Rr)&&Lr(t,Rr,{configurable:!0,value:e})},Ur=Ir.IteratorPrototype,Dr=function(){return this},Fr=function(t,e,r,n){var o=e+" Iterator";return t.prototype=Ee(Ur,{next:Fe(+!n,r)}),Cr(t,o,!1),_e[o]=Dr,t},Mr=A.String,Br=A.TypeError,Nr=Object.setPrototypeOf||("__proto__"in{}?function(){var t,e=!1,n={};try{(t=U(Object.getOwnPropertyDescriptor(Object.prototype,"__proto__").set))(n,[]),e=n instanceof Array}catch(t){}return function(n,o){return kt(n),function(t){if("object"==r(t)||ot(t))return t;throw Br("Can't set "+Mr(t)+" as a prototype")}(o),e?t(n,o):n.__proto__=o,n}}():void 0),Gr=ir.PROPER,zr=ir.CONFIGURABLE,Wr=Ir.IteratorPrototype,Vr=Ir.BUGGY_SAFARI_ITERATORS,qr=wt("iterator"),Hr="keys",Qr="values",$r="entries",Kr=function(){return this},Yr=Xt.f,Jr="Array Iterator",Xr=Ye.set,Zr=Ye.getterFor(Jr),tn=function(t,e,r,n,o,i,a){Fr(r,e,n);var u,c,s,f=function(t){if(t===o&&y)return y;if(!Vr&&t in p)return p[t];switch(t){case Hr:case Qr:case $r:return function(){return new r(this,t)}}return function(){return new r(this)}},l=e+" Iterator",h=!1,p=t.prototype,d=p[qr]||p["@@iterator"]||o&&p[o],y=!Vr&&d||f(o),v="Array"==e&&p.entries||d;if(v&&(u=jr(v.call(new t)))!==Object.prototype&&u.next&&(jr(u)!==Wr&&(Nr?Nr(u,Wr):ot(u[qr])||ur(u,qr,Kr)),Cr(u,l,!0)),Gr&&o==Qr&&d&&d.name!==Qr&&(zr?Me(p,"name",Qr):(h=!0,y=function(){return Lt(d,this)})),o)if(c={values:f(Qr),keys:i?y:f(Hr),entries:f($r)},a)for(s in c)(Vr||h||!(s in p))&&ur(p,s,c[s]);else Or({target:e,proto:!0,forced:Vr||h},c);return p[qr]!==y&&ur(p,qr,y,{name:o}),_e[e]=y,c}(Array,"Array",(function(t,e){Xr(this,{type:Jr,target:V(t),index:0,kind:e})}),(function(){var t=Zr(this),e=t.target,r=t.kind,n=t.index++;return!e||n>=e.length?(t.target=void 0,{value:void 0,done:!0}):"keys"==r?{value:n,done:!1}:"values"==r?{value:e[n],done:!1}:{value:[n,e[n]],done:!1}}),"values"),en=_e.Arguments=_e.Array;if(Ae("keys"),Ae("values"),Ae("entries"),Et&&"values"!==en.name)try{Yr(en,"name",{value:"values"})}catch(t){}var rn={};rn[wt("toStringTag")]="z";var nn="[object z]"===String(rn),on=wt("toStringTag"),an=A.Object,un="Arguments"==M(function(){return arguments}()),cn=nn?M:function(t){var e,r,n;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(r=function(t,e){try{return t[e]}catch(t){}}(e=an(t),on))?r:un?M(e):"Object"==(n=M(e))&&ot(e.callee)?"Arguments":n},sn=nn?{}.toString:function(){return"[object "+cn(this)+"]"};nn||ur(Object.prototype,"toString",sn,{unsafe:!0});var fn=A.Array,ln=Math.max,hn=function(t,e,r){for(var n,o,i,a,u=ae(t),c=oe(e,u),s=oe(void 0===r?u:r,u),f=fn(ln(s-c,0)),l=0;c<s;c++,l++)n=f,o=l,i=t[c],a=void 0,(a=qt(o))in n?Xt.f(n,a,Fe(0,i)):n[a]=i;return f.length=l,f},pn=sr.f,dn="object"==("undefined"==typeof window?"undefined":r(window))&&window&&Object.getOwnPropertyNames?Object.getOwnPropertyNames(window):[],yn={f:function(t){return dn&&"Window"==M(t)?function(t){try{return pn(t)}catch(t){return hn(dn)}}(t):pn(V(t))}},vn=_((function(){if("function"==typeof ArrayBuffer){var t=new ArrayBuffer(8);Object.isExtensible(t)&&Object.defineProperty(t,"a",{value:8})}})),gn=Object.isExtensible,bn=_((function(){gn(1)}))||vn?function(t){return!!St(t)&&((!vn||"ArrayBuffer"!=M(t))&&(!gn||gn(t)))}:gn,mn=!_((function(){return Object.isExtensible(Object.preventExtensions({}))})),wn=k((function(t){var e=Xt.f,n=!1,o=nt("meta"),i=0,a=function(t){e(t,o,{value:{objectID:"O"+i++,weakData:{}}})},u=t.exports={enable:function(){u.enable=function(){},n=!0;var t=sr.f,e=U([].splice),r={};r[o]=1,t(r).length&&(sr.f=function(r){for(var n=t(r),i=0,a=n.length;i<a;i++)if(n[i]===o){e(n,i,1);break}return n},Or({target:"Object",stat:!0,forced:!0},{getOwnPropertyNames:yn.f}))},fastKey:function(t,e){if(!St(t))return"symbol"==r(t)?t:("string"==typeof t?"S":"P")+t;if(!Z(t,o)){if(!bn(t))return"F";if(!e)return"E";a(t)}return t[o].objectID},getWeakData:function(t,e){if(!Z(t,o)){if(!bn(t))return!0;if(!e)return!1;a(t)}return t[o].weakData},onFreeze:function(t){return mn&&n&&bn(t)&&!Z(t,o)&&a(t),t}};se[o]=!0})),Sn=U(U.bind),On=function(t,e){return Bt(t),void 0===e?t:P?Sn(t,e):function(){return t.apply(e,arguments)}},xn=wt("iterator"),kn=Array.prototype,En=wt("iterator"),Tn=function(t){if(null!=t)return Nt(t,En)||Nt(t,"@@iterator")||_e[cn(t)]},jn=A.TypeError,An=function(t,e){var r=arguments.length<2?Tn(t):e;if(Bt(r))return kt(Lt(r,t));throw jn(Ft(t)+" is not iterable")},_n=function(t,e,r){var n,o;kt(t);try{if(!(n=Nt(t,"return"))){if("throw"===e)throw r;return r}n=Lt(n,t)}catch(t){o=!0,n=t}if("throw"===e)throw r;if(o)throw n;return kt(n),r},Pn=A.TypeError,In=function(t,e){this.stopped=t,this.result=e},Ln=In.prototype,Rn=function(t,e,n){var o,i,a,u,c,s,f,l,h=n&&n.that,p=!(!n||!n.AS_ENTRIES),d=!(!n||!n.IS_ITERATOR),y=!(!n||!n.INTERRUPTED),v=On(e,h),g=function(t){return o&&_n(o,"normal",t),new In(!0,t)},b=function(t){return p?(kt(t),y?v(t[0],t[1],g):v(t[0],t[1])):y?v(t,g):v(t)};if(d)o=t;else{if(!(i=Tn(t)))throw Pn(Ft(t)+" is not iterable");if(void 0!==(l=i)&&(_e.Array===l||kn[xn]===l)){for(a=0,u=ae(t);u>a;a++)if((c=b(t[a]))&&Rt(Ln,c))return c;return new In(!1)}o=An(t,i)}for(s=o.next;!(f=Lt(s,o)).done;){try{c=b(f.value)}catch(t){_n(o,"throw",t)}if("object"==r(c)&&c&&Rt(Ln,c))return c}return new In(!1)},Cn=A.TypeError,Un=function(t,e){if(Rt(e,t))return t;throw Cn("Incorrect invocation")},Dn=wt("iterator"),Fn=!1;try{var Mn=0,Bn={next:function(){return{done:!!Mn++}},return:function(){Fn=!0}};Bn[Dn]=function(){return this},Array.from(Bn,(function(){throw 2}))}catch(t){}var Nn=function(t,e,r){for(var n in e)ur(t,n,e[n],r);return t},Gn=Array.isArray||function(t){return"Array"==M(t)},zn=function(){},Wn=[],Vn=at("Reflect","construct"),qn=/^\s*(?:class|function)\b/,Hn=U(qn.exec),Qn=!qn.exec(zn),$n=function(t){if(!ot(t))return!1;try{return Vn(zn,Wn,t),!0}catch(t){return!1}},Kn=function(t){if(!ot(t))return!1;switch(cn(t)){case"AsyncFunction":case"GeneratorFunction":case"AsyncGeneratorFunction":return!1}try{return Qn||!!Hn(qn,Ce(t))}catch(t){return!0}};Kn.sham=!0;var Yn=!Vn||_((function(){var t;return $n($n.call)||!$n(Object)||!$n((function(){t=!0}))||t}))?Kn:$n,Jn=wt("species"),Xn=A.Array,Zn=function(t,e){return new(function(t){var e;return Gn(t)&&(e=t.constructor,(Yn(e)&&(e===Xn||Gn(e.prototype))||St(e)&&null===(e=e[Jn]))&&(e=void 0)),void 0===e?Xn:e}(t))(0===e?0:e)},to=U([].push),eo=function(t){var e=1==t,r=2==t,n=3==t,o=4==t,i=6==t,a=7==t,u=5==t||i;return function(c,s,f,l){for(var h,p,d=J(c),y=G(d),v=On(s,f),g=ae(y),b=0,m=l||Zn,w=e?m(c,g):r||a?m(c,0):void 0;g>b;b++)if((u||b in y)&&(p=v(h=y[b],b,d),t))if(e)w[b]=p;else if(p)switch(t){case 3:return!0;case 5:return h;case 6:return b;case 2:to(w,h)}else switch(t){case 4:return!1;case 7:to(w,h)}return i?-1:n||o?o:w}},ro={forEach:eo(0),map:eo(1),filter:eo(2),some:eo(3),every:eo(4),find:eo(5),findIndex:eo(6),filterReject:eo(7)},no=wn.getWeakData,oo=Ye.set,io=Ye.getterFor,ao=ro.find,uo=ro.findIndex,co=U([].splice),so=0,fo=function(t){return t.frozen||(t.frozen=new lo)},lo=function(){this.entries=[]},ho=function(t,e){return ao(t.entries,(function(t){return t[0]===e}))};lo.prototype={get:function(t){var e=ho(this,t);if(e)return e[1]},has:function(t){return!!ho(this,t)},set:function(t,e){var r=ho(this,t);r?r[1]=e:this.entries.push([t,e])},delete:function(t){var e=uo(this.entries,(function(e){return e[0]===t}));return~e&&co(this.entries,e,1),!!~e}};var po={getConstructor:function(t,e,r,n){var o=t((function(t,o){Un(t,i),oo(t,{type:e,id:so++,frozen:void 0}),null!=o&&Rn(o,t[n],{that:t,AS_ENTRIES:r})})),i=o.prototype,a=io(e),u=function(t,e,r){var n=a(t),o=no(kt(e),!0);return!0===o?fo(n).set(e,r):o[n.id]=r,t};return Nn(i,{delete:function(t){var e=a(this);if(!St(t))return!1;var r=no(t);return!0===r?fo(e).delete(t):r&&Z(r,e.id)&&delete r[e.id]},has:function(t){var e=a(this);if(!St(t))return!1;var r=no(t);return!0===r?fo(e).has(t):r&&Z(r,e.id)}}),Nn(i,r?{get:function(t){var e=a(this);if(St(t)){var r=no(t);return!0===r?fo(e).get(t):r?r[e.id]:void 0}},set:function(t,e){return u(this,t,e)}}:{add:function(t){return u(this,t,!0)}}),o}};!function(t,e,r){var n=-1!==t.indexOf("Map"),o=-1!==t.indexOf("Weak"),i=n?"set":"add",a=A[t],u=a&&a.prototype,c=a,s={},f=function(t){var e=U(u[t]);ur(u,t,"add"==t?function(t){return e(this,0===t?0:t),this}:"delete"==t?function(t){return!(o&&!St(t))&&e(this,0===t?0:t)}:"get"==t?function(t){return o&&!St(t)?void 0:e(this,0===t?0:t)}:"has"==t?function(t){return!(o&&!St(t))&&e(this,0===t?0:t)}:function(t,r){return e(this,0===t?0:t,r),this})};if(wr(t,!ot(a)||!(o||u.forEach&&!_((function(){(new a).entries().next()})))))c=r.getConstructor(e,t,n,i),wn.enable();else if(wr(t,!0)){var l=new c,h=l[i](o?{}:-0,1)!=l,p=_((function(){l.has(1)})),d=function(t,e){if(!e&&!Fn)return!1;var r=!1;try{var n={};n[Dn]=function(){return{next:function(){return{done:r=!0}}}},t(n)}catch(t){}return r}((function(t){new a(t)})),y=!o&&_((function(){for(var t=new a,e=5;e--;)t[i](e,e);return!t.has(-0)}));d||((c=e((function(t,e){Un(t,u);var r=function(t,e,r){var n,o;return Nr&&ot(n=e.constructor)&&n!==r&&St(o=n.prototype)&&o!==r.prototype&&Nr(t,o),t}(new a,t,c);return null!=e&&Rn(e,r[i],{that:r,AS_ENTRIES:n}),r}))).prototype=u,u.constructor=c),(p||y)&&(f("delete"),f("has"),n&&f("get")),(y||h)&&f(i),o&&u.clear&&delete u.clear}s[t]=c,Or({global:!0,forced:c!=a},s),Cr(c,t),o||r.setStrong(c,t,n)}("WeakSet",(function(t){return function(){return t(this,arguments.length?arguments[0]:void 0)}}),po);var yo={CSSRuleList:0,CSSStyleDeclaration:0,CSSValueList:0,ClientRectList:0,DOMRectList:0,DOMStringList:0,DOMTokenList:1,DataTransferItemList:0,FileList:0,HTMLAllCollection:0,HTMLCollection:0,HTMLFormElement:0,HTMLSelectElement:0,MediaList:0,MimeTypeArray:0,NamedNodeMap:0,NodeList:1,PaintRequestList:0,Plugin:0,PluginArray:0,SVGLengthList:0,SVGNumberList:0,SVGPathSegList:0,SVGPointList:0,SVGStringList:0,SVGTransformList:0,SourceBufferList:0,StyleSheetList:0,TextTrackCueList:0,TextTrackList:0,TouchList:0},vo=_t("span").classList,go=vo&&vo.constructor&&vo.constructor.prototype,bo=go===Object.prototype?void 0:go,mo=wt("iterator"),wo=wt("toStringTag"),So=tn.values,Oo=function(t,e){if(t){if(t[mo]!==So)try{Me(t,mo,So)}catch(e){t[mo]=So}if(t[wo]||Me(t,wo,e),yo[e])for(var r in tn)if(t[r]!==tn[r])try{Me(t,r,tn[r])}catch(e){t[r]=tn[r]}}};for(var xo in yo)Oo(A[xo]&&A[xo].prototype,xo);Oo(bo,"DOMTokenList");var ko=wt("iterator"),Eo=!_((function(){var t=new URL("b?a=1&b=2&c=3","http://a"),e=t.searchParams,r="";return t.pathname="c%20d",e.forEach((function(t,n){e.delete("b"),r+=n+t})),!e.sort||"http://a/c%20d?a=1&c=3"!==t.href||"3"!==e.get("c")||"a=1"!==String(new URLSearchParams("?a=1"))||!e[ko]||"a"!==new URL("https://a@b").username||"b"!==new URLSearchParams(new URLSearchParams("a=b")).get("a")||"xn--e1aybc"!==new URL("http://тест").host||"#%D0%B1"!==new URL("http://a#б").hash||"a1c3"!==r||"x"!==new URL("http://x",void 0).host})),To=A.String,jo=function(t){if("Symbol"===cn(t))throw TypeError("Cannot convert a Symbol value to a string");return To(t)},Ao=A.TypeError,_o=function(t,e){if(t<e)throw Ao("Not enough arguments");return t},Po=Math.floor,Io=function(t,e){for(var r,n,o=t.length,i=1;i<o;){for(n=i,r=t[i];n&&e(t[n-1],r)>0;)t[n]=t[--n];n!==i++&&(t[n]=r)}return t},Lo=function(t,e,r,n){for(var o=e.length,i=r.length,a=0,u=0;a<o||u<i;)t[a+u]=a<o&&u<i?n(e[a],r[u])<=0?e[a++]:r[u++]:a<o?e[a++]:r[u++];return t},Ro=function t(e,r){var n=e.length,o=Po(n/2);return n<8?Io(e,r):Lo(e,t(hn(e,0,o),r),t(hn(e,o),r),r)},Co=wt("iterator"),Uo="URLSearchParams",Do="URLSearchParamsIterator",Fo=Ye.set,Mo=Ye.getterFor(Uo),Bo=Ye.getterFor(Do),No=Object.getOwnPropertyDescriptor,Go=function(t){if(!Et)return A[t];var e=No(A,t);return e&&e.value},zo=Go("fetch"),Wo=Go("Request"),Vo=Go("Headers"),qo=Wo&&Wo.prototype,Ho=Vo&&Vo.prototype,Qo=A.RegExp,$o=A.TypeError,Ko=A.decodeURIComponent,Yo=A.encodeURIComponent,Jo=U("".charAt),Xo=U([].join),Zo=U([].push),ti=U("".replace),ei=U([].shift),ri=U([].splice),ni=U("".split),oi=U("".slice),ii=/\+/g,ai=Array(4),ui=function(t){return ai[t-1]||(ai[t-1]=Qo("((?:%[\\da-f]{2}){"+t+"})","gi"))},ci=function(t){try{return Ko(t)}catch(e){return t}},si=function(t){var e=ti(t,ii," "),r=4;try{return Ko(e)}catch(t){for(;r;)e=ti(e,ui(r--),ci);return e}},fi=/[!'()~]|%20/g,li={"!":"%21","'":"%27","(":"%28",")":"%29","~":"%7E","%20":"+"},hi=function(t){return li[t]},pi=function(t){return ti(Yo(t),fi,hi)},di=Fr((function(t,e){Fo(this,{type:Do,iterator:An(Mo(t).entries),kind:e})}),"Iterator",(function(){var t=Bo(this),e=t.kind,r=t.iterator.next(),n=r.value;return r.done||(r.value="keys"===e?n.key:"values"===e?n.value:[n.key,n.value]),r}),!0),yi=function(t){this.entries=[],this.url=null,void 0!==t&&(St(t)?this.parseObject(t):this.parseQuery("string"==typeof t?"?"===Jo(t,0)?oi(t,1):t:jo(t)))};yi.prototype={type:Uo,bindURL:function(t){this.url=t,this.update()},parseObject:function(t){var e,r,n,o,i,a,u,c=Tn(t);if(c)for(r=(e=An(t,c)).next;!(n=Lt(r,e)).done;){if(i=(o=An(kt(n.value))).next,(a=Lt(i,o)).done||(u=Lt(i,o)).done||!Lt(i,o).done)throw $o("Expected sequence with length 2");Zo(this.entries,{key:jo(a.value),value:jo(u.value)})}else for(var s in t)Z(t,s)&&Zo(this.entries,{key:s,value:jo(t[s])})},parseQuery:function(t){if(t)for(var e,r,n=ni(t,"&"),o=0;o<n.length;)(e=n[o++]).length&&(r=ni(e,"="),Zo(this.entries,{key:si(ei(r)),value:si(Xo(r,"="))}))},serialize:function(){for(var t,e=this.entries,r=[],n=0;n<e.length;)t=e[n++],Zo(r,pi(t.key)+"="+pi(t.value));return Xo(r,"&")},update:function(){this.entries.length=0,this.parseQuery(this.url.query)},updateURL:function(){this.url&&this.url.update()}};var vi=function(){Un(this,gi);var t=arguments.length>0?arguments[0]:void 0;Fo(this,new yi(t))},gi=vi.prototype;if(Nn(gi,{append:function(t,e){_o(arguments.length,2);var r=Mo(this);Zo(r.entries,{key:jo(t),value:jo(e)}),r.updateURL()},delete:function(t){_o(arguments.length,1);for(var e=Mo(this),r=e.entries,n=jo(t),o=0;o<r.length;)r[o].key===n?ri(r,o,1):o++;e.updateURL()},get:function(t){_o(arguments.length,1);for(var e=Mo(this).entries,r=jo(t),n=0;n<e.length;n++)if(e[n].key===r)return e[n].value;return null},getAll:function(t){_o(arguments.length,1);for(var e=Mo(this).entries,r=jo(t),n=[],o=0;o<e.length;o++)e[o].key===r&&Zo(n,e[o].value);return n},has:function(t){_o(arguments.length,1);for(var e=Mo(this).entries,r=jo(t),n=0;n<e.length;)if(e[n++].key===r)return!0;return!1},set:function(t,e){_o(arguments.length,1);for(var r,n=Mo(this),o=n.entries,i=!1,a=jo(t),u=jo(e),c=0;c<o.length;c++)(r=o[c]).key===a&&(i?ri(o,c--,1):(i=!0,r.value=u));i||Zo(o,{key:a,value:u}),n.updateURL()},sort:function(){var t=Mo(this);Ro(t.entries,(function(t,e){return t.key>e.key?1:-1})),t.updateURL()},forEach:function(t){for(var e,r=Mo(this).entries,n=On(t,arguments.length>1?arguments[1]:void 0),o=0;o<r.length;)n((e=r[o++]).value,e.key,this)},keys:function(){return new di(this,"keys")},values:function(){return new di(this,"values")},entries:function(){return new di(this,"entries")}},{enumerable:!0}),ur(gi,Co,gi.entries,{name:"entries"}),ur(gi,"toString",(function(){return Mo(this).serialize()}),{enumerable:!0}),Cr(vi,Uo),Or({global:!0,forced:!Eo},{URLSearchParams:vi}),!Eo&&ot(Vo)){var bi=U(Ho.has),mi=U(Ho.set),wi=function(t){if(St(t)){var e,r=t.body;if(cn(r)===Uo)return e=t.headers?new Vo(t.headers):new Vo,bi(e,"content-type")||mi(e,"content-type","application/x-www-form-urlencoded;charset=UTF-8"),Ee(t,{body:Fe(0,jo(r)),headers:Fe(0,e)})}return t};if(ot(zo)&&Or({global:!0,enumerable:!0,noTargetGet:!0,forced:!0},{fetch:function(t){return zo(t,arguments.length>1?wi(arguments[1]):{})}}),ot(Wo)){var Si=function(t){return Un(this,qo),new Wo(t,arguments.length>1?wi(arguments[1]):{})};qo.constructor=Si,Si.prototype=qo,Or({global:!0,forced:!0,noTargetGet:!0},{Request:Si})}}k((function(t){var e=function(t){var e,r=Object.prototype,n=r.hasOwnProperty,o="function"==typeof Symbol?Symbol:{},i=o.iterator||"@@iterator",a=o.asyncIterator||"@@asyncIterator",u=o.toStringTag||"@@toStringTag";function c(t,e,r){return Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{c({},"")}catch(t){c=function(t,e,r){return t[e]=r}}function s(t,e,r,n){var o=e&&e.prototype instanceof v?e:v,i=Object.create(o.prototype),a=new A(n||[]);return i._invoke=function(t,e,r){var n=l;return function(o,i){if(n===p)throw new Error("Generator is already running");if(n===d){if("throw"===o)throw i;return P()}for(r.method=o,r.arg=i;;){var a=r.delegate;if(a){var u=E(a,r);if(u){if(u===y)continue;return u}}if("next"===r.method)r.sent=r._sent=r.arg;else if("throw"===r.method){if(n===l)throw n=d,r.arg;r.dispatchException(r.arg)}else"return"===r.method&&r.abrupt("return",r.arg);n=p;var c=f(t,e,r);if("normal"===c.type){if(n=r.done?d:h,c.arg===y)continue;return{value:c.arg,done:r.done}}"throw"===c.type&&(n=d,r.method="throw",r.arg=c.arg)}}}(t,r,a),i}function f(t,e,r){try{return{type:"normal",arg:t.call(e,r)}}catch(t){return{type:"throw",arg:t}}}t.wrap=s;var l="suspendedStart",h="suspendedYield",p="executing",d="completed",y={};function v(){}function g(){}function b(){}var m={};c(m,i,(function(){return this}));var w=Object.getPrototypeOf,S=w&&w(w(_([])));S&&S!==r&&n.call(S,i)&&(m=S);var O=b.prototype=v.prototype=Object.create(m);function x(t){["next","throw","return"].forEach((function(e){c(t,e,(function(t){return this._invoke(e,t)}))}))}function k(t,e){function r(o,i,a,u){var c=f(t[o],t,i);if("throw"!==c.type){var s=c.arg,l=s.value;return l&&"object"==typeof l&&n.call(l,"__await")?e.resolve(l.__await).then((function(t){r("next",t,a,u)}),(function(t){r("throw",t,a,u)})):e.resolve(l).then((function(t){s.value=t,a(s)}),(function(t){return r("throw",t,a,u)}))}u(c.arg)}var o;this._invoke=function(t,n){function i(){return new e((function(e,o){r(t,n,e,o)}))}return o=o?o.then(i,i):i()}}function E(t,r){var n=t.iterator[r.method];if(n===e){if(r.delegate=null,"throw"===r.method){if(t.iterator.return&&(r.method="return",r.arg=e,E(t,r),"throw"===r.method))return y;r.method="throw",r.arg=new TypeError("The iterator does not provide a 'throw' method")}return y}var o=f(n,t.iterator,r.arg);if("throw"===o.type)return r.method="throw",r.arg=o.arg,r.delegate=null,y;var i=o.arg;return i?i.done?(r[t.resultName]=i.value,r.next=t.nextLoc,"return"!==r.method&&(r.method="next",r.arg=e),r.delegate=null,y):i:(r.method="throw",r.arg=new TypeError("iterator result is not an object"),r.delegate=null,y)}function T(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function j(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function A(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(T,this),this.reset(!0)}function _(t){if(t){var r=t[i];if(r)return r.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var o=-1,a=function r(){for(;++o<t.length;)if(n.call(t,o))return r.value=t[o],r.done=!1,r;return r.value=e,r.done=!0,r};return a.next=a}}return{next:P}}function P(){return{value:e,done:!0}}return g.prototype=b,c(O,"constructor",b),c(b,"constructor",g),g.displayName=c(b,u,"GeneratorFunction"),t.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===g||"GeneratorFunction"===(e.displayName||e.name))},t.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,b):(t.__proto__=b,c(t,u,"GeneratorFunction")),t.prototype=Object.create(O),t},t.awrap=function(t){return{__await:t}},x(k.prototype),c(k.prototype,a,(function(){return this})),t.AsyncIterator=k,t.async=function(e,r,n,o,i){void 0===i&&(i=Promise);var a=new k(s(e,r,n,o),i);return t.isGeneratorFunction(r)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},x(O),c(O,u,"Generator"),c(O,i,(function(){return this})),c(O,"toString",(function(){return"[object Generator]"})),t.keys=function(t){var e=[];for(var r in t)e.push(r);return e.reverse(),function r(){for(;e.length;){var n=e.pop();if(n in t)return r.value=n,r.done=!1,r}return r.done=!0,r}},t.values=_,A.prototype={constructor:A,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=e,this.done=!1,this.delegate=null,this.method="next",this.arg=e,this.tryEntries.forEach(j),!t)for(var r in this)"t"===r.charAt(0)&&n.call(this,r)&&!isNaN(+r.slice(1))&&(this[r]=e)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var r=this;function o(n,o){return u.type="throw",u.arg=t,r.next=n,o&&(r.method="next",r.arg=e),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var a=this.tryEntries[i],u=a.completion;if("root"===a.tryLoc)return o("end");if(a.tryLoc<=this.prev){var c=n.call(a,"catchLoc"),s=n.call(a,"finallyLoc");if(c&&s){if(this.prev<a.catchLoc)return o(a.catchLoc,!0);if(this.prev<a.finallyLoc)return o(a.finallyLoc)}else if(c){if(this.prev<a.catchLoc)return o(a.catchLoc,!0)}else{if(!s)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return o(a.finallyLoc)}}}},abrupt:function(t,e){for(var r=this.tryEntries.length-1;r>=0;--r){var o=this.tryEntries[r];if(o.tryLoc<=this.prev&&n.call(o,"finallyLoc")&&this.prev<o.finallyLoc){var i=o;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,y):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),y},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.finallyLoc===t)return this.complete(r.completion,r.afterLoc),j(r),y}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.tryLoc===t){var n=r.completion;if("throw"===n.type){var o=n.arg;j(r)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,r,n){return this.delegate={iterator:_(t),resultName:r,nextLoc:n},"next"===this.method&&(this.arg=e),y}},t}(t.exports);try{regeneratorRuntime=e}catch(t){"object"==typeof globalThis?globalThis.regeneratorRuntime=e:Function("r","regeneratorRuntime = r")(e)}}));var Oi="undefined"!=typeof globalThis&&globalThis||"undefined"!=typeof self&&self||void 0!==Oi&&Oi,xi="URLSearchParams"in Oi,ki="Symbol"in Oi&&"iterator"in Symbol,Ei="FileReader"in Oi&&"Blob"in Oi&&function(){try{return new Blob,!0}catch(t){return!1}}(),Ti="FormData"in Oi,ji="ArrayBuffer"in Oi;if(ji)var Ai=["[object Int8Array]","[object Uint8Array]","[object Uint8ClampedArray]","[object Int16Array]","[object Uint16Array]","[object Int32Array]","[object Uint32Array]","[object Float32Array]","[object Float64Array]"],_i=ArrayBuffer.isView||function(t){return t&&Ai.indexOf(Object.prototype.toString.call(t))>-1};function Pi(t){if("string"!=typeof t&&(t=String(t)),/[^a-z0-9\-#$%&'*+.^_`|~!]/i.test(t)||""===t)throw new TypeError('Invalid character in header field name: "'+t+'"');return t.toLowerCase()}function Ii(t){return"string"!=typeof t&&(t=String(t)),t}function Li(t){var e={next:function(){var e=t.shift();return{done:void 0===e,value:e}}};return ki&&(e[Symbol.iterator]=function(){return e}),e}function Ri(t){this.map={},t instanceof Ri?t.forEach((function(t,e){this.append(e,t)}),this):Array.isArray(t)?t.forEach((function(t){this.append(t[0],t[1])}),this):t&&Object.getOwnPropertyNames(t).forEach((function(e){this.append(e,t[e])}),this)}function Ci(t){if(t.bodyUsed)return Promise.reject(new TypeError("Already read"));t.bodyUsed=!0}function Ui(t){return new Promise((function(e,r){t.onload=function(){e(t.result)},t.onerror=function(){r(t.error)}}))}function Di(t){var e=new FileReader,r=Ui(e);return e.readAsArrayBuffer(t),r}function Fi(t){if(t.slice)return t.slice(0);var e=new Uint8Array(t.byteLength);return e.set(new Uint8Array(t)),e.buffer}function Mi(){return this.bodyUsed=!1,this._initBody=function(t){var e;this.bodyUsed=this.bodyUsed,this._bodyInit=t,t?"string"==typeof t?this._bodyText=t:Ei&&Blob.prototype.isPrototypeOf(t)?this._bodyBlob=t:Ti&&FormData.prototype.isPrototypeOf(t)?this._bodyFormData=t:xi&&URLSearchParams.prototype.isPrototypeOf(t)?this._bodyText=t.toString():ji&&Ei&&((e=t)&&DataView.prototype.isPrototypeOf(e))?(this._bodyArrayBuffer=Fi(t.buffer),this._bodyInit=new Blob([this._bodyArrayBuffer])):ji&&(ArrayBuffer.prototype.isPrototypeOf(t)||_i(t))?this._bodyArrayBuffer=Fi(t):this._bodyText=t=Object.prototype.toString.call(t):this._bodyText="",this.headers.get("content-type")||("string"==typeof t?this.headers.set("content-type","text/plain;charset=UTF-8"):this._bodyBlob&&this._bodyBlob.type?this.headers.set("content-type",this._bodyBlob.type):xi&&URLSearchParams.prototype.isPrototypeOf(t)&&this.headers.set("content-type","application/x-www-form-urlencoded;charset=UTF-8"))},Ei&&(this.blob=function(){var t=Ci(this);if(t)return t;if(this._bodyBlob)return Promise.resolve(this._bodyBlob);if(this._bodyArrayBuffer)return Promise.resolve(new Blob([this._bodyArrayBuffer]));if(this._bodyFormData)throw new Error("could not read FormData body as blob");return Promise.resolve(new Blob([this._bodyText]))},this.arrayBuffer=function(){if(this._bodyArrayBuffer){var t=Ci(this);return t||(ArrayBuffer.isView(this._bodyArrayBuffer)?Promise.resolve(this._bodyArrayBuffer.buffer.slice(this._bodyArrayBuffer.byteOffset,this._bodyArrayBuffer.byteOffset+this._bodyArrayBuffer.byteLength)):Promise.resolve(this._bodyArrayBuffer))}return this.blob().then(Di)}),this.text=function(){var t,e,r,n=Ci(this);if(n)return n;if(this._bodyBlob)return t=this._bodyBlob,e=new FileReader,r=Ui(e),e.readAsText(t),r;if(this._bodyArrayBuffer)return Promise.resolve(function(t){for(var e=new Uint8Array(t),r=new Array(e.length),n=0;n<e.length;n++)r[n]=String.fromCharCode(e[n]);return r.join("")}(this._bodyArrayBuffer));if(this._bodyFormData)throw new Error("could not read FormData body as text");return Promise.resolve(this._bodyText)},Ti&&(this.formData=function(){return this.text().then(Gi)}),this.json=function(){return this.text().then(JSON.parse)},this}Ri.prototype.append=function(t,e){t=Pi(t),e=Ii(e);var r=this.map[t];this.map[t]=r?r+", "+e:e},Ri.prototype.delete=function(t){delete this.map[Pi(t)]},Ri.prototype.get=function(t){return t=Pi(t),this.has(t)?this.map[t]:null},Ri.prototype.has=function(t){return this.map.hasOwnProperty(Pi(t))},Ri.prototype.set=function(t,e){this.map[Pi(t)]=Ii(e)},Ri.prototype.forEach=function(t,e){for(var r in this.map)this.map.hasOwnProperty(r)&&t.call(e,this.map[r],r,this)},Ri.prototype.keys=function(){var t=[];return this.forEach((function(e,r){t.push(r)})),Li(t)},Ri.prototype.values=function(){var t=[];return this.forEach((function(e){t.push(e)})),Li(t)},Ri.prototype.entries=function(){var t=[];return this.forEach((function(e,r){t.push([r,e])})),Li(t)},ki&&(Ri.prototype[Symbol.iterator]=Ri.prototype.entries);var Bi=["DELETE","GET","HEAD","OPTIONS","POST","PUT"];function Ni(t,e){if(!(this instanceof Ni))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');var r,n,o=(e=e||{}).body;if(t instanceof Ni){if(t.bodyUsed)throw new TypeError("Already read");this.url=t.url,this.credentials=t.credentials,e.headers||(this.headers=new Ri(t.headers)),this.method=t.method,this.mode=t.mode,this.signal=t.signal,o||null==t._bodyInit||(o=t._bodyInit,t.bodyUsed=!0)}else this.url=String(t);if(this.credentials=e.credentials||this.credentials||"same-origin",!e.headers&&this.headers||(this.headers=new Ri(e.headers)),this.method=(r=e.method||this.method||"GET",n=r.toUpperCase(),Bi.indexOf(n)>-1?n:r),this.mode=e.mode||this.mode||null,this.signal=e.signal||this.signal,this.referrer=null,("GET"===this.method||"HEAD"===this.method)&&o)throw new TypeError("Body not allowed for GET or HEAD requests");if(this._initBody(o),!("GET"!==this.method&&"HEAD"!==this.method||"no-store"!==e.cache&&"no-cache"!==e.cache)){var i=/([?&])_=[^&]*/;if(i.test(this.url))this.url=this.url.replace(i,"$1_="+(new Date).getTime());else{this.url+=(/\?/.test(this.url)?"&":"?")+"_="+(new Date).getTime()}}}function Gi(t){var e=new FormData;return t.trim().split("&").forEach((function(t){if(t){var r=t.split("="),n=r.shift().replace(/\+/g," "),o=r.join("=").replace(/\+/g," ");e.append(decodeURIComponent(n),decodeURIComponent(o))}})),e}function zi(t,e){if(!(this instanceof zi))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');e||(e={}),this.type="default",this.status=void 0===e.status?200:e.status,this.ok=this.status>=200&&this.status<300,this.statusText=void 0===e.statusText?"":""+e.statusText,this.headers=new Ri(e.headers),this.url=e.url||"",this._initBody(t)}Ni.prototype.clone=function(){return new Ni(this,{body:this._bodyInit})},Mi.call(Ni.prototype),Mi.call(zi.prototype),zi.prototype.clone=function(){return new zi(this._bodyInit,{status:this.status,statusText:this.statusText,headers:new Ri(this.headers),url:this.url})},zi.error=function(){var t=new zi(null,{status:0,statusText:""});return t.type="error",t};var Wi=[301,302,303,307,308];zi.redirect=function(t,e){if(-1===Wi.indexOf(e))throw new RangeError("Invalid status code");return new zi(null,{status:e,headers:{location:t}})};var Vi,qi=Oi.DOMException;try{new qi}catch(t){(qi=function(t,e){this.message=t,this.name=e;var r=Error(t);this.stack=r.stack}).prototype=Object.create(Error.prototype),qi.prototype.constructor=qi}function Hi(t,e){return new Promise((function(r,n){var o=new Ni(t,e);if(o.signal&&o.signal.aborted)return n(new qi("Aborted","AbortError"));var i=new XMLHttpRequest;function a(){i.abort()}i.onload=function(){var t,e,n={status:i.status,statusText:i.statusText,headers:(t=i.getAllResponseHeaders()||"",e=new Ri,t.replace(/\r?\n[\t ]+/g," ").split("\r").map((function(t){return 0===t.indexOf("\n")?t.substr(1,t.length):t})).forEach((function(t){var r=t.split(":"),n=r.shift().trim();if(n){var o=r.join(":").trim();e.append(n,o)}})),e)};n.url="responseURL"in i?i.responseURL:n.headers.get("X-Request-URL");var o="response"in i?i.response:i.responseText;setTimeout((function(){r(new zi(o,n))}),0)},i.onerror=function(){setTimeout((function(){n(new TypeError("Network request failed"))}),0)},i.ontimeout=function(){setTimeout((function(){n(new TypeError("Network request failed"))}),0)},i.onabort=function(){setTimeout((function(){n(new qi("Aborted","AbortError"))}),0)},i.open(o.method,function(t){try{return""===t&&Oi.location.href?Oi.location.href:t}catch(e){return t}}(o.url),!0),"include"===o.credentials?i.withCredentials=!0:"omit"===o.credentials&&(i.withCredentials=!1),"responseType"in i&&(Ei?i.responseType="blob":ji&&o.headers.get("Content-Type")&&-1!==o.headers.get("Content-Type").indexOf("application/octet-stream")&&(i.responseType="arraybuffer")),!e||"object"!=typeof e.headers||e.headers instanceof Ri?o.headers.forEach((function(t,e){i.setRequestHeader(e,t)})):Object.getOwnPropertyNames(e.headers).forEach((function(t){i.setRequestHeader(t,Ii(e.headers[t]))})),o.signal&&(o.signal.addEventListener("abort",a),i.onreadystatechange=function(){4===i.readyState&&o.signal.removeEventListener("abort",a)}),i.send(void 0===o._bodyInit?null:o._bodyInit)}))}Hi.polyfill=!0,Oi.fetch||(Oi.fetch=Hi,Oi.Headers=Ri,Oi.Request=Ni,Oi.Response=zi);var Qi=new Uint8Array(16);function $i(){if(!Vi&&!(Vi="undefined"!=typeof crypto&&crypto.getRandomValues&&crypto.getRandomValues.bind(crypto)||"undefined"!=typeof msCrypto&&"function"==typeof msCrypto.getRandomValues&&msCrypto.getRandomValues.bind(msCrypto)))throw new Error("crypto.getRandomValues() not supported. See https://github.com/uuidjs/uuid#getrandomvalues-not-supported");return Vi(Qi)}var Ki=/^(?:[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|00000000-0000-0000-0000-000000000000)$/i;function Yi(t){return"string"==typeof t&&Ki.test(t)}for(var Ji=[],Xi=0;Xi<256;++Xi)Ji.push((Xi+256).toString(16).substr(1));function Zi(t,e,r){var n=(t=t||{}).random||(t.rng||$i)();if(n[6]=15&n[6]|64,n[8]=63&n[8]|128,e){r=r||0;for(var o=0;o<16;++o)e[r+o]=n[o];return e}return function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:0,r=(Ji[t[e+0]]+Ji[t[e+1]]+Ji[t[e+2]]+Ji[t[e+3]]+"-"+Ji[t[e+4]]+Ji[t[e+5]]+"-"+Ji[t[e+6]]+Ji[t[e+7]]+"-"+Ji[t[e+8]]+Ji[t[e+9]]+"-"+Ji[t[e+10]]+Ji[t[e+11]]+Ji[t[e+12]]+Ji[t[e+13]]+Ji[t[e+14]]+Ji[t[e+15]]).toLowerCase();if(!Yi(r))throw TypeError("Stringified UUID is invalid");return r}(n)}
 /*! js-cookie v3.0.1 | MIT */function ta(t){for(var e=1;e<arguments.length;e++){var r=arguments[e];for(var n in r)t[n]=r[n]}return t}var ea=function t(e,r){function n(t,n,o){if("undefined"!=typeof document){"number"==typeof(o=ta({},r,o)).expires&&(o.expires=new Date(Date.now()+864e5*o.expires)),o.expires&&(o.expires=o.expires.toUTCString()),t=encodeURIComponent(t).replace(/%(2[346B]|5E|60|7C)/g,decodeURIComponent).replace(/[()]/g,escape);var i="";for(var a in o)o[a]&&(i+="; "+a,!0!==o[a]&&(i+="="+o[a].split(";")[0]));return document.cookie=t+"="+e.write(n,t)+i}}return Object.create({set:n,get:function(t){if("undefined"!=typeof document&&(!arguments.length||t)){for(var r=document.cookie?document.cookie.split("; "):[],n={},o=0;o<r.length;o++){var i=r[o].split("="),a=i.slice(1).join("=");try{var u=decodeURIComponent(i[0]);if(n[u]=e.read(a,u),t===u)break}catch(t){}}return t?n[t]:n}},remove:function(t,e){n(t,"",ta({},e,{expires:-1}))},withAttributes:function(e){return t(this.converter,ta({},this.attributes,e))},withConverter:function(e){return t(ta({},this.converter,e),this.attributes)}},{attributes:{value:Object.freeze(r)},converter:{value:Object.freeze(e)}})}({read:function(t){return'"'===t[0]&&(t=t.slice(1,-1)),t.replace(/(%[\dA-F]{2})+/gi,decodeURIComponent)},write:function(t){return encodeURIComponent(t).replace(/%(2[346BF]|3[AC-F]|40|5[BDE]|60|7[BCD])/g,decodeURIComponent)}},{path:"/"});
-/*! @timeone-group/error-logger-js 0.2.3 https://github.com/https://github.com/TimeOne-Group/error-logger-js#readme @license GPL-3.0 */class ra extends Error{constructor(t,...e){super(...e),this.severity=t}}var na=1,oa=2;const ia=(t,e="Logger")=>{if(t instanceof ra)if(t.severity===oa)console.warn(`[${e}] WARNING ${t.message}`);else console.error(`[${e}] ERROR ${t.message}`);else t instanceof Error?console.error(`[${e}] ERROR(js-core) ${t.message}`):console.error(`[${e}] ERROR Unabled to treat error`,t)};var aa=Math.ceil,ua=Math.floor,sa=function(t){return isNaN(t=+t)?0:(t>0?ua:aa)(t)},ca=function(t){if(null==t)throw TypeError("Can't call method on "+t);return t},fa=function(t){return function(e,r){var n,o,i=String(ca(e)),a=sa(r),u=i.length;return a<0||a>=u?t?"":void 0:(n=i.charCodeAt(a))<55296||n>56319||a+1===u||(o=i.charCodeAt(a+1))<56320||o>57343?t?i.charAt(a):n:t?i.slice(a,a+2):o-56320+(n-55296<<10)+65536}},la={codeAt:fa(!1),charAt:fa(!0)},ha="undefined"!=typeof globalThis?globalThis:"undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:{};function pa(t){var e={exports:{}};return t(e,e.exports),e.exports}var da=function(t){return t&&t.Math==Math&&t},ya=da("object"==("undefined"==typeof globalThis?"undefined":r(globalThis))&&globalThis)||da("object"==("undefined"==typeof window?"undefined":r(window))&&window)||da("object"==("undefined"==typeof self?"undefined":r(self))&&self)||da("object"==r(ha)&&ha)||function(){return this}()||Function("return this")(),va=function(t){try{return!!t()}catch(t){return!0}},ga=!va((function(){return 7!=Object.defineProperty({},1,{get:function(){return 7}})[1]})),ma=function(t){return"object"===r(t)?null!==t:"function"==typeof t},ba=ya.document,wa=ma(ba)&&ma(ba.createElement),Sa=function(t){return wa?ba.createElement(t):{}},Oa=!ga&&!va((function(){return 7!=Object.defineProperty(Sa("div"),"a",{get:function(){return 7}}).a})),xa=function(t){if(!ma(t))throw TypeError(String(t)+" is not an object");return t},ka=function(t,e){if(!ma(t))return t;var r,n;if(e&&"function"==typeof(r=t.toString)&&!ma(n=r.call(t)))return n;if("function"==typeof(r=t.valueOf)&&!ma(n=r.call(t)))return n;if(!e&&"function"==typeof(r=t.toString)&&!ma(n=r.call(t)))return n;throw TypeError("Can't convert object to primitive value")},Ta=Object.defineProperty,Ea={f:ga?Ta:function(t,e,r){if(xa(t),e=ka(e,!0),xa(r),Oa)try{return Ta(t,e,r)}catch(t){}if("get"in r||"set"in r)throw TypeError("Accessors not supported");return"value"in r&&(t[e]=r.value),t}},ja=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}},Aa=ga?function(t,e,r){return Ea.f(t,e,ja(1,r))}:function(t,e,r){return t[e]=r,t},_a=function(t,e){try{Aa(ya,t,e)}catch(r){ya[t]=e}return e},Pa="__core-js_shared__",Ia=ya[Pa]||_a(Pa,{}),La=Function.toString;"function"!=typeof Ia.inspectSource&&(Ia.inspectSource=function(t){return La.call(t)});var Ra,Ca,Ua,Da=Ia.inspectSource,Fa=ya.WeakMap,Ma="function"==typeof Fa&&/native code/.test(Da(Fa)),Na={}.hasOwnProperty,Ba=function(t,e){return Na.call(t,e)},Ga=pa((function(t){(t.exports=function(t,e){return Ia[t]||(Ia[t]=void 0!==e?e:{})})("versions",[]).push({version:"3.9.1",mode:"global",copyright:"© 2021 Denis Pushkarev (zloirock.ru)"})})),Wa=0,za=Math.random(),qa=function(t){return"Symbol("+String(void 0===t?"":t)+")_"+(++Wa+za).toString(36)},Va=Ga("keys"),Ha=function(t){return Va[t]||(Va[t]=qa(t))},Qa={},$a=ya.WeakMap;if(Ma){var Ja=Ia.state||(Ia.state=new $a),Ka=Ja.get,Ya=Ja.has,Xa=Ja.set;Ra=function(t,e){return e.facade=t,Xa.call(Ja,t,e),e},Ca=function(t){return Ka.call(Ja,t)||{}},Ua=function(t){return Ya.call(Ja,t)}}else{var Za=Ha("state");Qa[Za]=!0,Ra=function(t,e){return e.facade=t,Aa(t,Za,e),e},Ca=function(t){return Ba(t,Za)?t[Za]:{}},Ua=function(t){return Ba(t,Za)}}var tu,eu,ru={set:Ra,get:Ca,has:Ua,enforce:function(t){return Ua(t)?Ca(t):Ra(t,{})},getterFor:function(t){return function(e){var r;if(!ma(e)||(r=Ca(e)).type!==t)throw TypeError("Incompatible receiver, "+t+" required");return r}}},nu={}.propertyIsEnumerable,ou=Object.getOwnPropertyDescriptor,iu={f:ou&&!nu.call({1:2},1)?function(t){var e=ou(this,t);return!!e&&e.enumerable}:nu},au={}.toString,uu=function(t){return au.call(t).slice(8,-1)},su="".split,cu=va((function(){return!Object("z").propertyIsEnumerable(0)}))?function(t){return"String"==uu(t)?su.call(t,""):Object(t)}:Object,fu=function(t){return cu(ca(t))},lu=Object.getOwnPropertyDescriptor,hu={f:ga?lu:function(t,e){if(t=fu(t),e=ka(e,!0),Oa)try{return lu(t,e)}catch(t){}if(Ba(t,e))return ja(!iu.f.call(t,e),t[e])}},pu=pa((function(t){var e=ru.get,r=ru.enforce,n=String(String).split("String");(t.exports=function(t,e,o,i){var a,u=!!i&&!!i.unsafe,s=!!i&&!!i.enumerable,c=!!i&&!!i.noTargetGet;"function"==typeof o&&("string"!=typeof e||Ba(o,"name")||Aa(o,"name",e),(a=r(o)).source||(a.source=n.join("string"==typeof e?e:""))),t!==ya?(u?!c&&t[e]&&(s=!0):delete t[e],s?t[e]=o:Aa(t,e,o)):s?t[e]=o:_a(e,o)})(Function.prototype,"toString",(function(){return"function"==typeof this&&e(this).source||Da(this)}))})),du=ya,yu=function(t){return"function"==typeof t?t:void 0},vu=function(t,e){return arguments.length<2?yu(du[t])||yu(ya[t]):du[t]&&du[t][e]||ya[t]&&ya[t][e]},gu=Math.min,mu=function(t){return t>0?gu(sa(t),9007199254740991):0},bu=Math.max,wu=Math.min,Su=function(t){return function(e,r,n){var o,i=fu(e),a=mu(i.length),u=function(t,e){var r=sa(t);return r<0?bu(r+e,0):wu(r,e)}(n,a);if(t&&r!=r){for(;a>u;)if((o=i[u++])!=o)return!0}else for(;a>u;u++)if((t||u in i)&&i[u]===r)return t||u||0;return!t&&-1}},Ou={includes:Su(!0),indexOf:Su(!1)}.indexOf,xu=function(t,e){var r,n=fu(t),o=0,i=[];for(r in n)!Ba(Qa,r)&&Ba(n,r)&&i.push(r);for(;e.length>o;)Ba(n,r=e[o++])&&(~Ou(i,r)||i.push(r));return i},ku=["constructor","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","toLocaleString","toString","valueOf"],Tu=ku.concat("length","prototype"),Eu={f:Object.getOwnPropertyNames||function(t){return xu(t,Tu)}},ju={f:Object.getOwnPropertySymbols},Au=vu("Reflect","ownKeys")||function(t){var e=Eu.f(xa(t)),r=ju.f;return r?e.concat(r(t)):e},_u=function(t,e){for(var r=Au(e),n=Ea.f,o=hu.f,i=0;i<r.length;i++){var a=r[i];Ba(t,a)||n(t,a,o(e,a))}},Pu=/#|\.prototype\./,Iu=function(t,e){var r=Ru[Lu(t)];return r==Uu||r!=Cu&&("function"==typeof e?va(e):!!e)},Lu=Iu.normalize=function(t){return String(t).replace(Pu,".").toLowerCase()},Ru=Iu.data={},Cu=Iu.NATIVE="N",Uu=Iu.POLYFILL="P",Du=Iu,Fu=hu.f,Mu=function(t,e){var n,o,i,a,u,s=t.target,c=t.global,f=t.stat;if(n=c?ya:f?ya[s]||_a(s,{}):(ya[s]||{}).prototype)for(o in e){if(a=e[o],i=t.noTargetGet?(u=Fu(n,o))&&u.value:n[o],!Du(c?o:s+(f?".":"#")+o,t.forced)&&void 0!==i){if(r(a)===r(i))continue;_u(a,i)}(t.sham||i&&i.sham)&&Aa(a,"sham",!0),pu(n,o,a,t)}},Nu=function(t){return Object(ca(t))},Bu=!va((function(){function t(){}return t.prototype.constructor=null,Object.getPrototypeOf(new t)!==t.prototype})),Gu=Ha("IE_PROTO"),Wu=Object.prototype,zu=Bu?Object.getPrototypeOf:function(t){return t=Nu(t),Ba(t,Gu)?t[Gu]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?Wu:null},qu="process"==uu(ya.process),Vu=vu("navigator","userAgent")||"",Hu=ya.process,Qu=Hu&&Hu.versions,$u=Qu&&Qu.v8;$u?eu=(tu=$u.split("."))[0]+tu[1]:Vu&&(!(tu=Vu.match(/Edge\/(\d+)/))||tu[1]>=74)&&(tu=Vu.match(/Chrome\/(\d+)/))&&(eu=tu[1]);var Ju,Ku,Yu,Xu=eu&&+eu,Zu=!!Object.getOwnPropertySymbols&&!va((function(){return!Symbol.sham&&(qu?38===Xu:Xu>37&&Xu<41)})),ts=Zu&&!Symbol.sham&&"symbol"==r(Symbol.iterator),es=Ga("wks"),rs=ya.Symbol,ns=ts?rs:rs&&rs.withoutSetter||qa,os=function(t){return Ba(es,t)&&(Zu||"string"==typeof es[t])||(Zu&&Ba(rs,t)?es[t]=rs[t]:es[t]=ns("Symbol."+t)),es[t]},is=os("iterator"),as=!1;[].keys&&("next"in(Yu=[].keys())?(Ku=zu(zu(Yu)))!==Object.prototype&&(Ju=Ku):as=!0);var us=null==Ju||va((function(){var t={};return Ju[is].call(t)!==t}));us&&(Ju={}),Ba(Ju,is)||Aa(Ju,is,(function(){return this}));var ss,cs={IteratorPrototype:Ju,BUGGY_SAFARI_ITERATORS:as},fs=Object.keys||function(t){return xu(t,ku)},ls=ga?Object.defineProperties:function(t,e){xa(t);for(var r,n=fs(e),o=n.length,i=0;o>i;)Ea.f(t,r=n[i++],e[r]);return t},hs=vu("document","documentElement"),ps=Ha("IE_PROTO"),ds=function(){},ys=function(t){return"<script>"+t+"</"+"script>"},vs=function(){try{ss=document.domain&&new ActiveXObject("htmlfile")}catch(t){}var t,e;vs=ss?function(t){t.write(ys("")),t.close();var e=t.parentWindow.Object;return t=null,e}(ss):((e=Sa("iframe")).style.display="none",hs.appendChild(e),e.src=String("javascript:"),(t=e.contentWindow.document).open(),t.write(ys("document.F=Object")),t.close(),t.F);for(var r=ku.length;r--;)delete vs.prototype[ku[r]];return vs()};Qa[ps]=!0;var gs=Object.create||function(t,e){var r;return null!==t?(ds.prototype=xa(t),r=new ds,ds.prototype=null,r[ps]=t):r=vs(),void 0===e?r:ls(r,e)},ms=Ea.f,bs=os("toStringTag"),ws=function(t,e,r){t&&!Ba(t=r?t:t.prototype,bs)&&ms(t,bs,{configurable:!0,value:e})},Ss={},Os=cs.IteratorPrototype,xs=function(){return this},ks=Object.setPrototypeOf||("__proto__"in{}?function(){var t,e=!1,r={};try{(t=Object.getOwnPropertyDescriptor(Object.prototype,"__proto__").set).call(r,[]),e=r instanceof Array}catch(t){}return function(r,n){return xa(r),function(t){if(!ma(t)&&null!==t)throw TypeError("Can't set "+String(t)+" as a prototype")}(n),e?t.call(r,n):r.__proto__=n,r}}():void 0),Ts=cs.IteratorPrototype,Es=cs.BUGGY_SAFARI_ITERATORS,js=os("iterator"),As="keys",_s="values",Ps="entries",Is=function(){return this},Ls=function(t,e,r,n,o,i,a){!function(t,e,r){var n=e+" Iterator";t.prototype=gs(Os,{next:ja(1,r)}),ws(t,n,!1),Ss[n]=xs}(r,e,n);var u,s,c,f=function(t){if(t===o&&y)return y;if(!Es&&t in p)return p[t];switch(t){case As:case _s:case Ps:return function(){return new r(this,t)}}return function(){return new r(this)}},l=e+" Iterator",h=!1,p=t.prototype,d=p[js]||p["@@iterator"]||o&&p[o],y=!Es&&d||f(o),v="Array"==e&&p.entries||d;if(v&&(u=zu(v.call(new t)),Ts!==Object.prototype&&u.next&&(zu(u)!==Ts&&(ks?ks(u,Ts):"function"!=typeof u[js]&&Aa(u,js,Is)),ws(u,l,!0))),o==_s&&d&&d.name!==_s&&(h=!0,y=function(){return d.call(this)}),p[js]!==y&&Aa(p,js,y),Ss[e]=y,o)if(s={values:f(_s),keys:i?y:f(As),entries:f(Ps)},a)for(c in s)(Es||h||!(c in p))&&pu(p,c,s[c]);else Mu({target:e,proto:!0,forced:Es||h},s);return s},Rs=la.charAt,Cs="String Iterator",Us=ru.set,Ds=ru.getterFor(Cs);Ls(String,"String",(function(t){Us(this,{type:Cs,string:String(t),index:0})}),(function(){var t,e=Ds(this),r=e.string,n=e.index;return n>=r.length?{value:void 0,done:!0}:(t=Rs(r,n),e.index+=t.length,{value:t,done:!1})}));var Fs=function(t,e,r){if(function(t){if("function"!=typeof t)throw TypeError(String(t)+" is not a function")}(t),void 0===e)return t;switch(r){case 0:return function(){return t.call(e)};case 1:return function(r){return t.call(e,r)};case 2:return function(r,n){return t.call(e,r,n)};case 3:return function(r,n,o){return t.call(e,r,n,o)}}return function(){return t.apply(e,arguments)}},Ms=function(t){var e=t.return;if(void 0!==e)return xa(e.call(t)).value},Ns=function(t,e,r,n){try{return n?e(xa(r)[0],r[1]):e(r)}catch(e){throw Ms(t),e}},Bs=os("iterator"),Gs=Array.prototype,Ws=function(t){return void 0!==t&&(Ss.Array===t||Gs[Bs]===t)},zs=function(t,e,r){var n=ka(e);n in t?Ea.f(t,n,ja(0,r)):t[n]=r},qs={};qs[os("toStringTag")]="z";var Vs="[object z]"===String(qs),Hs=os("toStringTag"),Qs="Arguments"==uu(function(){return arguments}()),$s=Vs?uu:function(t){var e,r,n;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(r=function(t,e){try{return t[e]}catch(t){}}(e=Object(t),Hs))?r:Qs?uu(e):"Object"==(n=uu(e))&&"function"==typeof e.callee?"Arguments":n},Js=os("iterator"),Ks=function(t){if(null!=t)return t[Js]||t["@@iterator"]||Ss[$s(t)]},Ys=os("iterator"),Xs=!1;try{var Zs=0,tc={next:function(){return{done:!!Zs++}},return:function(){Xs=!0}};tc[Ys]=function(){return this},Array.from(tc,(function(){throw 2}))}catch(t){}var ec=function(t,e){if(!e&&!Xs)return!1;var r=!1;try{var n={};n[Ys]=function(){return{next:function(){return{done:r=!0}}}},t(n)}catch(t){}return r},rc=!ec((function(t){Array.from(t)}));Mu({target:"Array",stat:!0,forced:rc},{from:function(t){var e,r,n,o,i,a,u=Nu(t),s="function"==typeof this?this:Array,c=arguments.length,f=c>1?arguments[1]:void 0,l=void 0!==f,h=Ks(u),p=0;if(l&&(f=Fs(f,c>2?arguments[2]:void 0,2)),null==h||s==Array&&Ws(h))for(r=new s(e=mu(u.length));e>p;p++)a=l?f(u[p],p):u[p],zs(r,p,a);else for(i=(o=h.call(u)).next,r=new s;!(n=i.call(o)).done;p++)a=l?Ns(o,f,[n.value,p],!0):n.value,zs(r,p,a);return r.length=p,r}}),du.Array.from;var nc=!va((function(){return Object.isExtensible(Object.preventExtensions({}))})),oc=pa((function(t){var e=Ea.f,n=qa("meta"),o=0,i=Object.isExtensible||function(){return!0},a=function(t){e(t,n,{value:{objectID:"O"+ ++o,weakData:{}}})},u=t.exports={REQUIRED:!1,fastKey:function(t,e){if(!ma(t))return"symbol"==r(t)?t:("string"==typeof t?"S":"P")+t;if(!Ba(t,n)){if(!i(t))return"F";if(!e)return"E";a(t)}return t[n].objectID},getWeakData:function(t,e){if(!Ba(t,n)){if(!i(t))return!0;if(!e)return!1;a(t)}return t[n].weakData},onFreeze:function(t){return nc&&u.REQUIRED&&i(t)&&!Ba(t,n)&&a(t),t}};Qa[n]=!0})),ic=function(t,e){this.stopped=t,this.result=e},ac=function(t,e,n){var o,i,a,u,s,c,f,l=n&&n.that,h=!(!n||!n.AS_ENTRIES),p=!(!n||!n.IS_ITERATOR),d=!(!n||!n.INTERRUPTED),y=Fs(e,l,1+h+d),v=function(t){return o&&Ms(o),new ic(!0,t)},g=function(t){return h?(xa(t),d?y(t[0],t[1],v):y(t[0],t[1])):d?y(t,v):y(t)};if(p)o=t;else{if("function"!=typeof(i=Ks(t)))throw TypeError("Target is not iterable");if(Ws(i)){for(a=0,u=mu(t.length);u>a;a++)if((s=g(t[a]))&&s instanceof ic)return s;return new ic(!1)}o=i.call(t)}for(c=o.next;!(f=c.call(o)).done;){try{s=g(f.value)}catch(t){throw Ms(o),t}if("object"==r(s)&&s&&s instanceof ic)return s}return new ic(!1)},uc=function(t,e,r){if(!(t instanceof e))throw TypeError("Incorrect "+(r?r+" ":"")+"invocation");return t},sc=function(t,e,r){for(var n in e)pu(t,n,e[n],r);return t},cc=os("species"),fc=Ea.f,lc=oc.fastKey,hc=ru.set,pc=ru.getterFor,dc={getConstructor:function(t,e,r,n){var o=t((function(t,i){uc(t,o,e),hc(t,{type:e,index:gs(null),first:void 0,last:void 0,size:0}),ga||(t.size=0),null!=i&&ac(i,t[n],{that:t,AS_ENTRIES:r})})),i=pc(e),a=function(t,e,r){var n,o,a=i(t),s=u(t,e);return s?s.value=r:(a.last=s={index:o=lc(e,!0),key:e,value:r,previous:n=a.last,next:void 0,removed:!1},a.first||(a.first=s),n&&(n.next=s),ga?a.size++:t.size++,"F"!==o&&(a.index[o]=s)),t},u=function(t,e){var r,n=i(t),o=lc(e);if("F"!==o)return n.index[o];for(r=n.first;r;r=r.next)if(r.key==e)return r};return sc(o.prototype,{clear:function(){for(var t=i(this),e=t.index,r=t.first;r;)r.removed=!0,r.previous&&(r.previous=r.previous.next=void 0),delete e[r.index],r=r.next;t.first=t.last=void 0,ga?t.size=0:this.size=0},delete:function(t){var e=this,r=i(e),n=u(e,t);if(n){var o=n.next,a=n.previous;delete r.index[n.index],n.removed=!0,a&&(a.next=o),o&&(o.previous=a),r.first==n&&(r.first=o),r.last==n&&(r.last=a),ga?r.size--:e.size--}return!!n},forEach:function(t){for(var e,r=i(this),n=Fs(t,arguments.length>1?arguments[1]:void 0,3);e=e?e.next:r.first;)for(n(e.value,e.key,this);e&&e.removed;)e=e.previous},has:function(t){return!!u(this,t)}}),sc(o.prototype,r?{get:function(t){var e=u(this,t);return e&&e.value},set:function(t,e){return a(this,0===t?0:t,e)}}:{add:function(t){return a(this,t=0===t?0:t,t)}}),ga&&fc(o.prototype,"size",{get:function(){return i(this).size}}),o},setStrong:function(t,e,r){var n=e+" Iterator",o=pc(e),i=pc(n);Ls(t,e,(function(t,e){hc(this,{type:n,target:t,state:o(t),kind:e,last:void 0})}),(function(){for(var t=i(this),e=t.kind,r=t.last;r&&r.removed;)r=r.previous;return t.target&&(t.last=r=r?r.next:t.state.first)?"keys"==e?{value:r.key,done:!1}:"values"==e?{value:r.value,done:!1}:{value:[r.key,r.value],done:!1}:(t.target=void 0,{value:void 0,done:!0})}),r?"entries":"values",!r,!0),function(t){var e=vu(t),r=Ea.f;ga&&e&&!e[cc]&&r(e,cc,{configurable:!0,get:function(){return this}})}(e)}};!function(t,e,r){var n=-1!==t.indexOf("Map"),o=-1!==t.indexOf("Weak"),i=n?"set":"add",a=ya[t],u=a&&a.prototype,s=a,c={},f=function(t){var e=u[t];pu(u,t,"add"==t?function(t){return e.call(this,0===t?0:t),this}:"delete"==t?function(t){return!(o&&!ma(t))&&e.call(this,0===t?0:t)}:"get"==t?function(t){return o&&!ma(t)?void 0:e.call(this,0===t?0:t)}:"has"==t?function(t){return!(o&&!ma(t))&&e.call(this,0===t?0:t)}:function(t,r){return e.call(this,0===t?0:t,r),this})};if(Du(t,"function"!=typeof a||!(o||u.forEach&&!va((function(){(new a).entries().next()})))))s=r.getConstructor(e,t,n,i),oc.REQUIRED=!0;else if(Du(t,!0)){var l=new s,h=l[i](o?{}:-0,1)!=l,p=va((function(){l.has(1)})),d=ec((function(t){new a(t)})),y=!o&&va((function(){for(var t=new a,e=5;e--;)t[i](e,e);return!t.has(-0)}));d||((s=e((function(e,r){uc(e,s,t);var o=function(t,e,r){var n,o;return ks&&"function"==typeof(n=e.constructor)&&n!==r&&ma(o=n.prototype)&&o!==r.prototype&&ks(t,o),t}(new a,e,s);return null!=r&&ac(r,o[i],{that:o,AS_ENTRIES:n}),o}))).prototype=u,u.constructor=s),(p||y)&&(f("delete"),f("has"),n&&f("get")),(y||h)&&f(i),o&&u.clear&&delete u.clear}c[t]=s,Mu({global:!0,forced:s!=a},c),ws(s,t),o||r.setStrong(s,t,n)}("Set",(function(t){return function(){return t(this,arguments.length?arguments[0]:void 0)}}),dc);var yc=Vs?{}.toString:function(){return"[object "+$s(this)+"]"};Vs||pu(Object.prototype,"toString",yc,{unsafe:!0});var vc={CSSRuleList:0,CSSStyleDeclaration:0,CSSValueList:0,ClientRectList:0,DOMRectList:0,DOMStringList:0,DOMTokenList:1,DataTransferItemList:0,FileList:0,HTMLAllCollection:0,HTMLCollection:0,HTMLFormElement:0,HTMLSelectElement:0,MediaList:0,MimeTypeArray:0,NamedNodeMap:0,NodeList:1,PaintRequestList:0,Plugin:0,PluginArray:0,SVGLengthList:0,SVGNumberList:0,SVGPathSegList:0,SVGPointList:0,SVGStringList:0,SVGTransformList:0,SourceBufferList:0,StyleSheetList:0,TextTrackCueList:0,TextTrackList:0,TouchList:0},gc=os("unscopables"),mc=Array.prototype;null==mc[gc]&&Ea.f(mc,gc,{configurable:!0,value:gs(null)});var bc=function(t){mc[gc][t]=!0},wc="Array Iterator",Sc=ru.set,Oc=ru.getterFor(wc),xc=Ls(Array,"Array",(function(t,e){Sc(this,{type:wc,target:fu(t),index:0,kind:e})}),(function(){var t=Oc(this),e=t.target,r=t.kind,n=t.index++;return!e||n>=e.length?(t.target=void 0,{value:void 0,done:!0}):"keys"==r?{value:n,done:!1}:"values"==r?{value:e[n],done:!1}:{value:[n,e[n]],done:!1}}),"values");Ss.Arguments=Ss.Array,bc("keys"),bc("values"),bc("entries");var kc=os("iterator"),Tc=os("toStringTag"),Ec=xc.values;for(var jc in vc){var Ac=ya[jc],_c=Ac&&Ac.prototype;if(_c){if(_c[kc]!==Ec)try{Aa(_c,kc,Ec)}catch(t){_c[kc]=Ec}if(_c[Tc]||Aa(_c,Tc,jc),vc[jc])for(var Pc in xc)if(_c[Pc]!==xc[Pc])try{Aa(_c,Pc,xc[Pc])}catch(t){_c[Pc]=xc[Pc]}}}du.Set;var Ic,Lc,Rc,Cc=function(){function t(e){i(this,t),this.prefix=e}return u(t,[{key:"get",value:function(t){return"".concat(this.prefix,"_").concat(t)}},{key:"getIndex",value:function(){return this.get("INDEX")}}]),t}(),Uc=pa((function(t){var e=function(){var t=String.fromCharCode,e="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=",r="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+-$",n={};function o(t,e){if(!n[t]){n[t]={};for(var r=0;r<t.length;r++)n[t][t.charAt(r)]=r}return n[t][e]}var i={compressToBase64:function(t){if(null==t)return"";var r=i._compress(t,6,(function(t){return e.charAt(t)}));switch(r.length%4){default:case 0:return r;case 1:return r+"===";case 2:return r+"==";case 3:return r+"="}},decompressFromBase64:function(t){return null==t?"":""==t?null:i._decompress(t.length,32,(function(r){return o(e,t.charAt(r))}))},compressToUTF16:function(e){return null==e?"":i._compress(e,15,(function(e){return t(e+32)}))+" "},decompressFromUTF16:function(t){return null==t?"":""==t?null:i._decompress(t.length,16384,(function(e){return t.charCodeAt(e)-32}))},compressToUint8Array:function(t){for(var e=i.compress(t),r=new Uint8Array(2*e.length),n=0,o=e.length;n<o;n++){var a=e.charCodeAt(n);r[2*n]=a>>>8,r[2*n+1]=a%256}return r},decompressFromUint8Array:function(e){if(null==e)return i.decompress(e);for(var r=new Array(e.length/2),n=0,o=r.length;n<o;n++)r[n]=256*e[2*n]+e[2*n+1];var a=[];return r.forEach((function(e){a.push(t(e))})),i.decompress(a.join(""))},compressToEncodedURIComponent:function(t){return null==t?"":i._compress(t,6,(function(t){return r.charAt(t)}))},decompressFromEncodedURIComponent:function(t){return null==t?"":""==t?null:(t=t.replace(/ /g,"+"),i._decompress(t.length,32,(function(e){return o(r,t.charAt(e))})))},compress:function(e){return i._compress(e,16,(function(e){return t(e)}))},_compress:function(t,e,r){if(null==t)return"";var n,o,i,a={},u={},s="",c="",f="",l=2,h=3,p=2,d=[],y=0,v=0;for(i=0;i<t.length;i+=1)if(s=t.charAt(i),Object.prototype.hasOwnProperty.call(a,s)||(a[s]=h++,u[s]=!0),c=f+s,Object.prototype.hasOwnProperty.call(a,c))f=c;else{if(Object.prototype.hasOwnProperty.call(u,f)){if(f.charCodeAt(0)<256){for(n=0;n<p;n++)y<<=1,v==e-1?(v=0,d.push(r(y)),y=0):v++;for(o=f.charCodeAt(0),n=0;n<8;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1}else{for(o=1,n=0;n<p;n++)y=y<<1|o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o=0;for(o=f.charCodeAt(0),n=0;n<16;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1}0==--l&&(l=Math.pow(2,p),p++),delete u[f]}else for(o=a[f],n=0;n<p;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1;0==--l&&(l=Math.pow(2,p),p++),a[c]=h++,f=String(s)}if(""!==f){if(Object.prototype.hasOwnProperty.call(u,f)){if(f.charCodeAt(0)<256){for(n=0;n<p;n++)y<<=1,v==e-1?(v=0,d.push(r(y)),y=0):v++;for(o=f.charCodeAt(0),n=0;n<8;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1}else{for(o=1,n=0;n<p;n++)y=y<<1|o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o=0;for(o=f.charCodeAt(0),n=0;n<16;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1}0==--l&&(l=Math.pow(2,p),p++),delete u[f]}else for(o=a[f],n=0;n<p;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1;0==--l&&(l=Math.pow(2,p),p++)}for(o=2,n=0;n<p;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1;for(;;){if(y<<=1,v==e-1){d.push(r(y));break}v++}return d.join("")},decompress:function(t){return null==t?"":""==t?null:i._decompress(t.length,32768,(function(e){return t.charCodeAt(e)}))},_decompress:function(e,r,n){var o,i,a,u,s,c,f,l=[],h=4,p=4,d=3,y="",v=[],g={val:n(0),position:r,index:1};for(o=0;o<3;o+=1)l[o]=o;for(a=0,s=Math.pow(2,2),c=1;c!=s;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*c,c<<=1;switch(a){case 0:for(a=0,s=Math.pow(2,8),c=1;c!=s;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*c,c<<=1;f=t(a);break;case 1:for(a=0,s=Math.pow(2,16),c=1;c!=s;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*c,c<<=1;f=t(a);break;case 2:return""}for(l[3]=f,i=f,v.push(f);;){if(g.index>e)return"";for(a=0,s=Math.pow(2,d),c=1;c!=s;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*c,c<<=1;switch(f=a){case 0:for(a=0,s=Math.pow(2,8),c=1;c!=s;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*c,c<<=1;l[p++]=t(a),f=p-1,h--;break;case 1:for(a=0,s=Math.pow(2,16),c=1;c!=s;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*c,c<<=1;l[p++]=t(a),f=p-1,h--;break;case 2:return v.join("")}if(0==h&&(h=Math.pow(2,d),d++),l[f])y=l[f];else{if(f!==p)return null;y=i+i.charAt(0)}v.push(y),l[p++]=i+y.charAt(0),i=y,0==--h&&(h=Math.pow(2,d),d++)}}};return i}();null!=t&&(t.exports=e)})),Dc={},Fc={setItem:function(t,e){return Dc[t]=e,!0},getItem:function(t){return Dc[t]},removeItem:function(t){return delete Dc[t],!0}},Mc=function(){function t(e){switch(i(this,t),e){case"localStorage":case"sessionStorage":this.engine=window[e];break;case"InApp":this.engine=Fc;break;default:throw new ra(na,"Unknow engine")}}return u(t,[{key:"set",value:function(t,e){this.engine.setItem(t,Uc.compressToBase64(JSON.stringify(e)))}},{key:"get",value:function(t){var e=this.engine.getItem(t);if(e)try{return JSON.parse(Uc.decompressFromBase64(e))}catch(t){return ia(t),null}return null}},{key:"delete",value:function(t){return this.engine.removeItem(t)}}]),t}(),Nc="localStorage",Bc="storage-js",Gc=function(){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},r=e.storageEngine,n=e.defaultTTL,o=e.prefix;i(this,t),this.store=new Mc(r||Nc),this.defaultTTL=n||0,this.key=new Cc(o||Bc)}return u(t,[{key:"save",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=e.ttl;if(!t.id)throw new ra(na,"Object must have an id property");var n={object:t},o=(new Date).getTime();r?n.ttl=o+1e3*r:this.defaultTTL>0&&(n.ttl=o+1e3*this.defaultTTL),this.store.set(this.key.get(t.id),n),this.addToIndex(t.id)}},{key:"find",value:function(t){var e=this.store.get(this.key.get(t));return e&&(!e.ttl||(new Date).getTime()<=e.ttl)?e.object:{}}},{key:"findAll",value:function(){var t=this;return Array.from(this.loadIndex()).map((function(e){return t.find(e)})).filter((function(t){return!!t}))}},{key:"delete",value:function(t){return this.removeToIndex(t),this.store.delete(this.key.get(t))}},{key:"loadIndex",value:function(){var t=this.store.get(this.key.getIndex());return t?new Set(t):new Set}},{key:"addToIndex",value:function(t){var e=this.loadIndex();e.add(t),this.store.set(this.key.getIndex(),f(e))}},{key:"removeToIndex",value:function(t){var e=this.loadIndex();e.delete(t),this.store.set(this.key.getIndex(),f(e))}},{key:"resetIndex",value:function(){this.store.delete(this.key.getIndex())}}]),t}(),Wc={sdkName:"__ISDK",sdkScriptId:"__ISDK_ASSETS",consent:{name:"consent",ttl:390,status:{unknown:"unknown",optin:"optin",optout:"optout"}},event_consent_id:{name:"event_consent_id",ttl:390},subid:{name:"subid",queryname:"toSubid",ttl:40},cashback:{name:"cashback",queryname:"toCashback",ttl:30},default_storage_prefix:"to",default_ttl:390,urls:{conversion:(null===(Ic="https://back.service.sandbox.localhost/v1/b")?void 0:Ic.split(","))||[],statsConsent:(null===(Lc="https://tracking.service.sandbox.localhost/v1/log/consent")?void 0:Lc.split(","))||[],proofConsent:(null===(Rc="https://tracking.service.sandbox.localhost/v1/log/consent/proof")?void 0:Rc.split(","))||[]}},zc=regeneratorRuntime.mark(Jc),qc=new Gc({storageEngine:"localStorage",prefix:Wc.default_storage_prefix,defaultTTL:Wc.consent.ttl});function Vc(t){return"".concat(Wc.default_storage_prefix,"_").concat(t)}function Hc(t,e){var r=Wc[e]||{name:e,ttl:Wc.default_ttl};ea.set(Vc(r.name),t,{expires:r.ttl,sameSite:"strict"}),qc.save({id:e,value:t})}function Qc(t){var e,r=(null===(e=Wc[t])||void 0===e?void 0:e.name)||t,n=ea.get(Vc(r));if(n)return n;var o=qc.find(r);return null!=o&&o.value?null==o?void 0:o.value:null}function $c(t){var e,r=(null===(e=Wc[t])||void 0===e?void 0:e.name)||t;ea.remove(Vc(r)),qc.delete(r)}function Jc(){var t,e,r=arguments;return regeneratorRuntime.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:t=r.length>0&&void 0!==r[0]?r[0]:[],e=0;case 2:if(!(e<=t.length)){n.next=8;break}return n.next=5,t[e];case 5:e+=1,n.next=2;break;case 8:case"end":return n.stop()}}),zc)}var Kc=new WeakMap,Yc=new WeakMap,Xc=new WeakMap,Zc=new WeakMap,tf=new WeakMap,ef=new WeakMap,rf=new WeakMap,nf=new WeakMap,of=new WeakSet,af=new WeakSet,uf=new WeakSet,sf=new WeakSet,cf=new WeakSet,ff=new WeakSet,lf=new WeakSet,hf=new WeakSet,pf=new WeakSet,df=new WeakSet,yf=new WeakSet,vf=new WeakSet,gf=new WeakSet,mf=new WeakSet,bf=function(){function t(){i(this,t),O(this,mf),O(this,gf),O(this,vf),O(this,yf),O(this,df),O(this,pf),O(this,hf),O(this,lf),O(this,ff),O(this,cf),O(this,sf),O(this,uf),O(this,af),O(this,of),S(this,Kc,{writable:!0,value:[]}),S(this,Yc,{writable:!0,value:Jc(Wc.urls.conversion)}),S(this,Xc,{writable:!0,value:Jc(Wc.urls.statsConsent)}),S(this,Zc,{writable:!0,value:Jc(Wc.urls.proofConsent)}),S(this,tf,{writable:!0,value:v(this,Yc).next().value}),S(this,ef,{writable:!0,value:v(this,Xc).next().value}),S(this,rf,{writable:!0,value:v(this,Zc).next().value}),S(this,nf,{writable:!0,value:[]}),this.env="sandbox",b(this,of,wf).call(this),b(this,af,Sf).call(this,Wc.cashback),this.consent||b(this,pf,Af).call(this,Wc.consent.status.unknown),this.consent===Wc.consent.status.optin?b(this,af,Sf).call(this,Wc.subid):b(this,df,_f).call(this)}var e,r,n,a;return u(t,[{key:"consent",get:function(){return Qc(Wc.consent.name)}},{key:"subid",get:function(){return Qc(Wc.subid.name)||this.constructor.getProgramDataFromQueryParams(Wc.subid.queryname)}},{key:"eventConsentId",get:function(){return Qc(Wc.event_consent_id.name)}},{key:"cashbackSubid",get:function(){return Qc(Wc.cashback.name)||this.constructor.getProgramDataFromQueryParams(Wc.cashback.queryname)}},{key:"_setOptin",value:function(){b(this,pf,Af).call(this,Wc.consent.status.optin),b(this,af,Sf).call(this,Wc.subid)}},{key:"_setOptout",value:function(){b(this,pf,Af).call(this,Wc.consent.status.optout),b(this,df,_f).call(this)}},{key:"_setUnknown",value:function(){b(this,pf,Af).call(this,Wc.consent.status.unknown),b(this,df,_f).call(this)}},{key:"_setSale",value:(a=o(regeneratorRuntime.mark((function t(e){return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,b(this,mf,Rf).call(this,{data:e,method:"_setSale"});case 3:t.next=8;break;case 5:t.prev=5,t.t0=t.catch(0),b(this,vf,If).call(this,{error:t.t0,method:"_setSale",extra:e});case 8:case"end":return t.stop()}}),t,this,[[0,5]])}))),function(t){return a.apply(this,arguments)})},{key:"_setLead",value:(n=o(regeneratorRuntime.mark((function t(e){return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,b(this,mf,Rf).call(this,{data:e,method:"_setLead"});case 3:t.next=8;break;case 5:t.prev=5,t.t0=t.catch(0),b(this,vf,If).call(this,{error:t.t0,method:"_setLead",extra:e});case 8:case"end":return t.stop()}}),t,this,[[0,5]])}))),function(t){return n.apply(this,arguments)})},{key:"_setDbClick",value:(r=o(regeneratorRuntime.mark((function t(e){return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,b(this,mf,Rf).call(this,{data:e,method:"_setDbClick"});case 3:t.next=8;break;case 5:t.prev=5,t.t0=t.catch(0),b(this,vf,If).call(this,{error:t.t0,method:"_setDbClick",extra:e});case 8:case"end":return t.stop()}}),t,this,[[0,5]])}))),function(t){return r.apply(this,arguments)})},{key:"_setClick",value:(e=o(regeneratorRuntime.mark((function t(e){return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,b(this,mf,Rf).call(this,{data:e,method:"_setClick"});case 3:t.next=8;break;case 5:t.prev=5,t.t0=t.catch(0),b(this,vf,If).call(this,{error:t.t0,method:"_setClick",extra:e});case 8:case"end":return t.stop()}}),t,this,[[0,5]])}))),function(t){return e.apply(this,arguments)})},{key:"push",value:function(t){var e,r=l(e=t||[])||h(e)||p(e)||y(),n=r[0],o=r.slice(1);try{if("function"!=typeof this[n])throw new Error("Undefined function ".concat(n));this[n].apply(this,f(o))}catch(t){b(this,vf,If).call(this,{error:t,method:n})}}},{key:"getTrace",value:function(){return{env:this.env,progids:v(this,Kc),consent:this.consent,subid:this.subid,event_consent_id:this.eventConsentId,cashbackSubid:this.cashbackSubid,errors:b(this,gf,Lf).call(this),conversionUrls:Wc.urls.conversion}}}],[{key:"getProgramDataFromQueryParams",value:function(t){return new URLSearchParams(window.location.search).get(t)}}]),t}();function wf(){try{var t,e=null===(t=document.getElementById(Wc.sdkScriptId))||void 0===t?void 0:t.getAttribute("data-progids");e&&g(this,Kc,JSON.parse(e))}catch(t){b(this,vf,If).call(this,{error:t,method:"setProgids"})}}function Sf(t){var e=t.name,r=t.queryname,n=this.constructor.getProgramDataFromQueryParams(r);n&&Hc(n,e)}function Of(){g(this,tf,v(this,Yc).next().value)}function xf(){g(this,ef,v(this,Xc).next().value)}function kf(){g(this,rf,v(this,Zc).next().value)}function Tf(t){var r=this,n=t.url,i=t.setNextUrl,a=t.method,u=t.body;n?fetch(n,{method:a,body:JSON.stringify(e({},u))}).then(function(){var t=o(regeneratorRuntime.mark((function t(e){var n;return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(e.ok){t.next=5;break}return t.next=3,e.json();case 3:n=t.sent,b(r,vf,If).call(r,{error:{message:"".concat(null==e?void 0:e.status," - ").concat(null==n?void 0:n.message)},method:"#log"});case 5:case"end":return t.stop()}}),t)})));return function(e){return t.apply(this,arguments)}}()).catch((function(){i(),b(r,ff,Tf).call(r,{url:n,setNextUrl:i,body:u})})):b(this,vf,If).call(this,{error:{message:"Failed to contact server on ".concat(n)},method:"#log"})}function Ef(t){var e=this,r=[this.constructor.getProgramDataFromQueryParams(Wc.subid.queryname),this.constructor.getProgramDataFromQueryParams(Wc.cashback.queryname)].filter(Boolean);v(this,Kc).forEach((function(n){b(e,ff,Tf).call(e,{method:"POST",url:v(e,ef),setNextUrl:b(e,sf,xf),body:{progid:n,status:t,toSubids:r}})}))}function jf(){var t=Zi();Hc(t,Wc.event_consent_id.name);var e={event_consent_id:t};b(this,ff,Tf).call(this,{method:"POST",url:v(this,rf),setNextUrl:b(this,cf,kf),body:e})}function Af(t){var e=t!==this.consent,r=t===Wc.consent.status.optin;Hc(t,Wc.consent.name),e&&b(this,lf,Ef).call(this,t),e&&r&&b(this,hf,jf).call(this)}function _f(){$c(Wc.subid.name),$c(Wc.event_consent_id.name)}function Pf(){return this.cashbackSubid||this.constructor.getProgramDataFromQueryParams(Wc.subid.queryname)||this.subid&&this.consent===Wc.consent.status.optin}function If(t){v(this,nf).push(t)}function Lf(){return v(this,nf).map((function(t){var e=t.error,r=t.method,n=t.extra;return{message:'While calling "'.concat(r,'" method: ').concat(e.message),extra:n}}))}function Rf(t){return Cf.apply(this,arguments)}function Cf(){return Cf=o(regeneratorRuntime.mark((function t(r){var n,o,i,a,u,s,f,l,h,p,d,y,g,m;return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(n=r.data,o=void 0===n?{}:n,i=r.method,a=void 0===i?"setConversion":i,b(this,yf,Pf).call(this)){t.next=3;break}throw new Error("Make a conversion is not allowed. Check consent or ".concat(Wc.subid.queryname));case 3:if(u=o.progid,s=o.comid,f=o.iu,u&&s&&f){t.next=6;break}throw new Error("Missing progid or comid or iu. Those data are mandatory to make a conversion");case 6:if(v(this,tf)){t.next=8;break}throw new Error("Failed to contact server on ".concat(JSON.stringify(Wc.urls.conversion)));case 8:return l={type:"consent",value:this.subid},h={type:"cashback",value:this.cashbackSubid},p=[l,h].filter((function(t){return!!t.value})),d=e(e({},o),{},{event_consent_id:this.eventConsentId,toSubids:p}),y=Object.fromEntries(Object.entries(d).filter((function(t){return!!c(t,2)[1]}))),t.prev=13,t.next=16,fetch(v(this,tf),{method:"POST",headers:{accept:"application/json","Content-Type":"application/json"},body:JSON.stringify(y)});case 16:if((g=t.sent).ok){t.next=22;break}return t.next=20,g.json();case 20:m=t.sent,b(this,vf,If).call(this,{error:m,method:a,extra:e(e({},o),{},{url:v(this,tf)})});case 22:t.next=30;break;case 24:return t.prev=24,t.t0=t.catch(13),b(this,vf,If).call(this,{error:t.t0,method:a,extra:e(e({},o),{},{url:v(this,tf)})}),b(this,uf,Of).call(this),t.next=30,b(this,mf,Rf).call(this,{data:o,method:a});case 30:case"end":return t.stop()}}),t,this,[[13,24]])}))),Cf.apply(this,arguments)}var Uf=Wc.sdkName;var Df=Wc.sdkName;window.TOG_Storage=qc,function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:bf;try{if(void 0===window[Uf])window[Uf]=new t;else if("object"===r(window[Uf])&&Array.isArray(window[Uf])){var e=new t;window[Uf].forEach((function(t){return e.push(t)})),window[Uf]=e}else if(!("object"===r(window[Uf])&&window[Uf]instanceof t))throw new Error("".concat(Uf," error: Unknown type"))}catch(t){console.error(t.message)}}(bf),function t(e){if(document.getElementById("debug")){var r,n,o,i,a,u={progid:null===(r=window[Df])||void 0===r?void 0:r.progid,consent:null===(n=window[Df])||void 0===n?void 0:n.consent,subid:null===(o=window[Df])||void 0===o?void 0:o.subid,cashback:null===(i=window[Df])||void 0===i?void 0:i.cashback,to_event_consent_id:null===(a=window[Df])||void 0===a?void 0:a.eventConsentId};document.getElementById("debug").textContent=JSON.stringify(u,null,"\t")}else setTimeout((function(){return t(e)}),300);null==e||e()}(),console.log("SDK loaded")}();
+/*! @timeone-group/error-logger-js 0.2.3 https://github.com/https://github.com/TimeOne-Group/error-logger-js#readme @license GPL-3.0 */class ra extends Error{constructor(t,...e){super(...e),this.severity=t}}var na=1,oa=2;const ia=(t,e="Logger")=>{if(t instanceof ra)if(t.severity===oa)console.warn(`[${e}] WARNING ${t.message}`);else console.error(`[${e}] ERROR ${t.message}`);else t instanceof Error?console.error(`[${e}] ERROR(js-core) ${t.message}`):console.error(`[${e}] ERROR Unabled to treat error`,t)};var aa=Math.ceil,ua=Math.floor,ca=function(t){return isNaN(t=+t)?0:(t>0?ua:aa)(t)},sa=function(t){if(null==t)throw TypeError("Can't call method on "+t);return t},fa=function(t){return function(e,r){var n,o,i=String(sa(e)),a=ca(r),u=i.length;return a<0||a>=u?t?"":void 0:(n=i.charCodeAt(a))<55296||n>56319||a+1===u||(o=i.charCodeAt(a+1))<56320||o>57343?t?i.charAt(a):n:t?i.slice(a,a+2):o-56320+(n-55296<<10)+65536}},la={codeAt:fa(!1),charAt:fa(!0)},ha="undefined"!=typeof globalThis?globalThis:"undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:{};function pa(t){var e={exports:{}};return t(e,e.exports),e.exports}var da=function(t){return t&&t.Math==Math&&t},ya=da("object"==("undefined"==typeof globalThis?"undefined":r(globalThis))&&globalThis)||da("object"==("undefined"==typeof window?"undefined":r(window))&&window)||da("object"==("undefined"==typeof self?"undefined":r(self))&&self)||da("object"==r(ha)&&ha)||function(){return this}()||Function("return this")(),va=function(t){try{return!!t()}catch(t){return!0}},ga=!va((function(){return 7!=Object.defineProperty({},1,{get:function(){return 7}})[1]})),ba=function(t){return"object"===r(t)?null!==t:"function"==typeof t},ma=ya.document,wa=ba(ma)&&ba(ma.createElement),Sa=function(t){return wa?ma.createElement(t):{}},Oa=!ga&&!va((function(){return 7!=Object.defineProperty(Sa("div"),"a",{get:function(){return 7}}).a})),xa=function(t){if(!ba(t))throw TypeError(String(t)+" is not an object");return t},ka=function(t,e){if(!ba(t))return t;var r,n;if(e&&"function"==typeof(r=t.toString)&&!ba(n=r.call(t)))return n;if("function"==typeof(r=t.valueOf)&&!ba(n=r.call(t)))return n;if(!e&&"function"==typeof(r=t.toString)&&!ba(n=r.call(t)))return n;throw TypeError("Can't convert object to primitive value")},Ea=Object.defineProperty,Ta={f:ga?Ea:function(t,e,r){if(xa(t),e=ka(e,!0),xa(r),Oa)try{return Ea(t,e,r)}catch(t){}if("get"in r||"set"in r)throw TypeError("Accessors not supported");return"value"in r&&(t[e]=r.value),t}},ja=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}},Aa=ga?function(t,e,r){return Ta.f(t,e,ja(1,r))}:function(t,e,r){return t[e]=r,t},_a=function(t,e){try{Aa(ya,t,e)}catch(r){ya[t]=e}return e},Pa="__core-js_shared__",Ia=ya[Pa]||_a(Pa,{}),La=Function.toString;"function"!=typeof Ia.inspectSource&&(Ia.inspectSource=function(t){return La.call(t)});var Ra,Ca,Ua,Da=Ia.inspectSource,Fa=ya.WeakMap,Ma="function"==typeof Fa&&/native code/.test(Da(Fa)),Ba={}.hasOwnProperty,Na=function(t,e){return Ba.call(t,e)},Ga=pa((function(t){(t.exports=function(t,e){return Ia[t]||(Ia[t]=void 0!==e?e:{})})("versions",[]).push({version:"3.9.1",mode:"global",copyright:"© 2021 Denis Pushkarev (zloirock.ru)"})})),za=0,Wa=Math.random(),Va=function(t){return"Symbol("+String(void 0===t?"":t)+")_"+(++za+Wa).toString(36)},qa=Ga("keys"),Ha=function(t){return qa[t]||(qa[t]=Va(t))},Qa={},$a=ya.WeakMap;if(Ma){var Ka=Ia.state||(Ia.state=new $a),Ya=Ka.get,Ja=Ka.has,Xa=Ka.set;Ra=function(t,e){return e.facade=t,Xa.call(Ka,t,e),e},Ca=function(t){return Ya.call(Ka,t)||{}},Ua=function(t){return Ja.call(Ka,t)}}else{var Za=Ha("state");Qa[Za]=!0,Ra=function(t,e){return e.facade=t,Aa(t,Za,e),e},Ca=function(t){return Na(t,Za)?t[Za]:{}},Ua=function(t){return Na(t,Za)}}var tu,eu,ru={set:Ra,get:Ca,has:Ua,enforce:function(t){return Ua(t)?Ca(t):Ra(t,{})},getterFor:function(t){return function(e){var r;if(!ba(e)||(r=Ca(e)).type!==t)throw TypeError("Incompatible receiver, "+t+" required");return r}}},nu={}.propertyIsEnumerable,ou=Object.getOwnPropertyDescriptor,iu={f:ou&&!nu.call({1:2},1)?function(t){var e=ou(this,t);return!!e&&e.enumerable}:nu},au={}.toString,uu=function(t){return au.call(t).slice(8,-1)},cu="".split,su=va((function(){return!Object("z").propertyIsEnumerable(0)}))?function(t){return"String"==uu(t)?cu.call(t,""):Object(t)}:Object,fu=function(t){return su(sa(t))},lu=Object.getOwnPropertyDescriptor,hu={f:ga?lu:function(t,e){if(t=fu(t),e=ka(e,!0),Oa)try{return lu(t,e)}catch(t){}if(Na(t,e))return ja(!iu.f.call(t,e),t[e])}},pu=pa((function(t){var e=ru.get,r=ru.enforce,n=String(String).split("String");(t.exports=function(t,e,o,i){var a,u=!!i&&!!i.unsafe,c=!!i&&!!i.enumerable,s=!!i&&!!i.noTargetGet;"function"==typeof o&&("string"!=typeof e||Na(o,"name")||Aa(o,"name",e),(a=r(o)).source||(a.source=n.join("string"==typeof e?e:""))),t!==ya?(u?!s&&t[e]&&(c=!0):delete t[e],c?t[e]=o:Aa(t,e,o)):c?t[e]=o:_a(e,o)})(Function.prototype,"toString",(function(){return"function"==typeof this&&e(this).source||Da(this)}))})),du=ya,yu=function(t){return"function"==typeof t?t:void 0},vu=function(t,e){return arguments.length<2?yu(du[t])||yu(ya[t]):du[t]&&du[t][e]||ya[t]&&ya[t][e]},gu=Math.min,bu=function(t){return t>0?gu(ca(t),9007199254740991):0},mu=Math.max,wu=Math.min,Su=function(t){return function(e,r,n){var o,i=fu(e),a=bu(i.length),u=function(t,e){var r=ca(t);return r<0?mu(r+e,0):wu(r,e)}(n,a);if(t&&r!=r){for(;a>u;)if((o=i[u++])!=o)return!0}else for(;a>u;u++)if((t||u in i)&&i[u]===r)return t||u||0;return!t&&-1}},Ou={includes:Su(!0),indexOf:Su(!1)}.indexOf,xu=function(t,e){var r,n=fu(t),o=0,i=[];for(r in n)!Na(Qa,r)&&Na(n,r)&&i.push(r);for(;e.length>o;)Na(n,r=e[o++])&&(~Ou(i,r)||i.push(r));return i},ku=["constructor","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","toLocaleString","toString","valueOf"],Eu=ku.concat("length","prototype"),Tu={f:Object.getOwnPropertyNames||function(t){return xu(t,Eu)}},ju={f:Object.getOwnPropertySymbols},Au=vu("Reflect","ownKeys")||function(t){var e=Tu.f(xa(t)),r=ju.f;return r?e.concat(r(t)):e},_u=function(t,e){for(var r=Au(e),n=Ta.f,o=hu.f,i=0;i<r.length;i++){var a=r[i];Na(t,a)||n(t,a,o(e,a))}},Pu=/#|\.prototype\./,Iu=function(t,e){var r=Ru[Lu(t)];return r==Uu||r!=Cu&&("function"==typeof e?va(e):!!e)},Lu=Iu.normalize=function(t){return String(t).replace(Pu,".").toLowerCase()},Ru=Iu.data={},Cu=Iu.NATIVE="N",Uu=Iu.POLYFILL="P",Du=Iu,Fu=hu.f,Mu=function(t,e){var n,o,i,a,u,c=t.target,s=t.global,f=t.stat;if(n=s?ya:f?ya[c]||_a(c,{}):(ya[c]||{}).prototype)for(o in e){if(a=e[o],i=t.noTargetGet?(u=Fu(n,o))&&u.value:n[o],!Du(s?o:c+(f?".":"#")+o,t.forced)&&void 0!==i){if(r(a)===r(i))continue;_u(a,i)}(t.sham||i&&i.sham)&&Aa(a,"sham",!0),pu(n,o,a,t)}},Bu=function(t){return Object(sa(t))},Nu=!va((function(){function t(){}return t.prototype.constructor=null,Object.getPrototypeOf(new t)!==t.prototype})),Gu=Ha("IE_PROTO"),zu=Object.prototype,Wu=Nu?Object.getPrototypeOf:function(t){return t=Bu(t),Na(t,Gu)?t[Gu]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?zu:null},Vu="process"==uu(ya.process),qu=vu("navigator","userAgent")||"",Hu=ya.process,Qu=Hu&&Hu.versions,$u=Qu&&Qu.v8;$u?eu=(tu=$u.split("."))[0]+tu[1]:qu&&(!(tu=qu.match(/Edge\/(\d+)/))||tu[1]>=74)&&(tu=qu.match(/Chrome\/(\d+)/))&&(eu=tu[1]);var Ku,Yu,Ju,Xu=eu&&+eu,Zu=!!Object.getOwnPropertySymbols&&!va((function(){return!Symbol.sham&&(Vu?38===Xu:Xu>37&&Xu<41)})),tc=Zu&&!Symbol.sham&&"symbol"==r(Symbol.iterator),ec=Ga("wks"),rc=ya.Symbol,nc=tc?rc:rc&&rc.withoutSetter||Va,oc=function(t){return Na(ec,t)&&(Zu||"string"==typeof ec[t])||(Zu&&Na(rc,t)?ec[t]=rc[t]:ec[t]=nc("Symbol."+t)),ec[t]},ic=oc("iterator"),ac=!1;[].keys&&("next"in(Ju=[].keys())?(Yu=Wu(Wu(Ju)))!==Object.prototype&&(Ku=Yu):ac=!0);var uc=null==Ku||va((function(){var t={};return Ku[ic].call(t)!==t}));uc&&(Ku={}),Na(Ku,ic)||Aa(Ku,ic,(function(){return this}));var cc,sc={IteratorPrototype:Ku,BUGGY_SAFARI_ITERATORS:ac},fc=Object.keys||function(t){return xu(t,ku)},lc=ga?Object.defineProperties:function(t,e){xa(t);for(var r,n=fc(e),o=n.length,i=0;o>i;)Ta.f(t,r=n[i++],e[r]);return t},hc=vu("document","documentElement"),pc=Ha("IE_PROTO"),dc=function(){},yc=function(t){return"<script>"+t+"</"+"script>"},vc=function(){try{cc=document.domain&&new ActiveXObject("htmlfile")}catch(t){}var t,e;vc=cc?function(t){t.write(yc("")),t.close();var e=t.parentWindow.Object;return t=null,e}(cc):((e=Sa("iframe")).style.display="none",hc.appendChild(e),e.src=String("javascript:"),(t=e.contentWindow.document).open(),t.write(yc("document.F=Object")),t.close(),t.F);for(var r=ku.length;r--;)delete vc.prototype[ku[r]];return vc()};Qa[pc]=!0;var gc=Object.create||function(t,e){var r;return null!==t?(dc.prototype=xa(t),r=new dc,dc.prototype=null,r[pc]=t):r=vc(),void 0===e?r:lc(r,e)},bc=Ta.f,mc=oc("toStringTag"),wc=function(t,e,r){t&&!Na(t=r?t:t.prototype,mc)&&bc(t,mc,{configurable:!0,value:e})},Sc={},Oc=sc.IteratorPrototype,xc=function(){return this},kc=Object.setPrototypeOf||("__proto__"in{}?function(){var t,e=!1,r={};try{(t=Object.getOwnPropertyDescriptor(Object.prototype,"__proto__").set).call(r,[]),e=r instanceof Array}catch(t){}return function(r,n){return xa(r),function(t){if(!ba(t)&&null!==t)throw TypeError("Can't set "+String(t)+" as a prototype")}(n),e?t.call(r,n):r.__proto__=n,r}}():void 0),Ec=sc.IteratorPrototype,Tc=sc.BUGGY_SAFARI_ITERATORS,jc=oc("iterator"),Ac="keys",_c="values",Pc="entries",Ic=function(){return this},Lc=function(t,e,r,n,o,i,a){!function(t,e,r){var n=e+" Iterator";t.prototype=gc(Oc,{next:ja(1,r)}),wc(t,n,!1),Sc[n]=xc}(r,e,n);var u,c,s,f=function(t){if(t===o&&y)return y;if(!Tc&&t in p)return p[t];switch(t){case Ac:case _c:case Pc:return function(){return new r(this,t)}}return function(){return new r(this)}},l=e+" Iterator",h=!1,p=t.prototype,d=p[jc]||p["@@iterator"]||o&&p[o],y=!Tc&&d||f(o),v="Array"==e&&p.entries||d;if(v&&(u=Wu(v.call(new t)),Ec!==Object.prototype&&u.next&&(Wu(u)!==Ec&&(kc?kc(u,Ec):"function"!=typeof u[jc]&&Aa(u,jc,Ic)),wc(u,l,!0))),o==_c&&d&&d.name!==_c&&(h=!0,y=function(){return d.call(this)}),p[jc]!==y&&Aa(p,jc,y),Sc[e]=y,o)if(c={values:f(_c),keys:i?y:f(Ac),entries:f(Pc)},a)for(s in c)(Tc||h||!(s in p))&&pu(p,s,c[s]);else Mu({target:e,proto:!0,forced:Tc||h},c);return c},Rc=la.charAt,Cc="String Iterator",Uc=ru.set,Dc=ru.getterFor(Cc);Lc(String,"String",(function(t){Uc(this,{type:Cc,string:String(t),index:0})}),(function(){var t,e=Dc(this),r=e.string,n=e.index;return n>=r.length?{value:void 0,done:!0}:(t=Rc(r,n),e.index+=t.length,{value:t,done:!1})}));var Fc=function(t,e,r){if(function(t){if("function"!=typeof t)throw TypeError(String(t)+" is not a function")}(t),void 0===e)return t;switch(r){case 0:return function(){return t.call(e)};case 1:return function(r){return t.call(e,r)};case 2:return function(r,n){return t.call(e,r,n)};case 3:return function(r,n,o){return t.call(e,r,n,o)}}return function(){return t.apply(e,arguments)}},Mc=function(t){var e=t.return;if(void 0!==e)return xa(e.call(t)).value},Bc=function(t,e,r,n){try{return n?e(xa(r)[0],r[1]):e(r)}catch(e){throw Mc(t),e}},Nc=oc("iterator"),Gc=Array.prototype,zc=function(t){return void 0!==t&&(Sc.Array===t||Gc[Nc]===t)},Wc=function(t,e,r){var n=ka(e);n in t?Ta.f(t,n,ja(0,r)):t[n]=r},Vc={};Vc[oc("toStringTag")]="z";var qc="[object z]"===String(Vc),Hc=oc("toStringTag"),Qc="Arguments"==uu(function(){return arguments}()),$c=qc?uu:function(t){var e,r,n;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(r=function(t,e){try{return t[e]}catch(t){}}(e=Object(t),Hc))?r:Qc?uu(e):"Object"==(n=uu(e))&&"function"==typeof e.callee?"Arguments":n},Kc=oc("iterator"),Yc=function(t){if(null!=t)return t[Kc]||t["@@iterator"]||Sc[$c(t)]},Jc=oc("iterator"),Xc=!1;try{var Zc=0,ts={next:function(){return{done:!!Zc++}},return:function(){Xc=!0}};ts[Jc]=function(){return this},Array.from(ts,(function(){throw 2}))}catch(t){}var es=function(t,e){if(!e&&!Xc)return!1;var r=!1;try{var n={};n[Jc]=function(){return{next:function(){return{done:r=!0}}}},t(n)}catch(t){}return r},rs=!es((function(t){Array.from(t)}));Mu({target:"Array",stat:!0,forced:rs},{from:function(t){var e,r,n,o,i,a,u=Bu(t),c="function"==typeof this?this:Array,s=arguments.length,f=s>1?arguments[1]:void 0,l=void 0!==f,h=Yc(u),p=0;if(l&&(f=Fc(f,s>2?arguments[2]:void 0,2)),null==h||c==Array&&zc(h))for(r=new c(e=bu(u.length));e>p;p++)a=l?f(u[p],p):u[p],Wc(r,p,a);else for(i=(o=h.call(u)).next,r=new c;!(n=i.call(o)).done;p++)a=l?Bc(o,f,[n.value,p],!0):n.value,Wc(r,p,a);return r.length=p,r}}),du.Array.from;var ns=!va((function(){return Object.isExtensible(Object.preventExtensions({}))})),os=pa((function(t){var e=Ta.f,n=Va("meta"),o=0,i=Object.isExtensible||function(){return!0},a=function(t){e(t,n,{value:{objectID:"O"+ ++o,weakData:{}}})},u=t.exports={REQUIRED:!1,fastKey:function(t,e){if(!ba(t))return"symbol"==r(t)?t:("string"==typeof t?"S":"P")+t;if(!Na(t,n)){if(!i(t))return"F";if(!e)return"E";a(t)}return t[n].objectID},getWeakData:function(t,e){if(!Na(t,n)){if(!i(t))return!0;if(!e)return!1;a(t)}return t[n].weakData},onFreeze:function(t){return ns&&u.REQUIRED&&i(t)&&!Na(t,n)&&a(t),t}};Qa[n]=!0})),is=function(t,e){this.stopped=t,this.result=e},as=function(t,e,n){var o,i,a,u,c,s,f,l=n&&n.that,h=!(!n||!n.AS_ENTRIES),p=!(!n||!n.IS_ITERATOR),d=!(!n||!n.INTERRUPTED),y=Fc(e,l,1+h+d),v=function(t){return o&&Mc(o),new is(!0,t)},g=function(t){return h?(xa(t),d?y(t[0],t[1],v):y(t[0],t[1])):d?y(t,v):y(t)};if(p)o=t;else{if("function"!=typeof(i=Yc(t)))throw TypeError("Target is not iterable");if(zc(i)){for(a=0,u=bu(t.length);u>a;a++)if((c=g(t[a]))&&c instanceof is)return c;return new is(!1)}o=i.call(t)}for(s=o.next;!(f=s.call(o)).done;){try{c=g(f.value)}catch(t){throw Mc(o),t}if("object"==r(c)&&c&&c instanceof is)return c}return new is(!1)},us=function(t,e,r){if(!(t instanceof e))throw TypeError("Incorrect "+(r?r+" ":"")+"invocation");return t},cs=function(t,e,r){for(var n in e)pu(t,n,e[n],r);return t},ss=oc("species"),fs=Ta.f,ls=os.fastKey,hs=ru.set,ps=ru.getterFor,ds={getConstructor:function(t,e,r,n){var o=t((function(t,i){us(t,o,e),hs(t,{type:e,index:gc(null),first:void 0,last:void 0,size:0}),ga||(t.size=0),null!=i&&as(i,t[n],{that:t,AS_ENTRIES:r})})),i=ps(e),a=function(t,e,r){var n,o,a=i(t),c=u(t,e);return c?c.value=r:(a.last=c={index:o=ls(e,!0),key:e,value:r,previous:n=a.last,next:void 0,removed:!1},a.first||(a.first=c),n&&(n.next=c),ga?a.size++:t.size++,"F"!==o&&(a.index[o]=c)),t},u=function(t,e){var r,n=i(t),o=ls(e);if("F"!==o)return n.index[o];for(r=n.first;r;r=r.next)if(r.key==e)return r};return cs(o.prototype,{clear:function(){for(var t=i(this),e=t.index,r=t.first;r;)r.removed=!0,r.previous&&(r.previous=r.previous.next=void 0),delete e[r.index],r=r.next;t.first=t.last=void 0,ga?t.size=0:this.size=0},delete:function(t){var e=this,r=i(e),n=u(e,t);if(n){var o=n.next,a=n.previous;delete r.index[n.index],n.removed=!0,a&&(a.next=o),o&&(o.previous=a),r.first==n&&(r.first=o),r.last==n&&(r.last=a),ga?r.size--:e.size--}return!!n},forEach:function(t){for(var e,r=i(this),n=Fc(t,arguments.length>1?arguments[1]:void 0,3);e=e?e.next:r.first;)for(n(e.value,e.key,this);e&&e.removed;)e=e.previous},has:function(t){return!!u(this,t)}}),cs(o.prototype,r?{get:function(t){var e=u(this,t);return e&&e.value},set:function(t,e){return a(this,0===t?0:t,e)}}:{add:function(t){return a(this,t=0===t?0:t,t)}}),ga&&fs(o.prototype,"size",{get:function(){return i(this).size}}),o},setStrong:function(t,e,r){var n=e+" Iterator",o=ps(e),i=ps(n);Lc(t,e,(function(t,e){hs(this,{type:n,target:t,state:o(t),kind:e,last:void 0})}),(function(){for(var t=i(this),e=t.kind,r=t.last;r&&r.removed;)r=r.previous;return t.target&&(t.last=r=r?r.next:t.state.first)?"keys"==e?{value:r.key,done:!1}:"values"==e?{value:r.value,done:!1}:{value:[r.key,r.value],done:!1}:(t.target=void 0,{value:void 0,done:!0})}),r?"entries":"values",!r,!0),function(t){var e=vu(t),r=Ta.f;ga&&e&&!e[ss]&&r(e,ss,{configurable:!0,get:function(){return this}})}(e)}};!function(t,e,r){var n=-1!==t.indexOf("Map"),o=-1!==t.indexOf("Weak"),i=n?"set":"add",a=ya[t],u=a&&a.prototype,c=a,s={},f=function(t){var e=u[t];pu(u,t,"add"==t?function(t){return e.call(this,0===t?0:t),this}:"delete"==t?function(t){return!(o&&!ba(t))&&e.call(this,0===t?0:t)}:"get"==t?function(t){return o&&!ba(t)?void 0:e.call(this,0===t?0:t)}:"has"==t?function(t){return!(o&&!ba(t))&&e.call(this,0===t?0:t)}:function(t,r){return e.call(this,0===t?0:t,r),this})};if(Du(t,"function"!=typeof a||!(o||u.forEach&&!va((function(){(new a).entries().next()})))))c=r.getConstructor(e,t,n,i),os.REQUIRED=!0;else if(Du(t,!0)){var l=new c,h=l[i](o?{}:-0,1)!=l,p=va((function(){l.has(1)})),d=es((function(t){new a(t)})),y=!o&&va((function(){for(var t=new a,e=5;e--;)t[i](e,e);return!t.has(-0)}));d||((c=e((function(e,r){us(e,c,t);var o=function(t,e,r){var n,o;return kc&&"function"==typeof(n=e.constructor)&&n!==r&&ba(o=n.prototype)&&o!==r.prototype&&kc(t,o),t}(new a,e,c);return null!=r&&as(r,o[i],{that:o,AS_ENTRIES:n}),o}))).prototype=u,u.constructor=c),(p||y)&&(f("delete"),f("has"),n&&f("get")),(y||h)&&f(i),o&&u.clear&&delete u.clear}s[t]=c,Mu({global:!0,forced:c!=a},s),wc(c,t),o||r.setStrong(c,t,n)}("Set",(function(t){return function(){return t(this,arguments.length?arguments[0]:void 0)}}),ds);var ys=qc?{}.toString:function(){return"[object "+$c(this)+"]"};qc||pu(Object.prototype,"toString",ys,{unsafe:!0});var vs={CSSRuleList:0,CSSStyleDeclaration:0,CSSValueList:0,ClientRectList:0,DOMRectList:0,DOMStringList:0,DOMTokenList:1,DataTransferItemList:0,FileList:0,HTMLAllCollection:0,HTMLCollection:0,HTMLFormElement:0,HTMLSelectElement:0,MediaList:0,MimeTypeArray:0,NamedNodeMap:0,NodeList:1,PaintRequestList:0,Plugin:0,PluginArray:0,SVGLengthList:0,SVGNumberList:0,SVGPathSegList:0,SVGPointList:0,SVGStringList:0,SVGTransformList:0,SourceBufferList:0,StyleSheetList:0,TextTrackCueList:0,TextTrackList:0,TouchList:0},gs=oc("unscopables"),bs=Array.prototype;null==bs[gs]&&Ta.f(bs,gs,{configurable:!0,value:gc(null)});var ms=function(t){bs[gs][t]=!0},ws="Array Iterator",Ss=ru.set,Os=ru.getterFor(ws),xs=Lc(Array,"Array",(function(t,e){Ss(this,{type:ws,target:fu(t),index:0,kind:e})}),(function(){var t=Os(this),e=t.target,r=t.kind,n=t.index++;return!e||n>=e.length?(t.target=void 0,{value:void 0,done:!0}):"keys"==r?{value:n,done:!1}:"values"==r?{value:e[n],done:!1}:{value:[n,e[n]],done:!1}}),"values");Sc.Arguments=Sc.Array,ms("keys"),ms("values"),ms("entries");var ks=oc("iterator"),Es=oc("toStringTag"),Ts=xs.values;for(var js in vs){var As=ya[js],_s=As&&As.prototype;if(_s){if(_s[ks]!==Ts)try{Aa(_s,ks,Ts)}catch(t){_s[ks]=Ts}if(_s[Es]||Aa(_s,Es,js),vs[js])for(var Ps in xs)if(_s[Ps]!==xs[Ps])try{Aa(_s,Ps,xs[Ps])}catch(t){_s[Ps]=xs[Ps]}}}du.Set;var Is,Ls,Rs,Cs=function(){function t(e){i(this,t),this.prefix=e}return u(t,[{key:"get",value:function(t){return"".concat(this.prefix,"_").concat(t)}},{key:"getIndex",value:function(){return this.get("INDEX")}}]),t}(),Us=pa((function(t){var e=function(){var t=String.fromCharCode,e="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=",r="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+-$",n={};function o(t,e){if(!n[t]){n[t]={};for(var r=0;r<t.length;r++)n[t][t.charAt(r)]=r}return n[t][e]}var i={compressToBase64:function(t){if(null==t)return"";var r=i._compress(t,6,(function(t){return e.charAt(t)}));switch(r.length%4){default:case 0:return r;case 1:return r+"===";case 2:return r+"==";case 3:return r+"="}},decompressFromBase64:function(t){return null==t?"":""==t?null:i._decompress(t.length,32,(function(r){return o(e,t.charAt(r))}))},compressToUTF16:function(e){return null==e?"":i._compress(e,15,(function(e){return t(e+32)}))+" "},decompressFromUTF16:function(t){return null==t?"":""==t?null:i._decompress(t.length,16384,(function(e){return t.charCodeAt(e)-32}))},compressToUint8Array:function(t){for(var e=i.compress(t),r=new Uint8Array(2*e.length),n=0,o=e.length;n<o;n++){var a=e.charCodeAt(n);r[2*n]=a>>>8,r[2*n+1]=a%256}return r},decompressFromUint8Array:function(e){if(null==e)return i.decompress(e);for(var r=new Array(e.length/2),n=0,o=r.length;n<o;n++)r[n]=256*e[2*n]+e[2*n+1];var a=[];return r.forEach((function(e){a.push(t(e))})),i.decompress(a.join(""))},compressToEncodedURIComponent:function(t){return null==t?"":i._compress(t,6,(function(t){return r.charAt(t)}))},decompressFromEncodedURIComponent:function(t){return null==t?"":""==t?null:(t=t.replace(/ /g,"+"),i._decompress(t.length,32,(function(e){return o(r,t.charAt(e))})))},compress:function(e){return i._compress(e,16,(function(e){return t(e)}))},_compress:function(t,e,r){if(null==t)return"";var n,o,i,a={},u={},c="",s="",f="",l=2,h=3,p=2,d=[],y=0,v=0;for(i=0;i<t.length;i+=1)if(c=t.charAt(i),Object.prototype.hasOwnProperty.call(a,c)||(a[c]=h++,u[c]=!0),s=f+c,Object.prototype.hasOwnProperty.call(a,s))f=s;else{if(Object.prototype.hasOwnProperty.call(u,f)){if(f.charCodeAt(0)<256){for(n=0;n<p;n++)y<<=1,v==e-1?(v=0,d.push(r(y)),y=0):v++;for(o=f.charCodeAt(0),n=0;n<8;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1}else{for(o=1,n=0;n<p;n++)y=y<<1|o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o=0;for(o=f.charCodeAt(0),n=0;n<16;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1}0==--l&&(l=Math.pow(2,p),p++),delete u[f]}else for(o=a[f],n=0;n<p;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1;0==--l&&(l=Math.pow(2,p),p++),a[s]=h++,f=String(c)}if(""!==f){if(Object.prototype.hasOwnProperty.call(u,f)){if(f.charCodeAt(0)<256){for(n=0;n<p;n++)y<<=1,v==e-1?(v=0,d.push(r(y)),y=0):v++;for(o=f.charCodeAt(0),n=0;n<8;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1}else{for(o=1,n=0;n<p;n++)y=y<<1|o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o=0;for(o=f.charCodeAt(0),n=0;n<16;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1}0==--l&&(l=Math.pow(2,p),p++),delete u[f]}else for(o=a[f],n=0;n<p;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1;0==--l&&(l=Math.pow(2,p),p++)}for(o=2,n=0;n<p;n++)y=y<<1|1&o,v==e-1?(v=0,d.push(r(y)),y=0):v++,o>>=1;for(;;){if(y<<=1,v==e-1){d.push(r(y));break}v++}return d.join("")},decompress:function(t){return null==t?"":""==t?null:i._decompress(t.length,32768,(function(e){return t.charCodeAt(e)}))},_decompress:function(e,r,n){var o,i,a,u,c,s,f,l=[],h=4,p=4,d=3,y="",v=[],g={val:n(0),position:r,index:1};for(o=0;o<3;o+=1)l[o]=o;for(a=0,c=Math.pow(2,2),s=1;s!=c;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*s,s<<=1;switch(a){case 0:for(a=0,c=Math.pow(2,8),s=1;s!=c;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*s,s<<=1;f=t(a);break;case 1:for(a=0,c=Math.pow(2,16),s=1;s!=c;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*s,s<<=1;f=t(a);break;case 2:return""}for(l[3]=f,i=f,v.push(f);;){if(g.index>e)return"";for(a=0,c=Math.pow(2,d),s=1;s!=c;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*s,s<<=1;switch(f=a){case 0:for(a=0,c=Math.pow(2,8),s=1;s!=c;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*s,s<<=1;l[p++]=t(a),f=p-1,h--;break;case 1:for(a=0,c=Math.pow(2,16),s=1;s!=c;)u=g.val&g.position,g.position>>=1,0==g.position&&(g.position=r,g.val=n(g.index++)),a|=(u>0?1:0)*s,s<<=1;l[p++]=t(a),f=p-1,h--;break;case 2:return v.join("")}if(0==h&&(h=Math.pow(2,d),d++),l[f])y=l[f];else{if(f!==p)return null;y=i+i.charAt(0)}v.push(y),l[p++]=i+y.charAt(0),i=y,0==--h&&(h=Math.pow(2,d),d++)}}};return i}();null!=t&&(t.exports=e)})),Ds={},Fs={setItem:function(t,e){return Ds[t]=e,!0},getItem:function(t){return Ds[t]},removeItem:function(t){return delete Ds[t],!0}},Ms=function(){function t(e){switch(i(this,t),e){case"localStorage":case"sessionStorage":this.engine=window[e];break;case"InApp":this.engine=Fs;break;default:throw new ra(na,"Unknow engine")}}return u(t,[{key:"set",value:function(t,e){this.engine.setItem(t,Us.compressToBase64(JSON.stringify(e)))}},{key:"get",value:function(t){var e=this.engine.getItem(t);if(e)try{return JSON.parse(Us.decompressFromBase64(e))}catch(t){return ia(t),null}return null}},{key:"delete",value:function(t){return this.engine.removeItem(t)}}]),t}(),Bs="localStorage",Ns="storage-js",Gs=function(){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},r=e.storageEngine,n=e.defaultTTL,o=e.prefix;i(this,t),this.store=new Ms(r||Bs),this.defaultTTL=n||0,this.key=new Cs(o||Ns)}return u(t,[{key:"save",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=e.ttl;if(!t.id)throw new ra(na,"Object must have an id property");var n={object:t},o=(new Date).getTime();r?n.ttl=o+1e3*r:this.defaultTTL>0&&(n.ttl=o+1e3*this.defaultTTL),this.store.set(this.key.get(t.id),n),this.addToIndex(t.id)}},{key:"find",value:function(t){var e=this.store.get(this.key.get(t));return e&&(!e.ttl||(new Date).getTime()<=e.ttl)?e.object:{}}},{key:"findAll",value:function(){var t=this;return Array.from(this.loadIndex()).map((function(e){return t.find(e)})).filter((function(t){return!!t}))}},{key:"delete",value:function(t){return this.removeToIndex(t),this.store.delete(this.key.get(t))}},{key:"loadIndex",value:function(){var t=this.store.get(this.key.getIndex());return t?new Set(t):new Set}},{key:"addToIndex",value:function(t){var e=this.loadIndex();e.add(t),this.store.set(this.key.getIndex(),f(e))}},{key:"removeToIndex",value:function(t){var e=this.loadIndex();e.delete(t),this.store.set(this.key.getIndex(),f(e))}},{key:"resetIndex",value:function(){this.store.delete(this.key.getIndex())}}]),t}(),zs={sdkName:"__ISDK",sdkScriptId:"__ISDK_ASSETS",consent:{name:"consent",ttl:390,status:{unknown:"unknown",optin:"optin",optout:"optout"}},event_consent_id:{name:"event_consent_id",ttl:390},subid:{name:"subid",queryname:"toSubid",ttl:40},cashback:{name:"cashback",queryname:"toCashback",ttl:30},stats:{type:{visit:"visit",conversion:"conversion"}},default_storage_prefix:"to",default_ttl:390,urls:{conversion:(null===(Is="https://b.time1.me/v1/b")?void 0:Is.split(","))||[],stats:(null===(Ls="https://c.time1.me/v1/log/consent")?void 0:Ls.split(","))||[],proofConsent:(null===(Rs="https://c.time1.me/v1/log/consent/proof")?void 0:Rs.split(","))||[]}},Ws=regeneratorRuntime.mark(Ks),Vs=new Gs({storageEngine:"localStorage",prefix:zs.default_storage_prefix,defaultTTL:zs.consent.ttl});function qs(t){return"".concat(zs.default_storage_prefix,"_").concat(t)}function Hs(t,e){var r=zs[e]||{name:e,ttl:zs.default_ttl};ea.set(qs(r.name),t,{expires:r.ttl,sameSite:"strict"}),Vs.save({id:e,value:t})}function Qs(t){var e,r=(null===(e=zs[t])||void 0===e?void 0:e.name)||t,n=ea.get(qs(r));if(n)return n;var o=Vs.find(r);return null!=o&&o.value?null==o?void 0:o.value:null}function $s(t){var e,r=(null===(e=zs[t])||void 0===e?void 0:e.name)||t;ea.remove(qs(r)),Vs.delete(r)}function Ks(){var t,e,r=arguments;return regeneratorRuntime.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:t=r.length>0&&void 0!==r[0]?r[0]:[],e=0;case 2:if(!(e<=t.length)){n.next=8;break}return n.next=5,t[e];case 5:e+=1,n.next=2;break;case 8:case"end":return n.stop()}}),Ws)}function Ys(t){var e=Ks(t),r=e.next().value;return{next:function(){r=e.next().value},get url(){return r},urls:t}}var Js=new WeakMap,Xs=new WeakMap,Zs=new WeakMap,tf=new WeakMap,ef=new WeakMap,rf=new WeakSet,nf=new WeakSet,of=new WeakSet,af=new WeakSet,uf=new WeakSet,cf=new WeakSet,sf=new WeakSet,ff=new WeakSet,lf=new WeakSet,hf=new WeakSet,pf=new WeakSet,df=function(){function t(){i(this,t),O(this,pf),O(this,hf),O(this,lf),O(this,ff),O(this,sf),O(this,cf),O(this,uf),O(this,af),O(this,of),O(this,nf),O(this,rf),S(this,Js,{writable:!0,value:[]}),S(this,Xs,{writable:!0,value:Ys(zs.urls.conversion)}),S(this,Zs,{writable:!0,value:Ys(zs.urls.stats)}),S(this,tf,{writable:!0,value:Ys(zs.urls.proofConsent)}),S(this,ef,{writable:!0,value:[]}),this.env="production",m(this,rf,yf).call(this),m(this,nf,vf).call(this,zs.cashback),this.consent||m(this,cf,Sf).call(this,zs.consent.status.unknown),this.consent===zs.consent.status.optin?m(this,nf,vf).call(this,zs.subid):m(this,sf,Of).call(this)}var e,r,n,a;return u(t,[{key:"consent",get:function(){return Qs(zs.consent.name)}},{key:"subid",get:function(){return this.constructor.getProgramDataFromQueryParams(zs.subid.queryname)||Qs(zs.subid.name)}},{key:"cashbackSubid",get:function(){return this.constructor.getProgramDataFromQueryParams(zs.cashback.queryname)||Qs(zs.cashback.name)}},{key:"eventConsentId",get:function(){return Qs(zs.event_consent_id.name)}},{key:"_setOptin",value:function(){m(this,cf,Sf).call(this,zs.consent.status.optin),m(this,nf,vf).call(this,zs.subid)}},{key:"_setOptout",value:function(){m(this,cf,Sf).call(this,zs.consent.status.optout),m(this,sf,Of).call(this)}},{key:"_setUnknown",value:function(){m(this,cf,Sf).call(this,zs.consent.status.unknown),m(this,sf,Of).call(this)}},{key:"_setSale",value:(a=o(regeneratorRuntime.mark((function t(e){return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,m(this,pf,Tf).call(this,{data:e,caller:"_setSale"});case 3:t.next=8;break;case 5:t.prev=5,t.t0=t.catch(0),m(this,lf,kf).call(this,{error:t.t0,caller:"_setSale",extra:e});case 8:case"end":return t.stop()}}),t,this,[[0,5]])}))),function(t){return a.apply(this,arguments)})},{key:"_setLead",value:(n=o(regeneratorRuntime.mark((function t(e){return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,m(this,pf,Tf).call(this,{data:e,caller:"_setLead"});case 3:t.next=8;break;case 5:t.prev=5,t.t0=t.catch(0),m(this,lf,kf).call(this,{error:t.t0,caller:"_setLead",extra:e});case 8:case"end":return t.stop()}}),t,this,[[0,5]])}))),function(t){return n.apply(this,arguments)})},{key:"_setDbClick",value:(r=o(regeneratorRuntime.mark((function t(e){return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,m(this,pf,Tf).call(this,{data:e,caller:"_setDbClick"});case 3:t.next=8;break;case 5:t.prev=5,t.t0=t.catch(0),m(this,lf,kf).call(this,{error:t.t0,caller:"_setDbClick",extra:e});case 8:case"end":return t.stop()}}),t,this,[[0,5]])}))),function(t){return r.apply(this,arguments)})},{key:"_setClick",value:(e=o(regeneratorRuntime.mark((function t(e){return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,m(this,pf,Tf).call(this,{data:e,caller:"_setClick"});case 3:t.next=8;break;case 5:t.prev=5,t.t0=t.catch(0),m(this,lf,kf).call(this,{error:t.t0,caller:"_setClick",extra:e});case 8:case"end":return t.stop()}}),t,this,[[0,5]])}))),function(t){return e.apply(this,arguments)})},{key:"push",value:function(t){var e,r=l(e=t||[])||h(e)||p(e)||y(),n=r[0],o=r.slice(1);try{if("function"!=typeof this[n])throw new Error("Undefined function ".concat(n));this[n].apply(this,f(o))}catch(t){m(this,lf,kf).call(this,{error:t,method:n})}}},{key:"getTrace",value:function(){return{env:this.env,progids:v(this,Js),consent:this.consent,subid:this.subid,event_consent_id:this.eventConsentId,cashbackSubid:this.cashbackSubid,errors:m(this,hf,Ef).call(this),conversionUrls:zs.urls.conversion}}}],[{key:"getProgramDataFromQueryParams",value:function(t){return new URLSearchParams(window.location.search).get(t)}}]),t}();function yf(){try{var t,e=null===(t=document.getElementById(zs.sdkScriptId))||void 0===t?void 0:t.getAttribute("data-progids");e&&g(this,Js,JSON.parse(e))}catch(t){m(this,lf,kf).call(this,{error:t,method:"setProgids"})}}function vf(t){var e=t.name,r=t.queryname,n=this.constructor.getProgramDataFromQueryParams(r);n&&Hs(n,e)}function gf(t){return bf.apply(this,arguments)}function bf(){return(bf=o(regeneratorRuntime.mark((function t(e){var r,n,o,i,a,u;return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(r=e.urlIterator,n=e.body,o=void 0===n?{}:n,i=e.caller,null!=r&&r.url){t.next=4;break}return m(this,lf,kf).call(this,{error:{message:"Failed to contact server on ".concat(null==r?void 0:r.urls)},caller:i}),t.abrupt("return");case 4:return t.prev=4,t.next=7,fetch(null==r?void 0:r.url,{method:"POST",headers:{accept:"application/json","Content-Type":"application/json"},body:JSON.stringify(o)});case 7:if((a=t.sent).ok){t.next=13;break}return t.next=11,a.json();case 11:u=t.sent,m(this,lf,kf).call(this,{error:{message:" code ".concat(null==a?void 0:a.status," - ").concat(null==u?void 0:u.message)},caller:i,extra:{body:o,url:v(this,Xs).url}});case 13:t.next=19;break;case 15:t.prev=15,t.t0=t.catch(4),r.next(),m(this,of,gf).call(this,{urlIterator:r,body:o,caller:i});case 19:case"end":return t.stop()}}),t,this,[[4,15]])})))).apply(this,arguments)}function mf(t){var e=this,r=t.consent,n=t.type,o=[this.subid,this.cashbackSubid].filter(Boolean);v(this,Js).forEach((function(t){m(e,of,gf).call(e,{urlIterator:v(e,Zs),body:{type:n,progid:t,status:r,toSubids:o},caller:"#logStats"})}))}function wf(){var t=Zi();Hs(t,zs.event_consent_id.name);var e={event_consent_id:t,url:"".concat(window.location.hostname).concat(window.location.pathname)};m(this,of,gf).call(this,{urlIterator:v(this,tf),body:e,caller:"#setPOC"})}function Sf(t){var e=t!==this.consent,r=!this.eventConsentId&&this.subid&&t===zs.consent.status.optin;Hs(t,zs.consent.name),e&&m(this,af,mf).call(this,{consent:t,type:zs.stats.type.visit}),r&&m(this,uf,wf).call(this)}function Of(){$s(zs.subid.name),$s(zs.event_consent_id.name)}function xf(){return this.cashbackSubid||this.constructor.getProgramDataFromQueryParams(zs.subid.queryname)||this.subid&&this.consent===zs.consent.status.optin}function kf(t){v(this,ef).push(t)}function Ef(){return v(this,ef).map((function(t){var e=t.error,r=t.caller,n=t.extra;return{message:'While calling the method "'.concat(r,'": ').concat(e.message),extra:n}}))}function Tf(t){return jf.apply(this,arguments)}function jf(){return jf=o(regeneratorRuntime.mark((function t(r){var n,o,i,a,u,c,f,l,h,p,d,y,g,b;return regeneratorRuntime.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(o=r.data,i=void 0===o?{}:o,a=r.caller,u=void 0===a?"setConversion":a,m(this,ff,xf).call(this)){t.next=3;break}throw new Error("Make a conversion is not allowed. Check consent or ".concat(zs.subid.queryname));case 3:if(c=i.progid,f=i.comid,l=i.iu,c&&f&&l){t.next=6;break}throw new Error("Missing progid or comid or iu. Those data are mandatory to make a conversion");case 6:if(null!==(n=v(this,Xs))&&void 0!==n&&n.url){t.next=8;break}throw new Error("Failed to contact server on ".concat(JSON.stringify(null===(h=v(this,Xs))||void 0===h?void 0:h.urls)));case 8:return p={type:"consent",value:this.subid},d={type:"cashback",value:this.cashbackSubid},y=[p,d].filter((function(t){return!!t.value})),g=e(e({},i),{},{event_consent_id:this.eventConsentId,toSubids:y}),b=Object.fromEntries(Object.entries(g).filter((function(t){return!!s(t,2)[1]}))),t.next=15,m(this,of,gf).call(this,{urlIterator:v(this,Xs),body:b,caller:u});case 15:case"end":return t.stop()}}),t,this)}))),jf.apply(this,arguments)}var Af=zs.sdkName;var _f=zs.sdkName;window.TOG_Storage=Vs,function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:df;try{if(void 0===window[Af])window[Af]=new t;else if("object"===r(window[Af])&&Array.isArray(window[Af])){var e=new t;window[Af].forEach((function(t){return e.push(t)})),window[Af]=e}else if(!("object"===r(window[Af])&&window[Af]instanceof t))throw new Error("".concat(Af," error: Unknown type"))}catch(t){console.error(t.message)}}(df),function t(e){if(document.getElementById("debug")){var r,n,o,i,a,u={progid:null===(r=window[_f])||void 0===r?void 0:r.progid,consent:null===(n=window[_f])||void 0===n?void 0:n.consent,subid:null===(o=window[_f])||void 0===o?void 0:o.subid,cashback:null===(i=window[_f])||void 0===i?void 0:i.cashback,to_event_consent_id:null===(a=window[_f])||void 0===a?void 0:a.eventConsentId};document.getElementById("debug").textContent=JSON.stringify(u,null,"\t")}else setTimeout((function(){return t(e)}),300);null==e||e()}(),console.log("SDK loaded")}();
