+++ browserstack/stories/01_when_no_consent_defined/should_not_set_subid_if_consent_is_optout_or_unknown.mjs
@@ -12,7 +12,7 @@ import CONSTANTS from '../../../src/constants.mjs';
 import TEST_CONSTANTS from './constants.mjs';
 
 const testName = 'should_not_set_subid_if_consent_is_optout_or_unknown';
-const url = `${TEST_CONSTANTS.baseUrl}?${CONSTANTS.subid.queryname}=${TEST_CONSTANTS.subid}`;
+const url = `${TEST_CONSTANTS.baseUrl}`;
 
 export default async function shouldNotSetSubidIfConsentIsOptoutOrUnknown(driver) {
   printTestInConsole(TEST_CONSTANTS.groupTestName, testName);
