+++ browserstack/stories/01_when_no_consent_defined/should_set_subid_if_consent_is_optin.mjs
@@ -25,7 +25,7 @@ export default async function shouldSetSubidIfConsentIsOptin(driver) {
 
     expect(initialConsent).toEqual(CONSTANTS.consent.status.unknown);
     expect(initialprogid).toBeFalsy();
-    expect(initialSubid).toBeFalsy();
+    expect(initialSubid).toEqual(TEST_CONSTANTS.subid);
 
     await setOptin(driver);
