+++ browserstack/stories/03_when_optout_defined/index.mjs
@@ -1,2 +1,2 @@
-export { default as shouldNotSetSubid } from './should_not_set_subid.mjs';
+export { default as shouldSetSubidButNotWriteInStorage } from './should_set_subid_but_not_write_in_storage.mjs';
 export { default as shouldSetSdkCashbackSubid } from './should_set_sdk_cashbackSubid.mjs';
