+++ src/constants.mjs
@@ -24,11 +24,17 @@ export default {
     queryname: 'toCashback',
     ttl: 30,
   },
+  stats: {
+    type: {
+      visit: 'visit',
+      conversion: 'conversion',
+    },
+  },
   default_storage_prefix: 'to',
   default_ttl: 390,
   urls: {
     conversion: process.env.API_CONVERSION_URLS?.split(',') || [],
-    statsConsent: process.env.API_STATS_CONSENT_URLS?.split(',') || [],
+    stats: process.env.API_STATS_URLS?.split(',') || [],
     proofConsent: process.env.API_PROOF_CONSENT_URLS?.split(',') || [],
   },
 };
